% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addGeometry.R
\name{addGeometry}
\alias{addGeometry}
\title{addGeometry}
\usage{
addGeometry(x, clustermap)
}
\arguments{
\item{x}{Landuse data on cluster/cell resolution as a magclass object}

\item{clustermap}{A dataframe mapping with columns cluster, cell, and country}
}
\value{
A magclass object enriched with geometry information
}
\description{
Enriches land use data on cluster resolution geometry information as required for conversion
by magclass::as.SpatVector
}
\examples{
\dontrun{
landUse <- magpie4::land("fulldata.gdx", level = "cell")
clustermap <- readRDS(Sys.glob("clustermap_*.rds"))
landUseEnriched <- magpie4::addGeometry(landUse, clustermap)
attr(landUseEnriched, "geometry")
attr(landUseEnriched, "crs")
}

}
\author{
Jan Philipp Dietrich, Pascal Sauer
}
