% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addToDataChangelog.R
\name{addToDataChangelog}
\alias{addToDataChangelog}
\title{addToDataChangelog}
\usage{
addToDataChangelog(
  report,
  changelog,
  versionId,
  years = changelogYears(),
  variables = changelogVariables(),
  ...,
  maxEntries = 15,
  roundDigits = 2
)
}
\arguments{
\item{report}{data.frame as obtained by readRDS("report.rds")}

\item{changelog}{Path to the changelog file}

\item{versionId}{The model version identifier, e.g. a release number like 4.9.1 or a date like 2025-02-01}

\item{years}{For which years the variables should be read and put into the changelog}

\item{variables}{named vector of which variables to read from the report}

\item{...}{Reserved for future expansion.}

\item{maxEntries}{The maximum number of versionIds to keep in the changelog, the oldest one is removed first.}

\item{roundDigits}{Numbers are rounded to this many decimal places before being written to the changelog.}
}
\value{
Invisibly, the written changelog as data.frame
}
\description{
Prepend data from the given report to the changelog.
}
\author{
Pascal Sauer
}
