% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agEmployment.R
\name{agEmployment}
\alias{agEmployment}
\title{agEmployment}
\usage{
agEmployment(gdx, type = "absolute", detail = TRUE, level = "reg", file = NULL)
}
\arguments{
\item{gdx}{GDX file}

\item{type}{"absolute" for total number of people employed, "share" for share out of working age population}

\item{detail}{if TRUE, employment is disaggregated to crop products, livestock products and (if available) mitigation
measures, if FALSE only aggregated employment is reported, if byProduct then to all crop products}

\item{level}{spatial aggregation to report employment ("iso", "reg", "glo" or "regglo",
if type is "absolute" also "grid")}

\item{file}{a file name the output should be written to using write.magpie}
}
\value{
employment in agriculture as absolute value or as percentage of working age population
}
\description{
returns employment in crop+livestock production from MAgPIE results
}
\examples{
\dontrun{
x <- agEmployment(gdx)
}
}
\author{
Debbora Leip, David M Chen
}
