% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anthropometrics.R
\name{anthropometrics}
\alias{anthropometrics}
\title{anthropometrics}
\usage{
anthropometrics(
  gdx,
  indicator = "bodyheight",
  age = "adults",
  sex = FALSE,
  bmi_groups = FALSE,
  level = "iso",
  dir = ".",
  spamfiledirectory = "",
  final = TRUE,
  file = NULL,
  calibrated = TRUE
)
}
\arguments{
\item{gdx}{GDX file}

\item{indicator}{bodyheight, bodyweight, bodyweight_healthy, BMI(Body Mass Index) or PAL (physical activity level)}

\item{age}{if TRUE, demand is scaled down to age-groups and sex using food requirements}

\item{sex}{if FALSE, female and male are aggregated, if sex, results are divided into males and females}

\item{bmi_groups}{if TRUE, data is provided by BMI group}

\item{level}{Level of regional aggregation; "iso" ISO country codes, "reg" (regional), "glo" (global)}

\item{dir}{for gridded outputs: magpie output directory which contains a mapping file (rds) for disaggregation}

\item{spamfiledirectory}{deprecated. please use \code{dir} instead}

\item{final}{final results or preliminary results (the latter are the ones magpie uses for optimization before last iteration with demand model)}

\item{file}{a file name the output should be written to using write.magpie}

\item{calibrated}{if TRUE, uses the calibrated intake estimates for bodyweight estimation}
}
\value{
bodyweight (kg), bodyheight (cm), BMI or PAL as magpie objects
}
\description{
Calculates anthropometic indicators from the food demand model
}
\details{
Demand definitions are equivalent to FAO Food supply categories
}
\examples{

  \dontrun{
    x <- anthropometrics(gdx)
  }

}
\author{
Benjamin Leon Bodirsky
}
