% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BII.R
\name{BII}
\alias{BII}
\title{Biodiversity intactness index}
\usage{
BII(
  gdx,
  file = NULL,
  level = "glo",
  mode = "auto",
  bii_coeff = NULL,
  rr_layer = NULL,
  side_layers = NULL
)
}
\arguments{
\item{gdx}{GDX file}

\item{file}{a file name the output should be written to using write.magpie}

\item{level}{Level of regional aggregation; "reg" (regional), "glo" (global), "regglo" (regional and global).}

\item{mode}{"auto" (default), "MAgPIE" or "postprocessing". "MAgPIE" reports the BV based on values from the MAgPIE biodiversity module. "postprocessing" calculates the BV based on land information from MAgPIE. "auto" uses "MAgPIE" if available and falls back to "postprocessing" otherwise.}

\item{bii_coeff}{file containing BII coefficients. Only needed for mode="postprocessing". NULL tries to automatically detected the file.}

\item{rr_layer}{file containing the range-rarity layer. Only needed for mode="postprocessing". NULL tries to automatically detected the file.}

\item{side_layers}{file containing LUH2 side layers. Only needed for mode="postprocessing". NULL tries to automatically detected the file.}
}
\value{
Biodiversity intactness index (unitless)
}
\description{
calculated the area weighted biodiversity intactness index out of a MAgPIE gdx file
}
\details{
Calculates global and regional biodiversity intactness index
}
\examples{
\dontrun{
x <- BII(gdx)
}

}
\author{
Patrick v. Jeetze, Florian Humpenoeder
}
