% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioplasticDemand.R
\name{bioplasticDemand}
\alias{bioplasticDemand}
\title{bioplasticDemand}
\usage{
bioplasticDemand(
  gdx,
  type = "bioplastic",
  detail = FALSE,
  level = "regglo",
  file = NULL
)
}
\arguments{
\item{gdx}{GDX file}

\item{type}{"bioplastic" for bioplastic demand, "substrate" for biomass demand as substrate for bioplastic production}

\item{detail}{only relevant for type = "substrate". If TRUE, substrate demand is disaggregated by crop type, if 
FALSE only the aggregated demand is reported.}

\item{level}{spatial aggregation to report bioplastic/substrate demand (only "reg" or "regglo")}

\item{file}{a file name the output should be written to using write.magpie}
}
\description{
returns demand for bioplastic or demand for substrate for bioplastic production
}
\examples{
\dontrun{
x <- bioplasticDemand(gdx)
}
}
\author{
Debbora Leip
}
