% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bodyweight.R
\name{bodyweight}
\alias{bodyweight}
\title{bodyweight}
\usage{
bodyweight(
  gdx,
  level = "reg",
  age = FALSE,
  sex = FALSE,
  share = FALSE,
  population = NULL
)
}
\arguments{
\item{gdx}{GDX file}

\item{level}{Level of regional aggregation; "iso" ISO country codes, "reg" (regional), "glo" (global)}

\item{age}{if TRUE, demand is scaled down to age-groups and sex using food requirements}

\item{sex}{if FALSE, female and male are aggregated, if sex, results are divided into males and females}

\item{share}{if TRUE, data is provided by BMI group}

\item{population}{population information from GDX. Can be provided to speed up calculation process. Will be read
from GDX, if not provided.}
}
\value{
MAgPIE object with mio people or share of people in each weight category
}
\description{
Calculates the prevalence of underweight, normalweight, overweight (excluding obesity) and obesity. For more detailed body mass classifications see functions population or anthropometrics.
}
\details{
Demand definitions are equivalent to FAO Food supply categories
}
\examples{

  \dontrun{
    x <- bodyweight(gdx)
  }

}
\author{
Benjamin Leon Bodirsky
}
