% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carbonHWP.R
\name{carbonHWP}
\alias{carbonHWP}
\title{carbonHWP}
\usage{
carbonHWP(
  gdx,
  file = NULL,
  level = "cell",
  unit = "element",
  half_life = 35,
  cumulative = FALSE,
  baseyear = 1995
)
}
\arguments{
\item{gdx}{GDX file}

\item{file}{a file name the output should be written to using write.magpie}

\item{level}{Level of regional aggregation; "cell", "reg" (regional), "glo" (global), "regglo" (regional and global) or any secdforest aggregation level defined in superAggregate}

\item{unit}{element" or "gas"; "element": co2_c in Mt C/yr, n2o_n in Mt N/yr, ch4 in Mt CH4/yr; "gas": co2_c Mt CO2/yr, n2o_n in Mt NO2/yr, ch4 in Mt CH4/yr}

\item{half_life}{Half life in years for decay in wood products loosing half theor carbon content. (35 yrs is deafault)}

\item{cumulative}{Logical; Determines if cHWP emissions are reported annually (FALSE) or cumulative (TRUE). The starting point for cumulative emissions is y1995.}

\item{baseyear}{Baseyear used for cumulative emissions (default = 1995)}
}
\value{
carbon stocks in MtC from harvested timber
}
\description{
reads carbon stored in harvested timber out of a MAgPIE gdx file
}
\details{
Annual (and cumulative) Carbon stored in harvested wood products as well as slow emissions from half life deacy.
}
\examples{

  \dontrun{
    x <- carbonHWP(gdx)
  }
}
\author{
Abhijeet Mishra, Florian Humpenoeder
}
