% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carbonstock.R
\name{carbonstock}
\alias{carbonstock}
\title{carbonstock}
\usage{
carbonstock(
  gdx,
  file = NULL,
  level = "cell",
  sum_cpool = TRUE,
  sum_land = TRUE
)
}
\arguments{
\item{gdx}{GDX file}

\item{file}{a file name the output should be written to using write.magpie}

\item{level}{Level of regional aggregation; "cell", "reg" (regional), "glo" (global), "regglo" (regional and global) or any secdforest aggregation level defined in superAggregate}

\item{sum_cpool}{sum over carbon pool dimension (default = TRUE)}

\item{sum_land}{sum over land type dimension (default = TRUE)}
}
\value{
carbon stocks in MtC
}
\description{
reads carbon stocks out of a MAgPIE gdx file
}
\details{
carbon pools consist of vegetation carbon (vegc), litter carbon (litc) and soil carbon (soilc)
}
\examples{

  \dontrun{
    x <- carbonstock(gdx)
  }

}
\author{
Florian Humpenoeder
}
