% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellularFit.R
\name{cellularFit}
\alias{cellularFit}
\title{cellular fit}
\usage{
cellularFit(
  gdx,
  file = NULL,
  level = "cell",
  statistic = "MAE",
  variable = "land",
  dataset = "LUH2",
  water_aggr = FALSE
)
}
\arguments{
\item{gdx}{GDX file}

\item{file}{a file name the output should be written to using write.magpie}

\item{level}{level at which the regional and global bias should be calculated. Options "cell" or "grid"}

\item{statistic}{R2, MAE, MPE (mean percentage error - bias), MAPE (mean absolute percentage error)}

\item{variable}{variable to be evaulated: land (land types) or crop (crop types)}

\item{dataset}{dataset to compare with. LUH2 only option for variable land. LUH2 and MAPSPAM for the crop variable.}

\item{water_aggr}{if irrigation types for crops should be agregated or not}
}
\value{
returns selected statistic at regglo level for the historical part of the time horizon
}
\description{
cellular fit/error/bias calculations at regional and global level
}
\examples{

  \dontrun{
    x <- cellularFit(gdx)
  }

}
\author{
Edna J. Molina Bacca, Patrick v. Jeetze
}
