% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changelogVariables.R
\name{changelogVariables}
\alias{changelogVariables}
\title{changelogVariables}
\usage{
changelogVariables()
}
\description{
Returns a named vector of variables to put into the data changelog, see \code{\link{addToDataChangelog}}.
}
\author{
Pascal Sauer
}
