% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changelogYears.R
\name{changelogYears}
\alias{changelogYears}
\title{changelogYears}
\usage{
changelogYears()
}
\description{
Returns the years to put into the data changelog, see \code{\link{addToDataChangelog}}.
}
\seealso{
Other Data Changelog:
\code{\link[=addToDataChangelog]{addToDataChangelog()}},
\code{\link[=changelogVariables]{changelogVariables()}}
}
\author{
Pascal Sauer
}
\concept{Data Changelog}
