% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkLibrary.R
\name{checkLibrary}
\alias{checkLibrary}
\title{modelstat}
\usage{
checkLibrary(gdx, level = NULL)
}
\arguments{
\item{gdx}{GDX file}

\item{level}{Level of regional aggregation; "reg" (regional), "glo" (global), "regglo" (regional and global) or any other aggregation level defined in superAggregate}
}
\value{
A list with three entries:\itemize{\item{\code{notest}} Testing these functions was impossible because there wer missing arguments.\item{\code{error}} These functions could not be executed properly.\item{\code{fine}} Everything was fine with these functions.}
}
\description{
Function to check if the library functions work with the newest magpie version
}
\details{
This function simply tries to run all functions in the magpie library on the provided gdx file.
}
\examples{

  \dontrun{
    x <- modelstat(gdx)
  }

}
\author{
Markus Bonsch
}
