% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clearCacheMagpie4.R
\name{clearCacheMagpie4}
\alias{clearCacheMagpie4}
\title{clearCacheMagpie4}
\usage{
clearCacheMagpie4()
}
\value{
no return, but cache cleared
}
\description{
Clear memoise cache for all memoised functions in magpie4
}
\examples{
\dontrun{
x <- clearCacheMagpie4()
}

}
\author{
Benjamin Leon Bodirsky
}
