% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterOutputToShapefile.R
\name{clusterOutputToTerraVector}
\alias{clusterOutputToTerraVector}
\title{Convert cluster output to terra vector}
\usage{
clusterOutputToTerraVector(x, clustermap)
}
\arguments{
\item{x}{Landuse data on cluster/cell resolution as a magclass object}

\item{clustermap}{A dataframe mapping with columns cluster, cell, and country}
}
\value{
A SpatVector with the following columns: c("clusterId", "country", "region", "year", "landtype", "value")
}
\description{
Enriches land use data on cluster resolution with explicit spatial information by
creating a terra polygon for each cluster according to the given clustermap.
}
\examples{
\dontrun{
landUse <- magpie4::land("fulldata.gdx", level = "cell")
clustermap <- readRDS(Sys.glob("clustermap_*.rds"))
clusterPolygons <- magpie4::clusterOutputToTerraVector(landUse, clustermap)
terra::writeVector(clusterPolygons, "cluster_resolution.shp")
}

}
\author{
Pascal Führlich, Patrick v. Jeetze
}
