% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consumptionValue.R
\name{consumptionValue}
\alias{consumptionValue}
\title{consumptionValue}
\usage{
consumptionValue(
  gdx,
  file = NULL,
  level = "reg",
  products = "kall",
  product_aggr = TRUE,
  type = NULL,
  type_aggr = TRUE
)
}
\arguments{
\item{gdx}{GDX file}

\item{file}{a file name the output should be written to using write.magpie}

\item{level}{Level of regional aggregation; "reg" (regional), "glo" (global), "regglo" (regional and global) or any other aggregation level defined in superAggregate}

\item{products}{Selection of products (either by naming products, e.g. "tece", or naming a set,e.g."kcr")}

\item{product_aggr}{aggregate over products or not (boolean, default TRUE)}

\item{type}{Consumption type(s): "food", "feed", "processed", "other_util", "bioenergy", "seed", "waste", "dom_balanceflow; NULL returns all types}

\item{type_aggr}{aggregate over demand types or not (boolean, default TRUE)}
}
\value{
A MAgPIE object containing consumption value in million $US.
}
\description{
calculates consumption value of different types based on a MAgPIE gdx file.
}
\examples{

  \dontrun{
    x <- consumptionValue(gdx)
  }
}
\author{
Miodrag Stevanovic
}
