% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CostInputFactorsCrop.R
\name{costInputFactorsCrop}
\alias{costInputFactorsCrop}
\title{costInputFactorsCrop}
\usage{
costInputFactorsCrop(gdx, type = "annuity", file = NULL, level = "reg")
}
\arguments{
\item{gdx}{GDX file}

\item{type}{Type of capital investments accounting. It can either be total investments ("investment"),
or considering the annuity ("annuity") of the current time step. NULL in case the runs were not done with
the sticky realization.}

\item{file}{a file name the output should be written to using write.magpie}

\item{level}{Level of regional aggregation}
}
\value{
A MAgPIE object containing values related with overall value of production [million US$17]
}
\description{
Reads data to calculate Input factors with different approaches
}
\examples{
\dontrun{
x <- costInputFactorsCrop(gdx)
}

}
\author{
Edna Molina Bacca
}
