% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/costsAdaptationCrops.R
\name{costsAdaptationCrops}
\alias{costsAdaptationCrops}
\title{costsAdaptationCrops}
\usage{
costsAdaptationCrops(
  gdx,
  file = NULL,
  level = "regglo",
  type = "investment",
  dir = "."
)
}
\arguments{
\item{gdx}{GDX file}

\item{file}{a file name the output should be written to using write.magpie}

\item{level}{Level of regional aggregation; "reg" (regional), "glo" (global),
"regglo" (regional and global) or any other aggregation level defined in superAggregate}

\item{type}{Type of reporting, either "annuity" or total "investments"}

\item{dir}{directory with mapping for disaggregation to higher resolutions}
}
\value{
A MAgPIE object containing values related with adaptation costs for crops production
per ton produced [million US$05/tDM]
}
\description{
Reads data to calculate adaptation costs for cropland
}
\examples{
\dontrun{
x <- costsAdaptationCrops(gdx)
}

}
\author{
Edna Molina Bacca
}
