% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CostsAEI.R
\name{CostsAEI}
\alias{CostsAEI}
\title{CostsAEI}
\usage{
CostsAEI(gdx, file = NULL, level = "regglo")
}
\arguments{
\item{gdx}{GDX file}

\item{file}{a file name the output should be written to using write.magpie}

\item{level}{Level of regional aggregation ("reg", "glo", "regglo")}
}
\value{
MAgPIE object containing costs for AEI [million US$17]
}
\description{
reads AEI costs entering the objective function from a MAgPIE gdx file
}
\examples{

  \dontrun{
    x <- CostsAEI(gdx)
  }

}
\author{
Felicitas Beier
}
