% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CostsFertilizer.R
\name{CostsFertilizer}
\alias{CostsFertilizer}
\title{CostsFertilizer}
\usage{
CostsFertilizer(gdx, file = NULL, level = "regglo", disagg = TRUE)
}
\arguments{
\item{gdx}{GDX file}

\item{file}{a file name the output should be written to using write.magpie}

\item{level}{Level of regional aggregation ("reg", "glo", "regglo")}

\item{disagg}{whether costs should be disaggregated into the different crop types}
}
\value{
MAgPIE object containing fertilizer costs [million US$17]
}
\description{
reads costs entering the objective function from a MAgPIE gdx file
}
\examples{

  \dontrun{
    x <- CostsFertilizer(gdx)
  }

}
\author{
Debbora Leip
}
