% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/costsPresolve.R
\name{costsPresolve}
\alias{costsPresolve}
\title{costsPresolve}
\usage{
costsPresolve(gdx, file = NULL, level = "reg")
}
\arguments{
\item{gdx}{GDX file}

\item{file}{a file name the output should be written to using write.magpie}

\item{level}{Level of regional aggregation; "reg" (regional), "glo" (global), "regglo" (regional and global) or any other aggregation level defined in superAggregate}
}
\value{
A MAgPIE object containing the goal function costs in presolve mode [million US$17]
}
\description{
reads presovle costs (i.e. without bioenergy demand) entering the objective function from a MAgPIE gdx file
}
\details{
Presolve is without bioenergy demand. Hence costs from a MAgPIE run with bioenergy demand minus costs from presolve reflect costs that can be attributed to bioenergy production
}
\examples{
  \dontrun{
    x <- costsPresolve(gdx)
  }
}
\author{
Florian Humpenoeder
}
