% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/costsProductionCrops.R
\name{costsProductionCrops}
\alias{costsProductionCrops}
\title{costsProductionCrops}
\usage{
costsProductionCrops(
  gdx,
  file = NULL,
  level = "regglo",
  type = "investment",
  dir = "."
)
}
\arguments{
\item{gdx}{GDX file}

\item{file}{a file name the output should be written to using write.magpie}

\item{level}{Level of regional aggregation; "reg" (regional), "glo" (global),
"regglo" (regional and global) or any other aggregation level defined in gdxAggregate}

\item{type}{Type of reporting, either "annuity" or total "investments"}

\item{dir}{directory with mapping for disaggregation to higher resolutions}
}
\value{
A MAgPIE object containing values related with costs for crops production
per ton produced [million US$17/tDM]
}
\description{
Reads data to calculate production costs for crops, costs related with
investments are reported as annual average for both types (annuity, investment).
}
\examples{
\dontrun{
x <- costsProductionCrops(gdx)
}

}
\author{
Edna Molina Bacca
}
