% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CostsWithoutIncentives.R
\name{CostsWithoutIncentives}
\alias{CostsWithoutIncentives}
\title{CostsWithoutIncentives}
\usage{
CostsWithoutIncentives(gdx, file = NULL, level = "regglo")
}
\arguments{
\item{gdx}{GDX file}

\item{file}{a file name the output should be written to using write.magpie}

\item{level}{aggregation level, reg, glo or regglo}
}
\description{
calculates agricultural costs without taxes and incentives (i.e. GHG taxes and BII incentives)
}
\examples{
\dontrun{
x <- CostsWithoutIncentives(gdx)
}

}
\author{
David M Chen
}
