% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CostTransport.R
\name{CostTransport}
\alias{CostTransport}
\title{CostTransport}
\usage{
CostTransport(gdx, file = NULL, level = "cell", sum = FALSE)
}
\arguments{
\item{gdx}{GDX file}

\item{file}{a file name the output should be written to using write.magpie}

\item{level}{Level of regional aggregation; "reg" (regional), "glo" (global), "regglo" (regional and global) or any other aggregation level defined in superAggregate}

\item{sum}{total costs (TRUE) or detailed costs (FALSE)}
}
\value{
A MAgPIE object containing the transport costs [million US$05]
}
\description{
reads costs entering the objective function from a MAgPIE gdx file
}
\examples{

  \dontrun{
    x <- CostTransport(gdx)
  }

}
\author{
David Chen
}
