% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/croparea.R
\name{croparea}
\alias{croparea}
\title{croparea}
\usage{
croparea(
  gdx,
  file = NULL,
  level = "reg",
  products = "kcr",
  product_aggr = TRUE,
  water_aggr = TRUE,
  physical = TRUE
)
}
\arguments{
\item{gdx}{GDX file}

\item{file}{a file name the output should be written to using write.magpie}

\item{level}{Level of regional aggregation; "reg" (regional), "glo" (global),
"regglo" (regional and global) or
any other aggregation level defined in gdxAggregate}

\item{products}{Selection of products (either by naming products, e.g. "tece", or naming a set,e.g."kcr")}

\item{product_aggr}{aggregate over products or not (boolean)}

\item{water_aggr}{aggregate irrigated and non-irriagted production or not (boolean).}

\item{physical}{if TRUE (default) physical (cultivated) area is returned; if FALSE
harvested area, i.e. physical area scaled by the multicropping ratio.}
}
\value{
production as MAgPIE object (unit depends on attributes)
}
\description{
reads croparea out of a MAgPIE gdx file. Croparea excludes fallow land.
}
\examples{
\dontrun{
x <- croparea(gdx)
}

}
\seealso{
\code{\link{reportCroparea}}
}
\author{
Jan Philipp Dietrich, Florian Humpenoeder
}
