% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CropareaDiversityIndex.R
\name{CropareaDiversityIndex}
\alias{CropareaDiversityIndex}
\title{CropareaDiversityIndex}
\usage{
CropareaDiversityIndex(gdx, level = "reg", measurelevel = "cell")
}
\arguments{
\item{gdx}{GDX file}

\item{level}{Level of regional aggregation; "reg" (regional), "glo" (global),
"regglo" (regional and global) or
any other aggregation level defined in superAggregate}

\item{measurelevel}{level at which diversity is measured. "cell" means diversity
is measured at cellular level, even if lateron average diversity is aggregated
to regional level.}
}
\value{
MAgPIE object (unit depends on attributes)
}
\description{
calculates an index that measures the gini for croparea diversity
}
\examples{
\dontrun{
x <- CropareaDiversityIndex(gdx)
}

}
\seealso{
\code{\link{CropareaDiversityIndex}}
}
\author{
Benjamin Leon Bodirsky
}
