% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cropland.R
\name{cropland}
\alias{cropland}
\title{cropland}
\usage{
cropland(gdx, level = "reg", types = NULL, sum = FALSE)
}
\arguments{
\item{gdx}{GDX file}

\item{level}{Level of regional aggregation; "reg" (regional), "glo" (global), "regglo" (regional and global)
or any other aggregation level defined in gdxAggregate}

\item{types}{NULL or a vector of strings. If NULL, all land types are used. Options are
"crop_area", "crop_fallow", "crop_treecover"}

\item{sum}{determines whether output should be land-type-specific (FALSE) or aggregated over all types (TRUE).}
}
\value{
land as MAgPIE object (Mha)
}
\description{
reads cropland out of a MAgPIE gdx file. Cropland includes croparea plus fallow plus cropland with tree cover
}
\examples{
\dontrun{
x <- cropland(gdx)
}

}
\seealso{
\code{\link{reportLandUse}}
}
\author{
Benjamin Leon Bodirsky
}
