% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/croplandSCM.R
\name{croplandSCM}
\alias{croplandSCM}
\title{croplandSCM}
\usage{
croplandSCM(gdx, level = "reg", crop_aggr = TRUE)
}
\arguments{
\item{gdx}{GDX file}

\item{level}{aggregation level, "reg", "glo", "regglo", or "cell"}

\item{crop_aggr}{aggregate over crop types (TRUE) or report by crop (FALSE)}
}
\value{
SCM area as MAgPIE object (mio. ha)
}
\description{
reads cropland area under soil carbon management (SCM) out of a
             MAgPIE gdx file (Mha)
}
\examples{
\dontrun{
x <- croplandSCM(gdx)
}

}
\author{
Kristine Karstens
}
