% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/croplandTreeCover.R
\name{croplandTreeCover}
\alias{croplandTreeCover}
\title{croplandTreeCover}
\usage{
croplandTreeCover(gdx, level = "reg", sum_ac = TRUE, debugMode = FALSE)
}
\arguments{
\item{gdx}{GDX file}

\item{level}{aggregation level, reg, glo or regglo, cell or grid}

\item{sum_ac}{sum over age classes TRUE / FALSE}

\item{debugMode}{debug mode TRUE makes some consistency checks
between estimates for different resolutions}
}
\description{
calculates tree cover on cropland (Mha) from a MAgPIE gdx file
}
\examples{
\dontrun{
x <- fallow(gdx)
}

}
\author{
Florian Humpenoeder
}
