% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cshare.R
\name{cshare}
\alias{cshare}
\title{cshare}
\usage{
cshare(
  gdx,
  file = NULL,
  level = "reg",
  reference = "actual",
  noncrop_aggr = TRUE
)
}
\arguments{
\item{gdx}{GDX file}

\item{file}{a file name the output should be written to using write.magpie}

\item{level}{Level of regional aggregation;
"reg" (regional),
"glo" (global),
"regglo" (regional and global) or
any other aggregation level defined in superAggregate}

\item{reference}{default is "actual" (cshare in actual carbon stocks).
Other option is "target" (cshare in target carbon stocks).}

\item{noncrop_aggr}{aggregate non cropland types to 'noncropland'
(if FALSE all land types of pools59 will be reported)}
}
\value{
A MAgPIE object containing som values
}
\description{
Calculates soil carbon share in relation to potential natural
             vegetation based on a MAgPIE gdx file
}
\examples{
\dontrun{
x <- cshare(gdx)
}

}
\author{
Kristine Karstens
}
