% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demand.R
\name{demand}
\alias{demand}
\title{demand}
\usage{
demand(
  gdx,
  file = NULL,
  level = "reg",
  products = "kall",
  product_aggr = FALSE,
  attributes = "dm",
  type = NULL,
  type_aggr = FALSE
)
}
\arguments{
\item{gdx}{GDX file}

\item{file}{a file name the output should be written to using write.magpie}

\item{level}{Level of regional aggregation ("reg", "glo", "regglo")}

\item{products}{Selection of products (either by naming products, e.g. "tece", or naming a set,e.g."kcr")}

\item{product_aggr}{aggregate over products or not (boolean, default FALSE)}

\item{attributes}{dry matter: Mt ("dm"), gross energy: PJ ("ge"), reactive nitrogen: Mt ("nr"), phosphor: Mt ("p"), potash: Mt ("k"), wet matter: Mt ("wm"). Can also be a vector.}

\item{type}{Demand type(s): "food", "feed", "processed", "other_util", "bioenergy", "seed", "waste", "dom_balanceflow; NULL returns all types}

\item{type_aggr}{aggregate over demand types or not (boolean, default FALSE)}
}
\value{
demand as MAgPIE object (Unit depends on attributes)
}
\description{
Calculates MAgPIE demand out of a gdx file
}
\details{
Demand definitions are equivalent to FAO CBS categories
}
\examples{

  \dontrun{
    x <- demand(level="regglo", products="kcr")
  }

}
\author{
Benjamin Leon Bodirsky, Abhijeet Mishra, Miodrag Stevanovic
}
