% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demandBioenergy.R
\name{demandBioenergy}
\alias{demandBioenergy}
\title{demandBioenergy}
\usage{
demandBioenergy(gdx, file = NULL, level = "reg", sum = FALSE, round = NULL)
}
\arguments{
\item{gdx}{GDX file}

\item{file}{a file name the output should be written to using write.magpie}

\item{level}{Level of regional aggregation; "reg" (regional), "glo" (global), "regglo" (regional and global) or any other aggregation level defined in superAggregate}

\item{sum}{1st and 2nd generation bioenergy demand (FALSE) or total bioenergy demand (TRUE)}

\item{round}{NULL or number of digits}
}
\value{
A MAgPIE object containing bioenergy demand in EJ/yr
}
\description{
reads bioenergy demand from a MAgPIE gdx file
}
\examples{

  \dontrun{
    x <- demandBioenergy(gdx)
  }
}
\author{
Jan Philipp Dietrich, Florian Humpenoeder
}
