% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disaggregateLandConservation.R
\name{disaggregateLandConservation}
\alias{disaggregateLandConservation}
\title{disaggregateLandConservation}
\usage{
disaggregateLandConservation(gdx, cfg, mapping, wdpaHr, conservationPrioHr)
}
\arguments{
\item{gdx}{character, path to a fulldata.gdx}

\item{cfg}{list, config data usually obtained with gms::loadConfig("config.yml")}

\item{mapping}{dataframe, must include columns cell and cluster}

\item{wdpaHr}{magclass, World Database on Protected Areas data in high resolution}

\item{conservationPrioHr}{magclass, high resolution data on conservation priority areas}
}
\value{
magclass, high resolution land conservation area
}
\description{
Read land conservation data from a fulldata.gdx and disaggregate to high resolution (0.5 deg).
}
\author{
Patrick v. Jeetze, Pascal Sauer
}
