% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discountRates.R
\name{discountRates}
\alias{discountRates}
\title{discountRates}
\usage{
discountRates(gdx, file = NULL, level = "reg")
}
\arguments{
\item{gdx}{GDX file}

\item{file}{a file name the output should be written to using write.magpie}

\item{level}{Level of regional aggregation; "reg" (regional), "glo" (global), "regglo" (regional and global) or any other aggregation level defined in superAggregate}
}
\value{
A MAgPIE object containing discount rates used in the model
}
\description{
reads discount rates from a MAgPIE gdx file
}
\examples{

  \dontrun{
    x <- discountRates(gdx)
  }

}
\author{
Xiaoxi Wang
}
