% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Emissions.R
\name{Emissions}
\alias{Emissions}
\title{Emissions}
\usage{
Emissions(
  gdx,
  file = NULL,
  level = "reg",
  type = "co2_c",
  unit = "element",
  subcategories = FALSE,
  cumulative = FALSE,
  lowpass = NULL,
  inorg_fert_split = TRUE
)
}
\arguments{
\item{gdx}{GDX file}

\item{file}{a file name the output should be written to using write.magpie}

\item{level}{Level of regional aggregation; "reg" (regional), "glo" (global), "regglo" (regional and global) or any other aggregation level defined in superAggregate}

\item{type}{emission type(s): "co2_c", "n2o_n" or "ch4"}

\item{unit}{"element", "gas", "GWP100AR5", "GWP100AR6", "GWP*AR5", or "GWP*AR6"
"element": co2_c in Mt C/yr, n2o_n in Mt N/yr, ch4 in Mt CH4/yr
"gas":     co2_c Mt CO2/yr, n2o_n in Mt NO2/yr, ch4 in Mt CH4/yr
"GWP":     co2_c in Mt CO2/yr, n2o_n in Mt CO2eq/yr, ch4 in Mt CO2eq/yr}

\item{subcategories}{FALSE (default) or TRUE}

\item{cumulative}{Logical; Determines if emissions are reported annually (FALSE) or cumulative (TRUE). The starting point for cumulative emissions is y1995.}

\item{lowpass}{number of lowpass filter iterations}

\item{inorg_fert_split}{if TRUE then inorganic fertilizer emissions are further disaggregated into pasture- and cropland-related emissions.}
}
\value{
emissions as MAgPIE object (unit depends on \code{unit})
}
\description{
reads GHG emissions out of a MAgPIE gdx file
}
\examples{

  \dontrun{
    x <- Emissions(gdx)
  }
}
\author{
Florian Humpenoeder, Benjamin Leon Bodirsky
}
