% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EmissionsBeforeTechnicalMitigation.R
\name{EmissionsBeforeTechnicalMitigation}
\alias{EmissionsBeforeTechnicalMitigation}
\title{EmissionsBeforeTechnicalMitigation}
\usage{
EmissionsBeforeTechnicalMitigation(
  gdx,
  file = NULL,
  level = "reg",
  type = "co2_c",
  unit = "element",
  subcategories = FALSE
)
}
\arguments{
\item{gdx}{GDX file}

\item{file}{a file name the output should be written to using write.magpie}

\item{level}{Level of regional aggregation; "reg" (regional), "glo" (global), "regglo" (regional and global) or any other aggregation level defined in superAggregate}

\item{type}{emission type(s): "co2_c", "n2o_n" or "ch4" and in the case of unit="gas" "co2" and "n2o"}

\item{unit}{"element", "gas" or "co2eq"; "element": co2_c in Mt C/yr, n2o_n in Mt N/yr, ch4 in Mt CH4/yr; "gas": co2_c Mt CO2/yr, n2o_n in Mt NO2/yr, ch4 in Mt CH4/yr; "co2eq": co2_c in Mt CO2/yr, n2o_n in Mt CO2eq/yr, ch4 in Mt CO2eq/yr}

\item{subcategories}{FALSE (default) or TRUE}
}
\value{
emissions as MAgPIE object (unit depends on \code{unit})
}
\description{
reads GHG emissions before technical abatement out of a MAgPIE gdx file. Technical abatement includes all abatement done in the MACC curves, but exclude endogenous mitigation. These emissions are NOT the standard reporting emissions, but used for special purposes like remind-magpie coupling.
}
\examples{

  \dontrun{
    x <- EmissionsBeforeTechnicalMitigation(gdx)
  }
}
\author{
Florian Humpenoeder; Benjamin Leon Bodirsky
}
