% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expectVariablesPresent.R
\name{expectVariablesPresent}
\alias{expectVariablesPresent}
\title{expectVariablesPresent}
\usage{
expectVariablesPresent(report, variableNames)
}
\arguments{
\item{report}{A MAgPIE object containing the report to check.}

\item{variableNames}{A character vector of variable names that are expected
to be present in \code{report}.}
}
\value{
\code{NULL} invisibly. Issues a warning if expected variables are not present.
}
\description{
Checks whether a set of expected variable names is present in a
  MAgPIE report object. Variable names are normalized in that summation symbols
  are removed.
}
\author{
Patrick Rein
}
