% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expenditureIndexFood.R
\name{expenditureIndexFood}
\alias{expenditureIndexFood}
\title{expenditureIndexFood}
\usage{
expenditureIndexFood(
  gdx,
  file = NULL,
  level = "reg",
  products = "kfo",
  basketyear = "y2010",
  baseyear = "y2010",
  round = TRUE,
  ghgtax = TRUE,
  valueAdded = FALSE
)
}
\arguments{
\item{gdx}{GDX file}

\item{file}{File the output should be written to using write.magpie}

\item{level}{Level of regional aggregation; "reg" (regional),
"glo" (global), "regglo" (regional and global)
or any other aggregation level defined in mapping}

\item{products}{Selection of products (either by naming products,
e.g. "tece", or naming a set, e.g."kcr")}

\item{basketyear}{Year of reference food basket
(should be in the past for comparison of different runs to
have identical and comparable food basket)}

\item{baseyear}{Baseyear of the price index}

\item{round}{Rounded result (TRUE or FALSE)}

\item{ghgtax}{Correction of food price expenditure for ghg emission costs
(TRUE or FALSE)}

\item{valueAdded}{whether to include value added}
}
\value{
A MAgPIE object containing food price expenditure index, in 2017Int$PPP
}
\description{
calculates food expenditure index (baseyear = 100)
             corrected for ghg emission costs based on a MAgPIE gdx file
}
\examples{
\dontrun{
x <- expenditureIndexFood(gdx)
}

}
\author{
Felicitas Beier, David Chen
}
