% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractManureFuel.R
\name{extractManureFuel}
\alias{extractManureFuel}
\title{extractManureFuel}
\usage{
extractManureFuel(gdx, file = NULL)
}
\arguments{
\item{gdx}{GDX file}

\item{file}{a file name the output should be written to using write.magpie}
}
\value{
MAgPIE object with manure fuel at ISO level in PJ.
  Dimension 1: ISO country, Dimension 2: year, Dimension 3: "manurefuel"
}
\description{
Extracts manure used as direct combustion fuel from a MAgPIE GDX file
at ISO country level. Converts from nitrogen mass (Mt N) to energy (PJ) using
livestock-specific heating values and nitrogen contents from Hoyos-Sebá et al. (2024).
}
\examples{
\dontrun{
  x <- extractManureFuel(gdx)
}
}
\author{
Kristine Karstens
}
