% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractProcessingWoodResidues.R
\name{extractProcessingWoodResidues}
\alias{extractProcessingWoodResidues}
\title{extractProcessingWoodResidues}
\usage{
extractProcessingWoodResidues(gdx, file = NULL)
}
\arguments{
\item{gdx}{GDX file}

\item{file}{a file name the output should be written to using write.magpie}
}
\value{
MAgPIE object with potential wood processing residues at ISO level in PJ.
  Dimension 1: ISO country, Dimension 2: year, Dimension 3: "processingresidues"
}
\description{
Extracts the potential of wood processing residues (sawmill byproducts)
from a MAgPIE GDX file at ISO country level and converts to energy (PJ).
Processing residues are computed from industrial roundwood demand using a
30% residue rate (see \code{reportProcessingResiduesForestry} in magpie4).
}
\examples{
\dontrun{
  x <- extractProcessingWoodResidues(gdx)
}
}
\author{
Kristine Karstens
}
