% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractWoodFuel.R
\name{extractWoodFuel}
\alias{extractWoodFuel}
\title{extractWoodFuel}
\usage{
extractWoodFuel(gdx, file = NULL)
}
\arguments{
\item{gdx}{GDX file}

\item{file}{a file name the output should be written to using write.magpie}
}
\value{
MAgPIE object with wood fuel demand at ISO level in PJ.
  Dimension 1: ISO country, Dimension 2: year, Dimension 3: "woodfuel"
}
\description{
Extracts wood fuel demand from a MAgPIE GDX file at ISO country level
and converts from volumetric units (Mm³) to energy (PJ).
}
\examples{
\dontrun{
  x <- extractWoodFuel(gdx)
}
}
\author{
Kristine Karstens
}
