% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FactorCosts.R
\name{FactorCosts}
\alias{FactorCosts}
\title{FactorCosts}
\usage{
FactorCosts(gdx, products = "kli", file = NULL, level = "regglo")
}
\arguments{
\item{gdx}{GDX file}

\item{products}{products for which factor costs should be reported ("kli",
"kres", "fish", or "pastures")}

\item{file}{a file name the output should be written to using write.magpie}

\item{level}{Level of regional aggregation ("reg", "glo", "regglo")}
}
\value{
MAgPIE object containing factor costs [million US$05]
}
\description{
reads factor costs for livestock, residues or pasture entering
the objective function from a MAgPIE gdx file, and splits into labor costs
and capital costs in case the gdx file comes from a run using the sticky
factor costs realization
}
\examples{
\dontrun{
x <- FactorCosts(gdx)
}

}
\author{
Debbora Leip
}
