% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorCostShare.R
\name{factorCostShares}
\alias{factorCostShares}
\title{factorCostShares}
\usage{
factorCostShares(gdx, products = "kli", level = "reg", file = NULL)
}
\arguments{
\item{gdx}{GDX file}

\item{products}{products for which cost shares should be reported, kcr or kli}

\item{level}{spatial aggregation to report employment ("reg", "glo" or "regglo")}

\item{file}{a file name the output should be written to using write.magpie}
}
\value{
labor and capital cost share out of factor costs
}
\description{
returns labor and capital cost share out of factor costs (i.e. labor + capital)
}
\examples{
\dontrun{
x <- factorCostShares(gdx)
}
}
\author{
Debbora Leip
}
