% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feed.R
\name{feed}
\alias{feed}
\title{feed}
\usage{
feed(gdx, file = NULL, level = "reg", detail = T, nutrient = "dm")
}
\arguments{
\item{gdx}{GDX file}

\item{file}{a file name the output should be written to using write.magpie}

\item{level}{Level of regional aggregation; "reg" (regional), "glo" (global), "regglo" (regional and global) or any other aggregation level defined in superAggregate}

\item{detail}{if FALSE, only total feed demand per animal type is calculated without details on the type of feed}

\item{nutrient}{The nutrient in which the results shall be calculated}
}
\value{
feed demand by animal type as MAgPIE object (unit depends on selected nutrient attributes)
}
\description{
calculates feed demand by animal type out of a gdx file
}
\examples{
  \dontrun{
    x <- feed(gdx)
  }

}
\author{
Isabelle Weindl
}
