% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FoodDemandModuleConsumerPrices.R
\name{FoodDemandModuleConsumerPrices}
\alias{FoodDemandModuleConsumerPrices}
\title{FoodDemandModuleConsumerPrices}
\usage{
FoodDemandModuleConsumerPrices(gdx)
}
\arguments{
\item{gdx}{GDX file}
}
\value{
magpie object
}
\description{
Calculates food prices that enter demand module
}
\examples{

  \dontrun{
    x <- FoodDemandModuleConsumerPrices(gdx)
  }

}
\author{
Benjamin Leon Bodirsky
}
