% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FoodExpenditure.R
\name{FoodExpenditure}
\alias{FoodExpenditure}
\title{FoodExpenditure}
\usage{
FoodExpenditure(
  gdx,
  level = "reg",
  after_shock = TRUE,
  products = "kfo",
  product_aggr = TRUE,
  per_capita = TRUE,
  valueAdded = FALSE
)
}
\arguments{
\item{gdx}{GDX file}

\item{level}{spatial aggregation. can be "iso","reg","regglo","glo"}

\item{after_shock}{FALSE is using the exogenous real income and the prices before a shock, TRUE is using the endogeenous real income that takes into account food price change on real income, "after_price_before_demand" takes into account price changes on real income, but assumes no demand adjustment}

\item{products}{selected products or sets of products}

\item{product_aggr}{if true, aggregation over products}

\item{per_capita}{per capita or total population}

\item{valueAdded}{whether to add the value-added 
marketing margin to the total expenditures}
}
\value{
magpie object with per capita consumption
}
\description{
Calculates the food expenditure in USD per year
}
\examples{

  \dontrun{
    x <- FoodExpenditure(gdx)
  }

}
\author{
Benjamin Leon Bodirsky
}
