% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FoodExpenditureShare.R
\name{FoodExpenditureShare}
\alias{FoodExpenditureShare}
\title{FoodExpenditureShare}
\usage{
FoodExpenditureShare(
  gdx,
  level = "reg",
  after_shock = TRUE,
  products = "kfo",
  product_aggr = T
)
}
\arguments{
\item{gdx}{GDX file}

\item{level}{spatial aggregation. can be "iso","reg","regglo","glo"}

\item{after_shock}{FALSE is using the exogenous real income and the prices before a shock, TRUE is using the endogeenous real income that takes into account food price change on real income}

\item{products}{selected products or sets of products}

\item{product_aggr}{if true, aggregation over products}
}
\value{
magpie object with per capita consumption
}
\description{
Calculates the livestock share from the food demand model
}
\examples{

  \dontrun{
    x <- FoodExpenditureShare(gdx)
  }

}
\author{
Benjamin Leon Bodirsky
}
