% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foodmodelstat.R
\name{foodmodelstat}
\alias{foodmodelstat}
\title{foodmodelstat}
\usage{
foodmodelstat(gdx, file = NULL)
}
\arguments{
\item{gdx}{GDX file}

\item{file}{a file name the output should be written to using write.magpie}
}
\value{
A MAgPIE object containing number of iterations and convergence information for each time step
}
\description{
MAgPIE food model statistics with information about convergence and number of iterations
}
\examples{
\dontrun{
x <- foodmodelstat(gdx)
}

}
\author{
Jan Philipp Dietrich
}
