% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ForestYield.R
\name{ForestYield}
\alias{ForestYield}
\title{ForestYield}
\usage{
ForestYield(gdx, file = NULL, level = "cell")
}
\arguments{
\item{gdx}{GDX file}

\item{file}{a file name the output should be written to using write.magpie}

\item{level}{Level of regional aggregation; "cell", "reg" (regional), "glo" (global), "regglo" (regional and global) or any secdforest aggregation level defined in superAggregate}
}
\value{
Average growing stock density of harvested areas in m3/ha
}
\description{
Calculates the average growing stock density of harvested forest areas (m3/ha).
This is the ratio of timber production volume to harvested area, i.e. the average standing
volume removed per hectare harvested. Values are comparable to growing stock (typically
100-300 m3/ha) and should NOT be confused with forestry yield in the MAI sense (5-20 m3/ha/yr).
}
\details{
Average growing stock of harvested areas = timber production (m3) / harvested area (ha).
Differences to overall growing stock arise from selective harvesting of specific age classes.
}
\examples{

  \dontrun{
    x <- ForestYield(gdx)
  }
}
\author{
Abhijeet Mishra, Florian Humpenoeder
}
