% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDirectory.R
\name{getDirectory}
\alias{getDirectory}
\title{getDirectory}
\usage{
getDirectory(dir, spamfiledirectory)
}
\arguments{
\item{dir}{new directory input}

\item{spamfiledirectory}{old directory input}
}
\value{
a directory
}
\description{
support function to properly merge deprecated spamfiledirectory and dir input
}
