% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getReportDemandStandalone.R
\name{getReportDemandStandalone}
\alias{getReportDemandStandalone}
\title{getReportDemandStandalone}
\usage{
getReportDemandStandalone(
  gdx,
  file = NULL,
  scenario = NULL,
  detail = FALSE,
  ...
)
}
\arguments{
\item{gdx}{GDX file}

\item{file}{a file name the output should be written to using write.report. If NULL the report is returned instead as a MAgPIE object.}

\item{scenario}{Name of the scenario used for the list-structure of a reporting object (x$scenario$MAgPIE). If NULL the report is returned instead as a MAgPIE object.}

\item{detail}{Crop specific (TRUE) or aggregated outputs (FALSE)}

\item{...}{additional arguments for write.report. Will only be taken into account if argument "file" is not NULL.}
}
\value{
A MAgPIE object containing the report in the case that "file" is NULL.
}
\description{
Puts together a report based on a MAgPIE gdx file
}
\examples{

  \dontrun{
    x <- getReportDemandStandalone(gdx)
  }

}
\author{
Florian Humpenoeder
}
