% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getReportDietaryIndicators.R
\name{getReportDietaryIndicators}
\alias{getReportDietaryIndicators}
\title{getReportDietaryIndicators}
\usage{
getReportDietaryIndicators(gdx, scenario)
}
\arguments{
\item{gdx}{filepath of the GDX file}

\item{scenario}{character string describing the scenario configuration}
}
\value{
list of data.frames for the dietary indicators
}
\description{
reports dietary indicators on the country level. These are formatted as data.frames describing:
1. population, anthropometrics, and intake
2. caloric consumption by food category (i.e. with food waste)
}
\examples{

  \dontrun{
    getReportDietaryIndicators(gdx, scenario)
  }

}
\author{
Michael Crawford, Felicitas Beier, Benjamin Bodirsky
}
