% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getReportFableScenathon.R
\name{getReportFableScenathon}
\alias{getReportFableScenathon}
\title{getReportFableScenathon}
\usage{
getReportFableScenathon(gdx, file = NULL, iso = NULL)
}
\arguments{
\item{gdx}{a GDX file}

\item{file}{a file name the output should be written to using write.report.
If `NULL` the report is returned instead as a MAgPIE object. For the easier
reporting in Scenathon tabs, a .csv file extension is recommenended.}

\item{iso}{country/region selection. Default `NULL`, i.e. all `regglo` reporting}
}
\description{
Collects outputs from MAgPIE runs for FABLE Scenathon.
}
\examples{

  \dontrun{
    x <- getReportFableScenathon(gdx, file = "magpie2scenathon.csv", iso = "IND")
  }

}
\author{
Miodrag Stevanovic
}
