% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getReportFSECCosts.R
\name{getReportFSECCosts}
\alias{getReportFSECCosts}
\title{getReportFSECCosts}
\usage{
getReportFSECCosts(gdx, reportOutputDir = NULL, scenario = NULL)
}
\arguments{
\item{gdx}{a GDX file}

\item{reportOutputDir}{a folder name for the output to be written to. If NULL the report is not saved to
disk, and only returned to the calling function.}

\item{scenario}{the name of the scenario used. If NULL the report is not saved to disk, and only returned to the
calling function.}
}
\value{
A .csv containing the summed output of reportCostsAccounting on the region level
}
\description{
Reports cost indicators for the FSEC project
}
\examples{

  \dontrun{
    x <- getReportFSECCosts(gdx)
  }

}
\author{
Michael Crawford
}
