% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getReportFSECSimonDietz.R
\name{getReportFSECSimonDietz}
\alias{getReportFSECSimonDietz}
\title{getReportFSECSimonDietz}
\usage{
getReportFSECSimonDietz(
  magpieOutputDir,
  reportOutputDir = NULL,
  scenario = NULL
)
}
\arguments{
\item{magpieOutputDir}{a magpie output directory which contains all the files associate with the given scenario}

\item{reportOutputDir}{a folder name for the output to be written to. If NULL the report is not saved to
disk, and only returned to the calling function.}

\item{scenario}{the name of the scenario used. If NULL the report is not saved to disk, and only returned to the
calling function.}
}
\value{
A list of reports
}
\description{
Collects reports for Simon Dietz' social welfare function analysis
}
\examples{

  \dontrun{
    x <- getReportFSECSimonDietz(magpieOutputDir)
  }

}
\author{
Michael Crawford
}
