% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getReportGridEmissions.R
\name{getReportGridEmissions}
\alias{getReportGridEmissions}
\title{getReportGridEmissions}
\usage{
getReportGridEmissions(
  magpieOutputDir,
  reportOutputDir = NULL,
  scenario = NULL
)
}
\arguments{
\item{magpieOutputDir}{a magpie output directory which contains a mapping file (clustermap*.rds) for the
disaggregation of grid output}

\item{reportOutputDir}{a folder name for the output to be written to. If NULL the report is not saved to
disk, and only returned to the calling function.}

\item{scenario}{the name of the scenario used. If NULL the report is not saved to disk, and only returned to the
calling function.}
}
\value{
A list of MAgPIE objects containing the reports
}
\description{
Reports all grid-level emissions available for a magpie scenario
}
\examples{

  \dontrun{
    x <- getReportGridEmissions(gdx, magpieOutputDir)
  }

}
\author{
Michael Crawford
}
