% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getReportGridINMS.R
\name{getReportGridINMS}
\alias{getReportGridINMS}
\title{getReportGridINMS}
\usage{
getReportGridINMS(
  gdx,
  reportOutputDir = NULL,
  magpieOutputDir,
  scenario = NULL,
  filter = c(2, 7),
  version = "v13"
)
}
\arguments{
\item{gdx}{GDX file}

\item{reportOutputDir}{Directory in which the reports are to be saved. If NULL, a list of reports (MAgPIE objects) is returned instead}

\item{magpieOutputDir}{Directory containing the MAgPIE run which is to be processed}

\item{scenario}{Name of the scenario used for the list-structure of a reporting object (x$scenario$MAgPIE). If NULL a list of reports (MAgPIE objects) is returned instead.}

\item{filter}{Modelstat filter. Here you have to set the modelstat values for which results should be used. All values for time steps in which the modelstat is different or for which one of the previous modelstats were different are set to NA.}

\item{version}{Version number for this analysis}
}
\value{
A list of reports (MAgPIE objects)
}
\description{
Generates and saves a list of reports relevant to the INMS context
}
\examples{

  \dontrun{
    x <- getReportGridINMS(gdx)
  }

}
\author{
Benjamin Leon Bodirsky, Florian Humpenoeder, Michael Crawford
}
