% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getReportGridPollutants.R
\name{getReportGridPollutants}
\alias{getReportGridPollutants}
\title{getReportGridPollutants}
\usage{
getReportGridPollutants(
  gdx,
  reportOutputDir,
  magpieOutputDir,
  scenario,
  filter = c(2, 7)
)
}
\arguments{
\item{gdx}{GDX file}

\item{reportOutputDir}{a folder name that the output should be written to using write.report. If NULL the report is not saved to disk, and only returned as a magpie object.}

\item{magpieOutputDir}{magpie output directory which contains a mapping file (clustermap*.rds) for the disaggregation of grid output}

\item{scenario}{name of the scenario used for the list-structure of a reporting object (x$scenario$MAgPIE). If NULL the report is not saved to disk, and only returned as a magpie object.}

\item{filter}{Modelstat filter. Here you have to set the modelstat values for which results should be used. All values for time steps in which the modelstat is different or for which one of the previous modelstats were different are set to NA.}
}
\value{
A list of MAgPIE objects containing the reports
}
\description{
Reports key environmental pollution indicators, including nitrogen budgets for croplands and pastures
}
\examples{

  \dontrun{
    x <- getReportGridPollutants(gdx)
  }

}
\author{
Michael Crawford
}
