% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getReportMAgPIE2GAINS.R
\name{getReportMAgPIE2GAINS}
\alias{getReportMAgPIE2GAINS}
\title{getReportMAgPIE2GAINS}
\usage{
getReportMAgPIE2GAINS(
  gdx,
  folder = NULL,
  scenario = NULL,
  filter = c(2, 7),
  dir = ".",
  ...
)
}
\arguments{
\item{gdx}{GDX file}

\item{folder}{a folder name the output should be written to using write.report. If NULL the report is returned instead as a MAgPIE object.}

\item{scenario}{Name of the scenario used for the list-structure of a reporting object (x$scenario$MAgPIE). If NULL the report is returned instead as a MAgPIE object.}

\item{filter}{Modelstat filter. Here you have to set the modelstat values for which results should be used. All values for time steps in which the modelstat is different or for which one of the previous modelstats were different are set to NA.}

\item{dir}{for gridded outputs: magpie output directory which contains a mapping file (rds or spam) disaggregation}

\item{...}{additional arguments for write.report. Will only be taken into account if argument "file" is not NULL.}
}
\value{
A MAgPIE object containing the report in the case that "file" is NULL.
}
\description{
Puts together a report for the IIASA GAINS model based on a MAgPIE gdx file
}
\author{
Benjamin Leon Bodirsky, Florian Humpenoeder
}
