% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getReportMAgPIE2REMIND.R
\name{getReportMAgPIE2REMIND}
\alias{getReportMAgPIE2REMIND}
\title{getReportMAgPIE2REMIND}
\usage{
getReportMAgPIE2REMIND(gdx, file = NULL, scenario = NULL)
}
\arguments{
\item{gdx}{GDX file}

\item{file}{A file name the output should be written to using write.report.
If NULL the report is returned instead as a MAgPIE object.}

\item{scenario}{Name of the scenario used for the list-structure of a
reporting object (x$scenario$MAgPIE).
If NULL the report is returned instead as a MAgPIE object.}
}
\value{
A MAgPIE object containing the report.
}
\description{
Based on a MAgPIE gdx file, a report is generated containing
only the variables relevant for the coupling with REMIND. Basically a copy
of getReport, but calling less 'reportXY()' functions. Land-use-Change CO2 uses the
legacy-clearing reframe (the reportEmissions default).
}
\details{
Reports are organized with '|' as level delimiter and summation symbols
         for grouping subcategories into entities e.g. for stackplots.
         Notice the following hints for the summation symbol placement:
\itemize{
  \item Every name should just contain one summation symbol (mostly '+').
  \item The position of the symbol (counted in '|' from left side) will determine the level.
  \item Every subitem containing the same summation symbol in the same level
        with the same supercategory name will be summed.
  \item Items without any summation symbol will be silently ignored.
  \item Items with different summation symbols will be summed up separately.
  \item In most of the cases a summation symbol will be just placed
        before the last level (counted in '|' from left side).
  \item It is helpful to think about which group of items should be stacked in a stackplot.
}
  An example how a summation symbol placement could look like:
  \preformatted{  Toplevel
  Toplevel|+|Item 1
  Toplevel|+|Item 2
  Toplevel|Item 2|+|Subitem 1
  Toplevel|Item 2|+|Subitem 1
  Toplevel|++|Item A
  Toplevel|++|Item B
  Toplevel|Item ?}
}
\examples{
\dontrun{
x <- getReportMAgPIE2REMIND(gdx)
}

}
\author{
Florian Humpenoeder, David Klein
}
