% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grassyld.R
\name{grassyld}
\alias{grassyld}
\title{grassland yields}
\usage{
grassyld(gdx)
}
\arguments{
\item{gdx}{GDX file}
}
\value{
A MAgPIE object containing grassland yields values
}
\description{
Calculates grassland yields based on a MAgPIE gdx file
}
\examples{
\dontrun{
x <- grassyld(gdx)
}

}
\author{
Marcos Alves
}
