% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GrowingStock.R
\name{GrowingStock}
\alias{GrowingStock}
\title{GrowingStock}
\usage{
GrowingStock(gdx, file = NULL, level = "regglo", indicator = "relative")
}
\arguments{
\item{gdx}{GDX file}

\item{file}{a file name the output should be written to using write.magpie}

\item{level}{Level of regional aggregation; "cell", "reg" (regional), "glo" (global), "regglo" (regional and global) or any secdforest aggregation level defined in superAggregate}

\item{indicator}{If the reported numbers are relative (mio m3/ha) or absolute (mio. m3). Default is relative.}
}
\value{
Growing stock in m3 per ha
}
\description{
reads woody growing stock out of a MAgPIE gdx file
}
\details{
Growing stock for producing woody materials consist of growing stock from plantations (forestry), secondary and primary forest as well as other land (natveg)
}
\examples{

\dontrun{
x <- GrowingStock(gdx)
}

}
\author{
Abhijeet Mishra
}
