% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HANNP.R
\name{HANPP}
\alias{HANPP}
\title{HANPP}
\usage{
HANPP(gdx, level = "regglo")
}
\arguments{
\item{gdx}{GDX file}

\item{level}{Level of regional aggregation; "reg" (regional), "glo" (global), "regglo" (regional and global)}
}
\value{
magpie object
}
\description{
calculates the Human Appropriation of Net Primary Production (HANNP)
             based on outputs from MAgPIE gdx file and LPJmL data
}
\examples{
  \dontrun{
    x <- HANNP(gdx)
  }

}
\author{
Felicitas Beier, Isabelle Weindl
}
