% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harvested_area_timber.R
\name{harvested_area_timber}
\alias{harvested_area_timber}
\title{harvested_area_timber}
\usage{
harvested_area_timber(
  gdx,
  file = NULL,
  level = "cell",
  aggregateAgeClasses = TRUE
)
}
\arguments{
\item{gdx}{A fulldata.gdx of a magpie run, usually with endogenous forestry enabled}

\item{file}{a file name the output should be written to using write.magpie}

\item{level}{Level of regional aggregation; "cell", "reg" (regional), "glo"
(global), "regglo" (regional and global) or any secdforest aggregation
level defined in superAggregate}

\item{aggregateAgeClasses}{If TRUE, age classes are aggregated}
}
\value{
Area harvested for wood in Mha per year as a magpie object
}
\description{
Reads wood harvest area separated by source (primforest,
secdforest, forestry, other) and age classes from a gdx.
The data is on cluster level and the unit is Mha per year.
}
\author{
Abhijeet Mishra, Pascal Sauer
}
