% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hourlyLaborCosts.R
\name{hourlyLaborCosts}
\alias{hourlyLaborCosts}
\title{hourlyLaborCosts}
\usage{
hourlyLaborCosts(gdx, level = "reg", file = NULL)
}
\arguments{
\item{gdx}{GDX file}

\item{level}{spatial aggregation to report employment ("iso", "reg", "glo", or "regglo")}

\item{file}{a file name the output should be written to using write.magpie}
}
\value{
hourly labor costs in agriculture
}
\description{
returns hourly labor costs in agriculture from MAgPIE results
}
\examples{
\dontrun{
x <- hourlyLaborCosts(gdx)
}
}
\author{
Debbora Leip
}
