% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hunger.R
\name{Hunger}
\alias{Hunger}
\title{Hunger}
\usage{
Hunger(
  gdx,
  level = "reg",
  after_shock = TRUE,
  calibrated = FALSE,
  share = TRUE
)
}
\arguments{
\item{gdx}{GDX file}

\item{level}{spatial aggregation. can be "iso","reg","regglo","glo"}

\item{after_shock}{FALSE is using the exogenous real income and the prices before a shock, TRUE is using the endogeenous real income that takes into account food price change on real income}

\item{calibrated}{if calibrated is TRUE, kcal values are calibrated to better match historical years}

\item{share}{share of population that is undernourished}
}
\value{
magpie object with hunger (mio people) or hunger share
}
\description{
Calculates the share of people living in hunger.
}
\examples{

  \dontrun{
    x <- Hunger(gdx)
  }

}
\author{
Benjamin Leon Bodirsky
}
