% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/income.R
\name{income}
\alias{income}
\title{income}
\usage{
income(
  gdx,
  file = NULL,
  level = "reg",
  per_capita = TRUE,
  type = "ppp",
  after_shock = FALSE
)
}
\arguments{
\item{gdx}{GDX file}

\item{file}{a file name the output should be written to using write.magpie}

\item{level}{Level of regional aggregation; "reg" (regional), "glo" (global), "regglo" (regional and global) or any other aggregation level defined in superAggregate}

\item{per_capita}{income per capita or aggregated for the total population}

\item{type}{ppp for purchase power parity, mer for market exchange rate}

\item{after_shock}{FALSE is using the exogenous real income, TRUE is using the endogeenous real income that takes into account food price change on real income}
}
\value{
annual income as MAgPIE object (unit depends on per_capita: US$2005 MER/cap/yr (TRUE), US$2005 MER/yr (FALSE))
}
\description{
Calculates income based on a MAgPIE gdx file
}
\examples{

  \dontrun{
    x <- income(gdx)
  }

}
\author{
Florian Humpenoeder, Benjamin Bodirsky
}
