% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Intake.R
\name{Intake}
\alias{Intake}
\title{Intake}
\usage{
Intake(
  gdx,
  file = NULL,
  level = "reg",
  calibrated = TRUE,
  pregnancy = FALSE,
  per_capita = TRUE,
  age = FALSE,
  sex = FALSE,
  bmi_groups = FALSE,
  dir = ".",
  spamfiledirectory = ""
)
}
\arguments{
\item{gdx}{GDX file}

\item{file}{a file name the output should be written to using write.magpie}

\item{level}{Level of regional aggregation; "iso" ISO country codes, "reg" (regional), "glo" (global)}

\item{calibrated}{if FALSE, the true regression outputs are used, if TRUE the values calibrated to the start years are used}

\item{pregnancy}{if TRUE, adding the intake requriements for lactation and pregnancy}

\item{per_capita}{per capita or aggregated for the population}

\item{age}{if FALSE age and sex is aggregated}

\item{sex}{if TRUE, data is provided by sex}

\item{bmi_groups}{if TRUE data is proided by BMI group}

\item{dir}{for gridded outputs: magpie output directory which contains a mapping file (rds or spam) disaggregation}

\item{spamfiledirectory}{deprecated. please use \code{dir} instead}
}
\value{
calories as MAgPIE object (unit depends on per_capita: cal/cap/day (TRUE), cal/day (FALSE))
}
\description{
Calculates the per-capita kcal intake from the food demand model
}
\details{
Demand definitions are equivalent to FAO Food supply categories
}
\examples{

  \dontrun{
    x <- Intake(gdx)
  }

}
\author{
Benjamin Leon Bodirsky
}
