% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IntakeDetailed.R
\name{IntakeDetailed}
\alias{IntakeDetailed}
\title{IntakeDetailed}
\usage{
IntakeDetailed(
  gdx,
  file = NULL,
  level = "reg",
  product_aggr = FALSE,
  dir = ".",
  spamfiledirectory = ""
)
}
\arguments{
\item{gdx}{GDX file}

\item{file}{a file name the output should be written to using write.magpie}

\item{level}{Level of regional aggregation; "reg" (regional), "glo" (global), "regglo" (regional and global) or any other aggregation level defined in superAggregate}

\item{product_aggr}{aggregate over products or not (boolean)}

\item{dir}{for gridded outputs: magpie output directory which contains a mapping file (rds or spam) disaggregation}

\item{spamfiledirectory}{deprecated. please use \code{dir} instead}
}
\value{
Calories as MAgPIE object (unit: kcal/cap/day)
}
\description{
Calculates detailed or aggregated per-capita kcal intake including exogenous scenarios
}
\details{
Calculation of kcal food intake is possible for both exogenous diet scenarios and endogenous estimation from food demand model
}
\examples{

  \dontrun{
    x <- IntakeDetailed(gdx)
  }

}
\author{
Isabelle Weindl
}
