% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IntakeDetailedProtein.R
\name{IntakeDetailedProtein}
\alias{IntakeDetailedProtein}
\title{IntakeDetailedProtein}
\usage{
IntakeDetailedProtein(gdx, file = NULL, level = "reg", product_aggr = FALSE)
}
\arguments{
\item{gdx}{GDX file}

\item{file}{a file name the output should be written to using write.magpie}

\item{level}{Level of regional aggregation; "reg" (regional), "glo" (global), "regglo" (regional and global) or any other aggregation level defined in superAggregate}

\item{product_aggr}{aggregate over products or not (boolean)}
}
\value{
Protein intake as MAgPIE object (unit: grams/cap/day)
}
\description{
Calculates food-specific per-capita protein intake from magpie results in grams.
}
\examples{

  \dontrun{
    x <- IntakeDetailedProtein(gdx)
  }

}
\author{
Vartika Singh, Isabelle Weindl
}
