% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/land.R
\name{land}
\alias{land}
\title{land}
\usage{
land(
  gdx,
  file = NULL,
  level = "reg",
  types = NULL,
  subcategories = NULL,
  sum = FALSE,
  dir = ".",
  spamfiledirectory = ""
)
}
\arguments{
\item{gdx}{GDX file}

\item{file}{a file name the output should be written to using write.magpie}

\item{level}{Level of regional aggregation; "reg" (regional), "glo" (global), "regglo" (regional and global)
or any other aggregation level defined in gdxAggregate}

\item{types}{NULL or a vector of strings. If NULL, all land types are used. Options are "crop", "past",
"forestry", "primforest","secdforest, "urban", "other", "primother" and "secdother"}

\item{subcategories}{NULL or vector of strings. If NULL, no subcategories are returned. Meaningful options
are "crop, "forestry" and "other"}

\item{sum}{determines whether output should be land-type-specific (FALSE) or aggregated over all types (TRUE).}

\item{dir}{for gridded outputs: magpie output directory which contains a mapping file (rds) for disaggregation}

\item{spamfiledirectory}{deprecated. please use \code{dir} instead}
}
\value{
land as MAgPIE object (Mha)
}
\description{
reads land out of a MAgPIE gdx file
}
\examples{
\dontrun{
x <- land(gdx)
}

}
\seealso{
\code{\link{reportLandUse}}
}
\author{
Jan Philipp Dietrich, Florian Humpenoeder, Benjamin Leon Bodirsky, Patrick v. Jeetze
}
