% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/land_price.R
\name{land_price}
\alias{land_price}
\title{land_price}
\usage{
land_price(
  gdx,
  file = NULL,
  level = "reg",
  ignore_lowbound = FALSE,
  absolute = TRUE,
  digits = 4
)
}
\arguments{
\item{gdx}{GDX file}

\item{file}{a file name the output should be written to using write.magpie}

\item{level}{Level of regional aggregation; "reg" (regional), "glo" (global), "regglo" (regional and global) or any other aggregation level defined in superAggregate}

\item{ignore_lowbound}{Some shadow prices are positive (see Details), corresponding to a lower bound for that pool. \code{TRUE} sets them to 0. Default value: \code{FALSE}.}

\item{absolute}{Should the absolute value of all the marginals be taken into calculations? \code{TRUE} (default) of \code{FALSE}. See Details.}

\item{digits}{rounding accuracy for the output}
}
\value{
A MAgPIE object containing the land shadow prices (US$05/ha).
}
\description{
Calculates MAgPIE MAgPIE land shadow prices based on a gdx file
}
\details{
The land price is obtained through marginals of the "oq_cropland" constraint. The majority of these marginals are negative values, and a negligible number of them are positive. This is the consequence of the constraint binding either on upper or lower level. The parameter \code{ignore_lowbound} removes all the positive marginals from land price calculation (negligible), and parameter \code{absolute} transforms them into negative values (to be all together reported as positive values at the final calculation).
}
\examples{

  \dontrun{
    x <- land_price(level="regglo", products="kcr")
  }

}
\author{
Markus Bonsch, Misko Stevanovic
}
