% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landConservation.R
\name{landConservation}
\alias{landConservation}
\title{landConservation}
\usage{
landConservation(
  gdx,
  file = NULL,
  level = "cell",
  cumuRestor = FALSE,
  baseyear = 1995,
  annualRestor = FALSE,
  sum = FALSE,
  dir = "."
)
}
\arguments{
\item{gdx}{GDX file}

\item{file}{a file name the output should be written to using write.magpie}

\item{level}{Level of regional aggregation; "cell", "grid", "iso, "reg" (regional), "glo" (global), "regglo" (regional and global) or any secdforest aggregation level defined in superAggregate}

\item{cumuRestor}{Logical; Whether function should report cumulative land restoration.}

\item{baseyear}{Base year used for cumulative land restoration reporting (default = 1995)}

\item{annualRestor}{Logical; Whether function should report annual land restoration.}

\item{sum}{sum over land pools (default = FALSE)}

\item{dir}{for gridded outputs: magpie output directory which contains a mapping file (rds) for disaggregation}
}
\value{
protected area in Mha
}
\description{
reads land conservation information out of a MAgPIE gdx file.
Land restoration \code{'restore'} is reported in Mha/yr by default but can be
also reported both over the time step length and cumulatively.
}
\details{
protected areas in primforest, secdforest and other land
}
\examples{
\dontrun{
x <- landConservation(gdx)
}

}
\author{
Florian Humpenoeder, Patrick v. Jeetze
}
