% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landForestry.R
\name{landForestry}
\alias{landForestry}
\title{landForestry}
\usage{
landForestry(gdx, file = NULL, level = "reg")
}
\arguments{
\item{gdx}{GDX file}

\item{file}{a file name the output should be written to using write.magpie}

\item{level}{Level of regional aggregation; "reg" (regional), "glo" (global), "regglo" (regional and global) or any other aggregation level defined in superAggregate}
}
\value{
land as MAgPIE object (Mha)
}
\description{
reads and compiles forestry land subcategories from a  MAgPIE gdx file
}
\examples{

  \dontrun{
    x <- land(gdx)
  }

}
\seealso{
\code{\link{reportLandUse}}
}
\author{
Florian Humpenoeder
}
