% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lastIter.R
\name{lastIter}
\alias{lastIter}
\title{lastIter}
\usage{
lastIter(gdx, param, secondlast = FALSE)
}
\arguments{
\item{gdx}{GDX file}

\item{param}{Parameter to be returned}

\item{secondlast}{if TRUE, reads the secondlast iteration. For MAgPIE results, usually there is no last iteration as the food demand model reaches convergence before MAgPIE starts.}
}
\value{
magpie object
}
\description{
Returns the value of a parameter in the last iteration
}
\examples{

  \dontrun{
    x <- lastIter(gdx)
  }

}
\author{
Benjamin Leon Bodirsky
}
