% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LivestockDemStructure.R
\name{LivestockDemStructure}
\alias{LivestockDemStructure}
\title{LivestockDemStructure}
\usage{
LivestockDemStructure(
  gdx,
  file = NULL,
  level = "reg",
  after_shock = TRUE,
  calibrated = TRUE,
  attributes = "kcal",
  fish = FALSE
)
}
\arguments{
\item{gdx}{GDX file}

\item{file}{a file name the output should be written to using write.magpie}

\item{level}{Level of regional aggregation; "iso" ISO country codes, "reg" (regional), "glo" (global)}

\item{after_shock}{FALSE is using the exogenous real income and the prices before a shock, TRUE is using the endogeenous real income that takes into account food price change on real income}

\item{calibrated}{if FALSE, the true regression outputs are used, if TRUE the values calibrated to the start years are used}

\item{attributes}{unit: kilocalories per day ("kcal"), g protein per day ("protein"). Mt reactive nitrogen ("nr").}

\item{fish}{if true, livestock share includes fish, otherwhise not}
}
\value{
magpie object with the livestock demand structure in a region or country. Unit is dimensionsless, but value depends on chosen attribute
}
\description{
Calculates the share of different livestock commodities in total livestock product consumption on the basis of chosen attribute
}
\examples{

  \dontrun{
    x <- LivestockDemStructure(gdx)
  }

}
\author{
Isabelle Weindl
}
