% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localDemandShare.R
\name{localDemandShare}
\alias{localDemandShare}
\title{localDemandShare}
\usage{
localDemandShare(gdx, level = "reg", product_aggr = FALSE, file = NULL)
}
\arguments{
\item{gdx}{GDX file}

\item{level}{spatial aggregation to report employment ("reg", "glo" or "regglo")}

\item{product_aggr}{sum over products if TRUE}

\item{file}{a file name the output should be written to using write.magpie}
}
\value{
share or level of food consumed locally, in disaggregated transport module
}
\description{
returns labor and capital cost share out of factor costs (i.e. labor + capital)
}
\examples{
\dontrun{
x <- localDemandShare(gdx)
}
}
\author{
David M Chen
}
