% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localDemandShares.R
\name{localDemandShares}
\alias{localDemandShares}
\title{localDemandShares}
\usage{
localDemandShares(
  gdx,
  type = "local",
  level = "reg",
  product_aggr = TRUE,
  urb_aggr = TRUE,
  fvc_aggr = TRUE,
  feed_aggr = TRUE,
  file = NULL
)
}
\arguments{
\item{gdx}{GDX file}

\item{type}{Type of ratio that should be calculated
\itemize{
       \item \code{local}: How much local demand (taking into account industrial/rural split) is satisfied by local consumption
       \item \code{localtotal}: How much total gridded demand is satisfied by local consumption
       \item \code{potential}: How much total gridded demand is potentially
                               satisfied by gridded production
       \item \code{prod}: Not really used, to delete
       }}

\item{level}{spatial aggregation to report employment ("reg", "glo" or "regglo")}

\item{product_aggr}{sum over products if TRUE}

\item{urb_aggr}{sum over products if TRUE}

\item{fvc_aggr}{sum over products if TRUE}

\item{feed_aggr}{sum over food and feed if TRUE, buggy, keep TRUE for now}

\item{file}{a file name the output should be written to using write.magpie}
}
\value{
share or level of food consumed locally, in disaggregated transport module
}
\description{
reports local demand and production shares based on local consumption
}
\examples{
\dontrun{
x <- localDemandShares(gdx)
}
}
\author{
David M Chen
}
