% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_yeardiff.R
\name{m_yeardiff}
\alias{m_yeardiff}
\title{m_yeardiff}
\usage{
m_yeardiff(gdx)
}
\arguments{
\item{gdx}{GDX file}
}
\value{
a magpie object with the length of each timestep
}
\description{
Calculates the parameter m_yeardiff, which is a macro within MAgPIE.
}
\examples{

  \dontrun{
    x <- m_yeardiff(gdx)
  }

}
\author{
Benjamin Leon Bodirsky
}
