% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magpie4-package.R
\docType{package}
\name{magpie4-package}
\alias{magpie4-package}
\alias{magpie4}
\title{MAgPIE outputs R library for MAgPIE version 4.x}
\description{
Common output routines for extracting results from the MAgPIE framework (versions 4.x).
}
\author{
Maintainer: Jan Philipp Dietrich <dietrich@pik-potsdam.de>
}
