% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magpieReportAsQuitte.R
\name{magpieReportAsQuitte}
\alias{magpieReportAsQuitte}
\title{magpieReportAsQuitte}
\usage{
magpieReportAsQuitte(report)
}
\arguments{
\item{report}{A MAgPIE object containing the report to be converted.}
}
\value{
The report as a quitte object with the global region named "World".
}
\description{
Converts a MAgPIE report object to a quitte object.
}
\author{
Patrick Rein
}
