% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/malmquist.R
\name{malmquist}
\alias{malmquist}
\title{malmquist}
\usage{
malmquist(gdx, file = NULL)
}
\arguments{
\item{gdx}{GDX file}

\item{file}{a file name the output should be written to using write.magpie}
}
\value{
A MAgPIE object containing the malmquist index
}
\description{
calcluates malmquist index based on a MAgPIE gdx file
}
\examples{

  \dontrun{
    x <- malmquist(gdx)
  }
}
\author{
Xiaoxi Wang
}
