% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ManureExcretion.R
\name{ManureExcretion}
\alias{ManureExcretion}
\title{ManureExcretion}
\usage{
ManureExcretion(
  gdx,
  level = "reg",
  products = "kli",
  awms = c("grazing", "stubble_grazing", "fuel", "confinement"),
  agg = TRUE,
  disagg_lvst = "landbased",
  recycled = FALSE
)
}
\arguments{
\item{gdx}{GDX file}

\item{level}{aggregation level: glo, reg, cell, grid, iso}

\item{products}{livestock products}

\item{awms}{animal waste management systems: "grazing", "stubble_grazing", "fuel", "confinement"}

\item{agg}{aggregation over "awms" or over "products".}

\item{disagg_lvst}{grid-level disaggregation method (level "grid"/"iso" only):
"landbased" (default) weights ruminant manure by pasture or crop production and monogastric
manure by urban or cropland area; "glw" weights by the gridded livestock distribution
(f71_livestock_distribution_0.5.mz next to gdx) times the pasture share (grazing, fuel)
or cropland share (stubble_grazing).}

\item{recycled}{"glw" only: if TRUE, confinement manure is also weighted by the cropland share.}
}
\value{
MAgPIE object
}
\description{
downscales Manure Excretion
}
\examples{

  \dontrun{
    x <- ManureExcretion(gdx)
  }

}
\author{
Benjamin Leon Bodirsky, Bin Lin
}
