% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MapLandUseForDownscaling.R
\name{MapLandUseForDownscaling}
\alias{MapLandUseForDownscaling}
\title{MapLandUseForDownscaling}
\usage{
MapLandUseForDownscaling(
  gdx,
  dir = ".",
  file = NULL,
  downscaling_model = "SEALS",
  selectyears = c("y2015", "y2050"),
  grid = TRUE
)
}
\arguments{
\item{gdx}{GDX file}

\item{dir}{output directory which contains cellular magpie output.}

\item{file}{a file name the output should be written to using write.magpie}

\item{downscaling_model}{The only current option is "SEALS" (Spatial Economic Allocation Landscape Simulator).}

\item{selectyears}{Years to provide data for.}

\item{grid}{Whether gridded output should be used. Otherwise output will be returned at cluster level.}
}
\value{
Shares of different land use classes per grid sell as magpie object.
}
\description{
Modifies magpie land use output so that it can be fed into a downscaling model for generating high resolution land use maps.
}
\examples{
\dontrun{
x <- MapLandUseForDownscaling(dir = ".", downscaling_model = "SEALS")
}

}
\author{
Patrick v. Jeetze
}
