% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_comments.R
\name{metadata_comments}
\alias{metadata_comments}
\title{metadata_comments}
\usage{
metadata_comments(x, unit, description, comment, note)
}
\arguments{
\item{x}{magpie object (magpie4)}

\item{unit}{provide unit}

\item{description}{provide short description}

\item{comment}{optional comment}

\item{note}{optional note}
}
\value{
vector of comments following order of input (unit, description, comment, note - further: origin, creation data)
}
\description{
set metadata comments to magpie4 objects
}
\examples{

  \dontrun{
    x <- metadata_comments(x,unit,description,comment,note)
  }

}
\author{
Benjamin Bodirsky, Jannes Breier
}
