% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetForestChange.R
\name{NetForestChange}
\alias{NetForestChange}
\title{NetForestChange}
\usage{
NetForestChange(gdx, file = NULL, level = "cell", lowpass = NULL)
}
\arguments{
\item{gdx}{GDX file}

\item{file}{a file name the output should be written to using write.magpie}

\item{level}{Level of regional aggregation; "reg" (regional), "glo" (global),
"regglo" (regional and global) or any other aggregation level defined in superAggregateX}

\item{lowpass}{number of lowpass filter iterations (default = NULL)}
}
\value{
Net Forest Change as MAgPIE object (Mha per year)
}
\description{
Calculates net and gross forest area change based on a MAgPIE gdx file
}
\examples{
\dontrun{
x <- NetForestChange(gdx)
}

}
\author{
Florian Humpenoeder
}
