% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NitrogenBudget.R
\name{NitrogenBudget}
\alias{NitrogenBudget}
\title{NitrogenBudget}
\usage{
NitrogenBudget(
  gdx,
  include_emissions = FALSE,
  level = "reg",
  debug = FALSE,
  cropTypes = FALSE,
  threshold = 0.05,
  progress = TRUE,
  disagg_lvst = "landbased"
)
}
\arguments{
\item{gdx}{GDX file}

\item{include_emissions}{TRUE also divides the N surplus into different emissions}

\item{level}{aggregation level, reg, glo or regglo, cell, iso or grid}

\item{debug}{debug mode TRUE makes some consistency checks between estimates for different resolutions.}

\item{cropTypes}{FALSE for aggregate results; TRUE for crop-specific results}

\item{threshold}{passed to mstools::toolFertilizerDistribution}

\item{progress}{passed to mstools::toolFertilizerDistribution}

\item{disagg_lvst}{grid-level disaggregation of the manure applied to cropland (confinement
manure recycled to cropland and stubble grazing), passed to \code{\link{ManureExcretion}}:
"landbased" (default) or "glw" (gridded livestock distribution times cropland share).}
}
\description{
calculates projections of Nitrogen Budgets for Croplands (Tg Nr per) from a MAgPIE gdx file
}
\examples{
\dontrun{
x <- NitrogenBudget(gdx)
}

}
\author{
Benjamin Leon Bodirsky, Michael Crawford, Edna J. Molina Bacca, Florian Humpenoeder, Bin Lin
}
