% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NitrogenBudgetNonagland.R
\name{NitrogenBudgetNonagland}
\alias{NitrogenBudgetNonagland}
\title{NitrogenBudgetNonagland}
\usage{
NitrogenBudgetNonagland(gdx, level = "reg", dir = ".")
}
\arguments{
\item{gdx}{GDX file}

\item{level}{aggregation level, reg, glo or regglo}

\item{dir}{for gridded outputs: magpie output directory which contains a mapping file (rds) for disaggregation}
}
\description{
calculates projections of Nitrogen Budgets for non-agricutlural land from a MAgPIE gdx file
}
\examples{

  \dontrun{
    x <- NitrogenBudgetNonagland(gdx)
  }

}
\author{
Benjamin Leon Bodirsky, Edna J. Molina Bacca
}
