% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NitrogenBudgetPasture.R
\name{NitrogenBudgetPasture}
\alias{NitrogenBudgetPasture}
\title{NitrogenBudgetPasture}
\usage{
NitrogenBudgetPasture(gdx, include_emissions = FALSE, level = "reg", dir = ".")
}
\arguments{
\item{gdx}{GDX file}

\item{include_emissions}{TRUE also divides the N surplus into different emissions}

\item{level}{aggregation level, reg, glo or regglo, cell, grid, iso}

\item{dir}{for gridded outputs: magpie output directory which contains a mapping file (rds or spam) disaggregation}
}
\description{
calculates projections of Nitrogen Budgets for Croplands from a MAgPIE gdx file
}
\examples{
\dontrun{
x <- NitrogenBudgetPasture(gdx)
}

}
\author{
Benjamin Leon Bodirsky, Edna J. Molina Bacca
}
