% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OtherLand.R
\name{OtherLand}
\alias{OtherLand}
\title{OtherLand}
\usage{
OtherLand(gdx, level = "reg")
}
\arguments{
\item{gdx}{GDX file}

\item{level}{Level of regional aggregation; "cell", "reg" (regional), "glo" (global), "regglo" (regional and global) or any aggregation level defined in superAggregate. In addition "climate" for the 3 climate regions tropical, temperate and boreal is available.}
}
\value{
Other land area in Mha
}
\description{
Disaggregation of other land into initial, restored and recovered land based on a MAgPIE gdx file
}
\details{
initial, restored and recovered land
}
\examples{

  \dontrun{
    x <- PeatlandArea(gdx)
  }
}
\author{
Florian Humpenoeder
}
