% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/out.R
\name{out}
\alias{out}
\title{out}
\usage{
out(x,file)
}
\arguments{
\item{x}{an object that can be converted to a MAgPIE object}

\item{file}{file name of a file it should be written to. NULL, if x should
be returned instead to be written to a file.}
}
\value{
NULL or x as MAgPIE object
}
\description{
Function to safely returns parameters. Function returns either the output or
writes it to a file. Please use this function when you write own GDX output
functions.
}
\author{
Jan Philipp Dietrich
}
