% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputPerWorker.R
\name{outputPerWorker}
\alias{outputPerWorker}
\title{outputPerWorker}
\usage{
outputPerWorker(gdx, level = "reg", file = NULL)
}
\arguments{
\item{gdx}{GDX file}

\item{level}{spatial aggregation to report employment ("reg", "glo",
"regglo", or custom region aggregation)}

\item{file}{a file name the output should be written to using write.magpie}
}
\value{
output per worker as magpie object
}
\description{
returns output per worker in crop+livestock production
}
\examples{
\dontrun{
x <- outputPerWorker(gdx)
}
}
\author{
Debbora Leip
}
