% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeatlandArea.R
\name{PeatlandArea}
\alias{PeatlandArea}
\title{PeatlandArea}
\usage{
PeatlandArea(gdx, file = NULL, level = "cell")
}
\arguments{
\item{gdx}{GDX file}

\item{file}{a file name the output should be written to using write.magpie}

\item{level}{Level of regional aggregation; "cell", "reg" (regional), "glo" (global), "regglo" (regional and global) or any aggregation level defined in superAggregate. In addition "climate" for the 3 climate regions tropical, temperate and boreal is available.}
}
\value{
Peatland area in Mha
}
\description{
reads peatland area out of a MAgPIE gdx file
}
\details{
Intact, degraded and rewettet peatland area
}
\examples{

  \dontrun{
    x <- PeatlandArea(gdx)
  }
}
\author{
Florian Humpenoeder
}
