% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeatlandEmissions.R
\name{PeatlandEmissions}
\alias{PeatlandEmissions}
\title{PeatlandEmissions}
\usage{
PeatlandEmissions(
  gdx,
  file = NULL,
  level = "cell",
  unit = "gas",
  cumulative = FALSE,
  baseyear = 1995,
  lowpass = 0,
  sum = TRUE
)
}
\arguments{
\item{gdx}{GDX file}

\item{file}{a file name the output should be written to using write.magpie}

\item{level}{Level of regional aggregation; "cell", "reg" (regional), "glo" (global), "regglo" (regional and global) or any aggregation level defined in superAggregate. In addition "climate" for the 3 climate regions tropical, temperate and boreal is available.}

\item{unit}{global warming potential (gwp) or gas (gas)}

\item{cumulative}{FALSE (default) or TRUE}

\item{baseyear}{Baseyear used for cumulative emissions (default = 1995)}

\item{lowpass}{number of lowpass filter iterations (default = 0)}

\item{sum}{sum over 3 dimension TRUE (default) or FALSE}
}
\value{
Peatland GHG emissions in Mt CO2eq (if unit="gwp") or Mt of the respective gas (if unit="gas")
}
\description{
reads peatland GHG emissions out of a MAgPIE gdx file
}
\details{
Peatland GHG emissions: CO2, DOC, CH4 and N2O
}
\examples{

  \dontrun{
    x <- PeatlandArea(gdx)
  }
}
\author{
Florian Humpenoeder
}
