% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlantationEstablishment.R
\name{PlantationEstablishment}
\alias{PlantationEstablishment}
\title{PlantationEstablishment}
\usage{
PlantationEstablishment(gdx, file = NULL, level = "cell")
}
\arguments{
\item{gdx}{GDX file}

\item{file}{a file name the output should be written to using write.magpie}

\item{level}{Level of regional aggregation; "cell", "reg" (regional), "glo" (global), "regglo" (regional and global) or any secdforest aggregation level defined in superAggregate}
}
\value{
Area newly for timber production
}
\description{
reads carbon stocks in harvested timber out of a MAgPIE gdx file
}
\details{
Area newly established in current time step for future timber production
}
\examples{

  \dontrun{
    x <- PlantationEstablishment(gdx)
  }
}
\author{
Abhijeet Mishra
}
