% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population.R
\name{population}
\alias{population}
\title{population}
\usage{
population(
  gdx,
  file = NULL,
  level = "reg",
  age = FALSE,
  sex = FALSE,
  bmi_groups = FALSE
)
}
\arguments{
\item{gdx}{GDX file}

\item{file}{a file name the output should be written to using write.magpie}

\item{level}{Level of regional aggregation; "reg" (regional),
"glo" (global), "regglo" (regional and global)
or any other aggregation level defined in superAggregate}

\item{age}{if TRUE, population is split up by age groups}

\item{sex}{if TRUE, population is split up by sex}

\item{bmi_groups}{if TRUE, the population will be split up in body-mass-index groups.}
}
\value{
population as MAgPIE object (million people)
}
\description{
reads population out of a MAgPIE gdx file
}
\examples{
\dontrun{
x <- population(gdx)
}

}
\seealso{
\code{\link{reportPopulation}}
}
\author{
Florian Humpenoeder, Benjamin Bodirsky, Isabelle Weindl
}
