% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PriceGHG.R
\name{PriceGHG}
\alias{PriceGHG}
\title{PriceGHG}
\usage{
PriceGHG(gdx, file = NULL, level = "reg", aggr = "max")
}
\arguments{
\item{gdx}{GDX file}

\item{file}{a file name the output should be written to using write.magpie}

\item{level}{Level of regional aggregation; "reg" (regional), "glo" (global), "regglo" (regional and global) or any other aggregation level defined in superAggregate}

\item{aggr}{aggregation used for global value; "max" (maxium value accross regions) or "weight" (weighted by population)}
}
\value{
GHG emission prices as MAgPIE object (US$2005/tCO2, US$2005/tN2O, US$2005/tCH4)
}
\description{
reads GHG emission prices out of a MAgPIE gdx file
}
\examples{

  \dontrun{
    x <- PriceGHG(gdx)
  }

}
\seealso{
\code{\link{reportPriceGHG}}
}
\author{
Florian Humpenoeder, Amsalu W. Yalew
}
