% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priceIndexFood.R
\name{priceIndexFood}
\alias{priceIndexFood}
\title{priceIndexFood}
\usage{
priceIndexFood(
  gdx,
  file = NULL,
  level = "reg",
  index = "lasp",
  chain = FALSE,
  baseyear = "y2005",
  round = TRUE,
  product_aggr = TRUE
)
}
\arguments{
\item{gdx}{GDX file}

\item{file}{a file name the output should be written to using write.magpie}

\item{level}{Level of regional aggregation; "reg" (regional), "glo" (global), "regglo" (regional and global) or any other aggregation level defined in superAggregate}

\item{index}{"lasp" (Laspeyres-Index: baseyear weighting), "paas" (Paasche-Index: current weighting), "fish" (Fisher-Index: geometric mean of "lasp" and "paas")}

\item{chain}{Chain Index: if true, the base period for each time period is the immediately preceding time period. Can be combined with all of the above indices}

\item{baseyear}{baseyear of the price index. type model to take baseyear 2010 with literature prices}

\item{round}{shall the results be rounded?}

\item{product_aggr}{aggregate over products or not (boolean)}
}
\value{
A MAgPIE object containing price indices for consumers or producers (depending on type)
}
\description{
calcluates price indicies based on a MAgPIE gdx file
}
\examples{

  \dontrun{
    x <- priceIndexFood(gdx)
  }

}
\author{
Jan Philipp Dietrich, Florian Humpenoeder, Benjamin Bodirsky
}
