% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prices.R
\name{prices}
\alias{prices}
\title{prices}
\usage{
prices(
  gdx,
  file = NULL,
  level = "reg",
  products = "kall",
  product_aggr = FALSE,
  attributes = "dm",
  type = "consumer",
  glo_weight = "production"
)
}
\arguments{
\item{gdx}{GDX file}

\item{file}{a file name the output should be written to using write.magpie}

\item{level}{Level of regional aggregation; "reg" (regional), "glo" (global),
"regglo" (regional and global) or any other aggregation level defined in superAggregate}

\item{products}{Selection of products (either by naming products, e.g. "tece", or naming a set,e.g."kcr")}

\item{product_aggr}{aggregate over products or not (boolean)}

\item{attributes}{USD05MER per ton X (dm,nr,p,k,wm) except gross energy (ge) where it is USD05MER per GJ}

\item{type}{"consumer" or "producer" prices. Producers' prices are calculated on the regional
level as a sum of regional trade equation marginal values and respective global trade equation
marginal values.For the non traded commodities, both global and regional producers prices are
set to zero instead of NaN.}

\item{glo_weight}{Decides the calculation of global prices. Weighting schemes are applied for
estimation of global producer price. If \code{"export"} prices are calculated as average of regional
exporters' prices, weighted by the export volumes. If \code{"production"} (default), prices are
calculated as average of regional prices weighted by regional production. Alternatively, if
\code{"free_trade"}, the global prices are directly taken from the shadow prices of the global
trade constraint, and no averaging is performed.}
}
\value{
A MAgPIE object containing the consumer's or producers' prices (unit depends on attributes)
}
\description{
calcluates prices based on a MAgPIE gdx file
}
\examples{
\dontrun{
x <- prices(gdx)
}

}
\author{
Misko Stevanovic, Florian Humpenoeder, Jan Philipp Dietrich, Xiaoxi Wang
}
