% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/productionProfit.R
\name{productionProfit}
\alias{productionProfit}
\title{productionProfit}
\usage{
productionProfit(gdx, file = NULL, level = "reg", dir = ".")
}
\arguments{
\item{gdx}{GDX file}

\item{file}{a file name the output should be written to using write.magpie}

\item{level}{Level of regional aggregation; "reg" (regional), "glo" (global), "regglo" (regional and global) or any other aggregation level defined in superAggregate}

\item{dir}{for gridded outputs: magpie output directory which contains a mapping file (rds or spam) disaggregation}
}
\value{
A MAgPIE object containing producers profit in million $US.
}
\description{
calcluates aggregate producer profit based on a MAgPIE gdx file.
}
\examples{

  \dontrun{
    x <- productionProfit(gdx)
  }

}
\author{
Miodrag Stevanovic
}
