% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/productionRevenue.R
\name{productionRevenue}
\alias{productionRevenue}
\title{productionRevenue}
\usage{
productionRevenue(
  gdx,
  file = NULL,
  level = "reg",
  products = "kall",
  product_aggr = TRUE
)
}
\arguments{
\item{gdx}{GDX file}

\item{file}{a file name the output should be written to using write.magpie}

\item{level}{Level of regional aggregation; "reg" (regional), "glo" (global), "regglo" (regional and global) or any other aggregation level defined in superAggregate}

\item{products}{Selection of products (either by naming products, e.g. "tece", or naming a set,e.g."kcr")}

\item{product_aggr}{ggregate over products or not (boolean, default TRUE)}
}
\value{
A MAgPIE object containing prodcution revenues.
}
\description{
calcluates production revenue based on a MAgPIE gdx file.
}
\examples{

  \dontrun{
    x <- productionRevenue(gdx)
  }
}
\author{
Miodrag Stevanovic
}
