% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readGDXBilateral.R
\name{readGDXBilateral}
\alias{readGDXBilateral}
\title{readGDXBilateral}
\usage{
readGDXBilateral(gdx, symbol)
}
\arguments{
\item{gdx}{gdx file to report from}

\item{symbol}{name of the symbol to read from gdx}
}
\description{
Internal support function to read bilateral spatial data from GDX files.
}
\author{
David M Chen
}
