% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relativeHourlyLaborCosts.R
\name{relativeHourlyLaborCosts}
\alias{relativeHourlyLaborCosts}
\title{relativeHourlyLaborCosts}
\usage{
relativeHourlyLaborCosts(gdx, level = "reg", file = NULL)
}
\arguments{
\item{gdx}{GDX file}

\item{level}{spatial aggregation to report ("iso", "reg", "glo", or "regglo")}

\item{file}{a file name the output should be written to using write.magpie}
}
\value{
labor costs per ag. worker in relation to GDP pc
}
\description{
calculates labor costs per ag. worker in relation to GDP pc
}
\examples{
\dontrun{
x <- relativeHourlyLaborCosts(gdx)
}
}
\author{
Debbora Leip
}
