% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportAAI.R
\name{reportAAI}
\alias{reportAAI}
\title{reportAAI}
\usage{
reportAAI(gdx, level = "regglo")
}
\arguments{
\item{gdx}{GDX file}

\item{level}{aggregation level of returned data ("regglo" by default)}
}
\value{
Area actually irrigated as MAgPIE object. Unit: see names
}
\description{
reports area actually irrigated
}
\section{Area actually irrigated variables}{
\tabular{lll}{
   Name \tab Unit \tab Meta \cr
   Resources|Land Cover|Cropland|Area actually irrigated \tab million ha \tab Cropland area actually receiving irrigation \cr
}
}

\examples{
\dontrun{
x <- reportAEI(gdx)
}

}
\author{
Stephen Wirth, Anne Biewald
}
