% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportAEI.R
\name{reportAEI}
\alias{reportAEI}
\title{reportAEI}
\usage{
reportAEI(gdx)
}
\arguments{
\item{gdx}{GDX file}
}
\value{
Area equipped for Irrigation as MAgPIE object. Unit: see names
}
\description{
reports Area equipped for Irrigation
}
\examples{

  \dontrun{
    x <- reportAEI(gdx)
  }

}
\author{
Stephen Wirth
}
