% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportAgGDP.R
\name{reportAgGDP}
\alias{reportAgGDP}
\title{reportAgGDP}
\usage{
reportAgGDP(gdx)
}
\arguments{
\item{gdx}{GDX file}
}
\value{
Magpie object
}
\description{
reports MAgPIE Agricultural GDP Mio. USD05 MER
}
\examples{
\dontrun{
x <- reportAgGDP(gdx)
}
}
\author{
Edna J. Molina Bacca
}
