% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportAgLandscapeIntactness.R
\name{reportAgLandscapeIntactness}
\alias{reportAgLandscapeIntactness}
\title{reportAgLandscapeIntactness}
\usage{
reportAgLandscapeIntactness(gdx)
}
\arguments{
\item{gdx}{GDX file}
}
\value{
Agricultural landscape intactness index as MAgPIE object
}
\description{
reports agricultural landscape intactness index
}
\examples{
\dontrun{
x <- reportAgLandscapeIntactness(gdx)
}

}
\author{
Patrick v. Jeetze
}
