% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportBII.R
\name{reportBII}
\alias{reportBII}
\title{reportBII}
\usage{
reportBII(gdx, dir = ".")
}
\arguments{
\item{gdx}{GDX file}

\item{dir}{magpie output directory that contains gridded BII data}
}
\value{
Biodiversity intactness index as MAgPIE object
}
\description{
reports biodiversity intactness index
}
\examples{
\dontrun{
x <- reportBII(gdx)
}

}
\author{
Patrick v. Jeetze, Florian Humpenoeder
}
