% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportBiochar.R
\name{reportBiochar}
\alias{reportBiochar}
\title{reportBiochar}
\usage{
reportBiochar(gdx)
}
\arguments{
\item{gdx}{GDX file}
}
\value{
Biochar production and carbon storage as MAgPIE object
}
\description{
reports production and carbon storage of biochar
}
\examples{

  \dontrun{
    x <- reportBiochar(gdx)
  }
}
\author{
Kristine Karstens, Isabelle Weindl
}
