% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportBiodiversityHotspotIntactness.R
\name{reportBiodiversityHotspotIntactness}
\alias{reportBiodiversityHotspotIntactness}
\title{reportBiodiversityHotspotIntactness}
\usage{
reportBiodiversityHotspotIntactness(gdx, dir = ".")
}
\arguments{
\item{gdx}{GDX file}

\item{dir}{magpie output directory that contains gridded BII data}
}
\value{
Biodiversity hotspot intactness index as MAgPIE object
}
\description{
reports biodiversity hotspot intactness intactness index
}
\examples{
\dontrun{
x <- reportBiodiversityHotspotIntactness(gdx)
}

}
\author{
Patrick v. Jeetze
}
