% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportBiogasFeedstock.R
\name{reportBiogasFeedstock}
\alias{reportBiogasFeedstock}
\title{reportBiogasFeedstock}
\usage{
reportBiogasFeedstock(gdx, file = NULL)
}
\arguments{
\item{gdx}{GDX file}

\item{file}{a file name the output should be written to using write.magpie}
}
\value{
MAgPIE object with biogas feedstock potential ("manure", "fodder") at ISO level in PJ.
}
\description{
Reports biogas feedstock from a MAgPIE GDX file at ISO country level.
Currently covers two sources: manure allocated to digesters and fodder crops
(placeholder, all zeros until MAgPIE tracks foddr/forage crop allocation to biogas).
Manure is converted from nitrogen mass (Mt N) to energy (PJ) using livestock-specific
heating values and nitrogen contents from Hoyos-Sebá et al. (2024).
}
\examples{
\dontrun{
  x <- reportBiogasFeedstock(gdx)
}
}
\author{
Kristine Karstens
}
