% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportBioplasticDemand.R
\name{reportBioplasticDemand}
\alias{reportBioplasticDemand}
\title{reportBioplasticDemand}
\usage{
reportBioplasticDemand(gdx, detail = TRUE, level = "regglo")
}
\arguments{
\item{gdx}{GDX file}

\item{detail}{only relevant for substrate demand. If TRUE, substrate demand is disaggregated by crop type, if 
FALSE only the aggregated demand is reported.}

\item{level}{spatial aggregation to report bioplastic/substrate demand (only "reg" or "regglo")}
}
\value{
bioplastic and bioplastic substrate demand as MAgPIE object
}
\description{
reports  demand for bioplastic and demand for substrate for bioplastic production from MAgPIE results
}
\examples{

  \dontrun{
    x <- reportBioplasticDemand(gdx)
  }

}
\author{
Debbora Leip
}
