% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportCostInputsCrop.R
\name{reportCostInputsCrop}
\alias{reportCostInputsCrop}
\title{reportCostInputsCrop}
\usage{
reportCostInputsCrop(gdx)
}
\arguments{
\item{gdx}{GDX file}
}
\value{
Magpie object associated with overall costs and value of production
}
\description{
reports MAgPIE costs
}
\examples{
\dontrun{
x <- reportCostInputsCrop(gdx)
}
}
\author{
Edna J. Molina Bacca
}
