% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportCosts.R
\name{reportCosts}
\alias{reportCosts}
\title{reportCosts}
\usage{
reportCosts(gdx)
}
\arguments{
\item{gdx}{GDX file}
}
\value{
consumption value as MAgPIE object Unit: see names
}
\description{
reports MAgPIE costs
}
\examples{
  \dontrun{
    x <- reportCosts(gdx)
  }
}
\author{
Florian Humpenoeder
}
