% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportCostsAccounting.R
\name{reportCostsAccounting}
\alias{reportCostsAccounting}
\title{reportCostsAccounting}
\usage{
reportCostsAccounting(gdx)
}
\arguments{
\item{gdx}{GDX file}
}
\value{
Costs accounting including total investments
}
\description{
reports MAgPIE costs including total investments
}
\examples{
\dontrun{
x <- reportCostsAccounting(gdx)
}

}
\author{
Edna J. Molina Bacca
}
