% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportCostsAdaptationCrops.R
\name{reportCostsAdaptationCrops}
\alias{reportCostsAdaptationCrops}
\title{reportCostsAdaptationCrops}
\usage{
reportCostsAdaptationCrops(gdx, type = "investment")
}
\arguments{
\item{gdx}{GDX file}

\item{type}{Type of reporting, either "annuity" or total "investments"}
}
\value{
costs related to adaptation,  (million US$05/yr/tDM)
}
\description{
reports costs relevant to evaluate adaptation in crops
}
\examples{
\dontrun{
x <- reportCostsAdaptationCrops(gdx)
}

}
\author{
Edna J. Molina Bacca
}
