% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportCostsFertilizer.R
\name{reportCostsFertilizer}
\alias{reportCostsFertilizer}
\title{reportCostsFertilizer}
\usage{
reportCostsFertilizer(gdx)
}
\arguments{
\item{gdx}{GDX file}
}
\value{
magpie object with fertilizer costs
}
\description{
reports MAgPIE nitrogen fertilizer costs disaggregated to crop categories
}
\examples{

  \dontrun{
    x <- reportCostsFertilizer(gdx)
  }

}
\author{
Debbora Leip
}
