% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportCostsMACCS.R
\name{reportCostsMACCS}
\alias{reportCostsMACCS}
\title{reportCostsMACCS}
\usage{
reportCostsMACCS(gdx)
}
\arguments{
\item{gdx}{GDX file}
}
\value{
magpie object with mitigation costs
}
\description{
reports MAgPIE mitigation costs disaggregated into labor and capital
}
\examples{

  \dontrun{
    x <- reportCostsMACCS(gdx)
  }

}
\author{
Debbora Leip
}
