% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportCostsOptimization.R
\name{reportCostsOptimization}
\alias{reportCostsOptimization}
\title{reportCostsOptimization}
\usage{
reportCostsOptimization(gdx)
}
\arguments{
\item{gdx}{GDX file}
}
\value{
reports the costs entering the optimization
}
\description{
reports MAgPIE costs
}
\examples{
\dontrun{
x <- reportCostsOptimization(gdx)
}

}
\author{
Edna J. Molina Bacca
}
