% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportCostsPresolve.R
\name{reportCostsPresolve}
\alias{reportCostsPresolve}
\title{reportCostsPresolve}
\usage{
reportCostsPresolve(gdx, level = "regglo")
}
\arguments{
\item{gdx}{GDX file}

\item{level}{An aggregation level for the spatial dimension. Can be any level
available via superAggregateX.}
}
\value{
consumption value as MAgPIE object Unit: see names
}
\description{
reports MAgPIE costs
}
\section{Presolve cost variables}{
\tabular{lll}{
   Name \tab Unit \tab Meta \cr
   Costs|PreSolve|Total \tab million US$2017 \tab Cumulative costs from presolve phase \cr
}
}

\examples{
  \dontrun{
    x <- reportCostsPresolve(gdx)
  }

}
\author{
Florian Humpenoeder
}
