% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportCostsProductionCrops.R
\name{reportCostsProductionCrops}
\alias{reportCostsProductionCrops}
\title{reportCostsProductionCrops}
\usage{
reportCostsProductionCrops(gdx, type = "investment")
}
\arguments{
\item{gdx}{GDX file}

\item{type}{Type of reporting, either "annuity" or total "investments"}
}
\value{
costs related to crops production,  (million US$05/yr/tDM)
}
\description{
reports costs about production crops
}
\examples{
\dontrun{
x <- reportCostsProductionCrops(gdx)
}

}
\author{
Edna J. Molina Bacca
}
