% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportCostsTrade.R
\name{reportCostsTrade}
\alias{reportCostsTrade}
\title{reportCostsTrade}
\usage{
reportCostsTrade(gdx, level = "regglo", sum = FALSE)
}
\arguments{
\item{gdx}{GDX file}

\item{level}{Level of regional aggregation; "reg" (regional), "glo" (global),
"regglo" (regional and global) or any other aggregation level defined in superAggregate}

\item{sum}{whether to sum across subcategories}
}
\value{
A MAgPIE object containing values related to trade costs (million US$2017/yr)
}
\description{
Reports MAgPIE bilateral trade cost components
}
\author{
David M Chen
}
