% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportCostsWithoutIncentives.R
\name{reportCostsWithoutIncentives}
\alias{reportCostsWithoutIncentives}
\title{reportCostsWithoutIncentives}
\usage{
reportCostsWithoutIncentives(gdx, level = "regglo")
}
\arguments{
\item{gdx}{GDX file}

\item{level}{spatial aggregation: "reg", "glo", "regglo"}
}
\value{
magpie object
}
\description{
reports Costs Without Incentives
}
\examples{
\dontrun{
x <- reportCostsWithoutIncentives(gdx)
}


}
\author{
David Chen
}
