% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportCostTransport.R
\name{reportCostTransport}
\alias{reportCostTransport}
\title{reportCostTransport}
\usage{
reportCostTransport(gdx)
}
\arguments{
\item{gdx}{GDX file}
}
\value{
consumption value as MAgPIE object Unit: see names
}
\description{
reports MAgPIE costs
}
\examples{

  \dontrun{
    x <- reportCostTransport(gdx)
  }
}
\author{
David Chen
}
