% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportCroparea.R
\name{reportCroparea}
\alias{reportCroparea}
\title{reportCroparea}
\usage{
reportCroparea(gdx, detail = FALSE)
}
\arguments{
\item{gdx}{GDX file}

\item{detail}{if detail=FALSE, the subcategories of groups are not reported (e.g. "soybean" within "oilcrops")}
}
\value{
Croparea as MAgPIE object (million Ha/yr)
}
\description{
reports croparea
}
\examples{

  \dontrun{
    x <- reportCroparea(gdx)
  }

}
\author{
Florian Humpenoeder
}
