% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportCropareaGrid.R
\name{reportCropareaGrid}
\alias{reportCropareaGrid}
\title{reportCropareaGrid}
\usage{
reportCropareaGrid(gdx)
}
\arguments{
\item{gdx}{GDX file}
}
\value{
Croparea as MAgPIE object (million ha/yr)
}
\description{
reports croparea
}
\examples{
\dontrun{
x <- reportCropareaGrid(gdx)
}

}
\author{
Benjamin Bodirsky
}
