% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportCropDiversity.R
\name{reportCropDiversity}
\alias{reportCropDiversity}
\title{reportCropDiversity}
\usage{
reportCropDiversity(gdx, grid = FALSE)
}
\arguments{
\item{gdx}{GDX file}

\item{grid}{Set to TRUE, if outputs should be reported on 0.5 degree grid level}
}
\value{
Crop diversity as MAgPIE object
}
\description{
reports crop diversity
}
\section{Crop diversity variables}{
\tabular{lll}{
   Name \tab Unit \tab Meta \cr
   Biodiversity|Shannon crop area diversity index \tab unitless \tab Crop type diversity based on area shares (higher = more diverse) \cr
   Biodiversity|Inverted Simpson crop area diversity index \tab unitless \tab Crop type diversity based on area shares (higher = more diverse) \cr
}
}

\examples{
\dontrun{
x <- reportCropDiversity(gdx)
}

}
\author{
Patrick v. Jeetze
}
