% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportCroplandSCM.R
\name{reportCroplandSCM}
\alias{reportCroplandSCM}
\title{reportCroplandSCM}
\usage{
reportCroplandSCM(gdx, level = "regglo")
}
\arguments{
\item{gdx}{GDX file}

\item{level}{aggregation level ("regglo" by default)}
}
\value{
SCM area indicators as MAgPIE object (million ha)
}
\description{
Reports cropland area under soil carbon management (SCM)
}
\section{SCM area variables}{
\tabular{lll}{
   Name \tab Unit \tab Meta \cr
   Resources|Land Cover|Cropland|Croparea under Soil Carbon Management \tab million ha \tab Total cropland area under soil carbon management \cr
   Resources|Land Cover|Cropland|Croparea under Soil Carbon Management|Share of Croparea \tab 1 \tab Share of cropland area under SCM \cr
}
}

\examples{
\dontrun{
x <- reportCroplandSCM(gdx)
}

}
\author{
Kristine Karstens
}
