% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportCropResidues2ndBE.R
\name{reportCropResidues2ndBE}
\alias{reportCropResidues2ndBE}
\title{reportCropResidues2ndBE}
\usage{
reportCropResidues2ndBE(
  gdx,
  file = NULL,
  collectionFraction = 0.25,
  minDensityForExtraction = 0.1
)
}
\arguments{
\item{gdx}{GDX file}

\item{file}{a file name the output should be written to using write.magpie}

\item{collectionFraction}{fraction of available residues that can be physically collected (default: 0.25)}

\item{minDensityForExtraction}{minimum residue density (tDM/ha) below which extraction is not economic (default: 0.1)}
}
\value{
MAgPIE object with crop residue potential for 2nd gen bioenergy (res_cereals, res_fibrous, res_nonfibrous) at ISO level in PJ.
}
\description{
Reports crop residues available for 2nd generation bioenergy from a
MAgPIE GDX file at ISO country level. Applies soil cover constraints (minimum 30\%
soil cover retained, Lutz et al. 2019) and a collection fraction to estimate
sustainably harvestable residue biomass, converted to energy (PJ).
}
\examples{
\dontrun{
  x <- reportCropResidues2ndBE(gdx)
}
}
\references{
Lutz et al. (2019), GMD, \url{https://gmd.copernicus.org/articles/12/2419/2019/}
}
\seealso{
\code{\link{ResidueBiomass}}, \code{\link{ResidueUsage}}
}
\author{
Kristine Karstens
}
