% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportDemandNr.R
\name{reportDemandNr}
\alias{reportDemandNr}
\title{reportDemandNr}
\usage{
reportDemandNr(gdx, detail = FALSE)
}
\arguments{
\item{gdx}{GDX file}

\item{detail}{if detail=F, the subcategories of groups are not reported (e.g. "soybean" within "oilcrops")}
}
\value{
demand as MAgPIE object (Mt DM)
}
\description{
Similar to reportDemand, but for nitrogen. reports Demand for Food, Feed, Processing, Material, Bioenergy, Seed and Supply Chain Loss
}
\examples{

  \dontrun{
    x <- reportDemand()
  }

}
\author{
Benjamin Leon Bodirsky
}
