% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportEmissionsBeforeTechnicalMitigation.R
\name{reportEmissionsBeforeTechnicalMitigation}
\alias{reportEmissionsBeforeTechnicalMitigation}
\title{reportEmissionsBeforeTechnicalMitigation}
\usage{
reportEmissionsBeforeTechnicalMitigation(gdx)
}
\arguments{
\item{gdx}{GDX file}
}
\value{
MAgPIE object (Unit: Mt CO2/yr, Mt N2O/yr and Mt CH4/yr)
}
\description{
reports GHG emissions before technical mitigation. Technical abatement includes all abatement done in the MACC curves, but exclude endogenous mitigation. These emissions are NOT the standard reporting emissions, but used for special purposes like remind-magpie coupling.
}
\examples{

  \dontrun{
    x <- reportEmissionsBeforeTechnicalMitigation(gdx)
  }
}
\author{
Florian Humpenoeder, Benjamin Leon Bodirsky
}
