% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportExpenditureFoodIndex.R
\name{reportExpenditureFoodIndex}
\alias{reportExpenditureFoodIndex}
\title{reportExpenditureFoodIndex}
\usage{
reportExpenditureFoodIndex(gdx, baseyear = "y2010", basketyear = "y2010")
}
\arguments{
\item{gdx}{GDX file}

\item{baseyear}{Baseyear of the price index}

\item{basketyear}{Year of reference food basket
(should be in the past for comparison of different runs to
have identical and comparable food basket)}
}
\value{
Food expenditure index as MAgPIE object
}
\description{
reports food expenditure index and food expenditure index
             corrected for emission costs
}
\examples{
\dontrun{
x <- reportPriceFoodIndex(gdx)
}

}
\author{
Felicitas Beier
}
