% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportFactorCostShares.R
\name{reportFactorCostShares}
\alias{reportFactorCostShares}
\title{reportFactorCostShares}
\usage{
reportFactorCostShares(gdx, type = "optimization", level = "regglo")
}
\arguments{
\item{gdx}{GDX file}

\item{type}{\itemize{
  \item "requirements": shares from factor requirements
  \item "optimization": cost shares between labor and capital costs in optimization
  \item "accounting": cost shares based on accounting of labor and capital costs
}}

\item{level}{spatial aggregation: "reg", "glo", "regglo"}
}
\value{
labor and capital cost shares as MAgPIE object
}
\description{
reports labor and capital cost share out of factor costs from MAgPIE results
}
\examples{
\dontrun{
x <- reportFactorCostShares(gdx)
}

}
\author{
Debbora Leip
}
