% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportFeed.R
\name{reportFeed}
\alias{reportFeed}
\title{reportFeed}
\usage{
reportFeed(gdx, detail = T)
}
\arguments{
\item{gdx}{GDX file}

\item{detail}{if detail=F, the subcategories of groups are not reported (e.g. "soybean" within "oilcrops")}
}
\value{
feed demand as MAgPIE object (Mt DM)
}
\description{
reportes feed demand by animal type
}
\examples{

  \dontrun{
    x <- reportFeed()
  }


}
\author{
Isabelle Weindl
}
