% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportFertilizerNitrogen.R
\name{reportFertilizerNitrogen}
\alias{reportFertilizerNitrogen}
\title{reportFertilizerNitrogen}
\usage{
reportFertilizerNitrogen(gdx, level = "regglo")
}
\arguments{
\item{gdx}{GDX file}

\item{level}{level of output}
}
\description{
Reports inorganic nitrogen application on crops
}
\examples{
\dontrun{
x <- reportFertilizerNitrogen(gdx)
}

}
\seealso{
\code{\link{NitrogenBudget}}
}
\author{
David M Chen
}
