% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportFireEmissions.R
\name{reportFireEmissions}
\alias{reportFireEmissions}
\title{reportFireEmissions.R}
\usage{
reportFireEmissions(gdx, level = "reg")
}
\arguments{
\item{gdx}{GDX file}

\item{level}{Level of regional aggregation; "reg" (regional), "glo" (global), "regglo" (regional and global)}
}
\value{
emissions as MAgPIE object (unit: Mt X/yr, plus cumulative Mt X/yr)
}
\description{
reads land cover data and deforestation data to produce future estimates of fire emissions
based on extrapolating GFED data from 2003-2016
}
\examples{
  \dontrun{
    x <- reportFireEmissions(gdx, level = "glo")
  }
}
\author{
Michael Crawford
}
