% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportForestYield.R
\name{reportForestYield}
\alias{reportForestYield}
\title{reportForestYield}
\usage{
reportForestYield(gdx)
}
\arguments{
\item{gdx}{GDX file}
}
\value{
Yield from Forests for timber production
}
\description{
reports MAgPIE harvested area for timber.
}
\examples{

  \dontrun{
    x <- reportForestYield(gdx)
  }

}
\author{
Abhijeet Mishra
}
