% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportGrasslandSoilCarbon.R
\name{reportGraslandSoilCarbon}
\alias{reportGraslandSoilCarbon}
\title{reportGraslandSoilCarbon}
\usage{
reportGraslandSoilCarbon(gdx)
}
\arguments{
\item{gdx}{GDX file}
}
\value{
Cattle values as magpie objetc
}
\description{
reports cattle related numbers
}
\examples{
\dontrun{
x <- reportGraslandSoilCarbon(gdx)
}
}
\author{
Marcos Alves
}
