% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportGrasslandYields.R
\name{reportGrasslandYields}
\alias{reportGrasslandYields}
\title{reportGrasslandYields}
\usage{
reportGrasslandYields(gdx, level = "regglo")
}
\arguments{
\item{gdx}{GDX file}

\item{level}{aggregation level of returned data ("regglo" by default)}
}
\value{
yield as MAgPIE object (Mt DM/ha)
}
\description{
reportGrasslandYields
}
\section{Grassland yield variables}{
\tabular{lll}{
   Name \tab Unit \tab Meta \cr
   Productivity|Yield|+|Pasture \tab t DM/ha \tab Pasture grassland yield \cr
   Productivity|Yield|+|Range \tab t DM/ha \tab Range grassland yield \cr
}
}

\examples{

  \dontrun{
    x <- reportGrasslandYields(gdx)
  }

}
\author{
Marcos Alves
}
