% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportGrassStats.R
\name{reportGrassStats}
\alias{reportGrassStats}
\title{reportGrassStats}
\usage{
reportGrassStats(gdx)
}
\arguments{
\item{gdx}{GDX file}
}
\value{
magpie object
}
\description{
report evaluation values for pasture management implementation
}
\examples{
\dontrun{
x <- reportGrassStats(gdx)
}


}
\author{
Marcos Alves
}
