% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportGridCroparea.R
\name{reportGridCroparea}
\alias{reportGridCroparea}
\title{reportGridCroparea}
\usage{
reportGridCroparea(gdx, dir = ".", spamfiledirectory = "")
}
\arguments{
\item{gdx}{GDX file}

\item{dir}{for gridded outputs: magpie output directory which contains a mapping file (rds or spam) disaggregation}

\item{spamfiledirectory}{deprecated. please use \code{dir} instead}
}
\value{
area of cropland as MAgPIE object (million ha)
}
\description{
reports Croparea from gridded (disaggregated) output
}
\examples{

  \dontrun{
    x <- reportGridCroparea(gdx)
  }

}
\author{
Jannes Breier
}
