% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportGridLand.R
\name{reportGridLand}
\alias{reportGridLand}
\title{reportGridLand}
\usage{
reportGridLand(gdx)
}
\arguments{
\item{gdx}{GDX file}
}
\value{
land-use as MAgPIE object (million ha)
}
\description{
reports land-use from gridded (disaggregated) output
}
\examples{

  \dontrun{
    x <- reportGridLand(gdx)
  }

}
\author{
Jannes Breier
}
