% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportGrowingStock.R
\name{reportGrowingStock}
\alias{reportGrowingStock}
\title{reportGrowingStock}
\usage{
reportGrowingStock(gdx, indicator = "relative", detail = FALSE)
}
\arguments{
\item{gdx}{GDX file}

\item{indicator}{If the reported numbers are relative (mio m3/ha) or absolute (mio. m3). Default is relative.}

\item{detail}{if detail=FALSE, the subcategories of groups are not reported.}
}
\value{
production as MAgPIE object. Unit: see names
}
\description{
reports Growing stocks for woody materials
}
\examples{

  \dontrun{
    x <- reportGrowingStock(gdx)
  }

}
\author{
Abhijeet Mishra
}
