% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportharvested_area_timber.R
\name{reportharvested_area_timber}
\alias{reportharvested_area_timber}
\title{reportharvested_area_timber}
\usage{
reportharvested_area_timber(gdx)
}
\arguments{
\item{gdx}{GDX file}
}
\value{
Area harvested for timber production
}
\description{
reports MAgPIE harvested area for timber.
}
\examples{

  \dontrun{
    x <- reportharvested_area_timber(gdx)
  }

}
\author{
Abhijeet Mishra
}
