% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportHourlyLaborCosts.R
\name{reportHourlyLaborCosts}
\alias{reportHourlyLaborCosts}
\title{reportHourlyLaborCosts}
\usage{
reportHourlyLaborCosts(gdx, level = "regglo")
}
\arguments{
\item{gdx}{GDX file}

\item{level}{spatial aggregation: "reg", "glo", "regglo"}
}
\value{
hourly labor costs as MAgPIE object
}
\description{
reports hourly labor costs in agriculture from MAgPIE results
}
\section{Hourly labor cost variables}{
\tabular{lll}{
   Name \tab Unit \tab Meta \cr
   Labor|Wages|Hourly labor costs \tab US$2017/h \tab Hourly labor costs in agriculture \cr
}
}

\examples{

  \dontrun{
    x <- reportHourlyLaborCosts(gdx)
  }


}
\author{
Debbora Leip
}
