% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportIncome.R
\name{reportIncome}
\alias{reportIncome}
\title{reportIncome}
\usage{
reportIncome(gdx, type = "ppp", level = "regglo")
}
\arguments{
\item{gdx}{GDX file}

\item{type}{ppp for purchase power parity, mer for market exchange rate}

\item{level}{spatial aggregation: "reg", "glo", "regglo", "iso"}
}
\value{
Annual per capita and total income as MAgPIE object (US$2017 MER/cap/yr and million US$17 PPP/yr)
}
\description{
reports income
}
\section{Income variables}{
\tabular{lll}{
   Name \tab Unit \tab Meta \cr
   Income per capita PPP \tab US$2017 PPP/cap/yr \tab GDP per capita converted to USD using purchasing power parity (PPP) \cr
   Income PPP \tab million US$2017 PPP/yr \tab GDP converted to USD using purchasing power parity (PPP) \cr
   Income per capita MER \tab US$2017 MER/cap/yr \tab GDP per capita converted to USD at market exchange rate (MER) \cr
   Income MER \tab million US$2017 MER/yr \tab GDP converted to USD at market exchange rate (MER) \cr
}
}

\examples{
\dontrun{
x <- reportIncome(gdx)
}

}
\author{
Florian Humpenoeder, Isabelle Weindl, Felicitas Beier
}
