% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportIntakeDetailed.R
\name{reportIntakeDetailed}
\alias{reportIntakeDetailed}
\title{reportIntakeDetailed}
\usage{
reportIntakeDetailed(gdx, detail = TRUE, level = "regglo")
}
\arguments{
\item{gdx}{GDX file}

\item{detail}{if detail=F, the subcategories of groups are not reported (e.g. "soybean" within "oilcrops")}

\item{level}{spatial aggregation: "reg", "glo", "regglo", "iso"}
}
\value{
per-capita calorie intake as MAgPIE object (cal/cap/day)
}
\description{
reports detailed or aggregated per-capita cal intake including exogenous scenarios
}
\examples{

  \dontrun{
    x <- reportIntakeDetailed(gdx)
  }

}
\author{
Isabelle Weindl
}
