% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportKBAIntactness.R
\name{reportKBAIntactness}
\alias{reportKBAIntactness}
\title{reportKBAIntactness}
\usage{
reportKBAIntactness(gdx)
}
\arguments{
\item{gdx}{GDX file}
}
\value{
Key biodiversity area intactness index as MAgPIE object
}
\description{
reports key biodiversity area intactness index
}
\examples{
\dontrun{
x <- reportKBAIntactness(gdx)
}

}
\author{
Patrick v. Jeetze
}
