% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportKcal.R
\name{reportKcal}
\alias{reportKcal}
\title{reportKcal}
\usage{
reportKcal(gdx, detail = FALSE)
}
\arguments{
\item{gdx}{GDX file}

\item{detail}{if detail=F, the subcategories of groups are not reported (e.g. "soybean" within "oilcrops")}
}
\value{
per-capita calories as MAgPIE object (kcal/cap/day)
}
\description{
reports per-capita calories food supply (including household waste)
}
\examples{

  \dontrun{
    x <- reportKcal(gdx)
  }

}
\author{
Benjamin Leon Bodirsky, Kristine karstens, Abhijeet Mishra
}
