% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportLaborCostsEmpl.R
\name{reportLaborCostsEmpl}
\alias{reportLaborCostsEmpl}
\title{reportLaborCostsEmpl}
\usage{
reportLaborCostsEmpl(gdx, level = "regglo")
}
\arguments{
\item{gdx}{GDX file}

\item{level}{aggregation level of returned data ("regglo" by default)}
}
\value{
magpie object with labor costs
}
\description{
reports MAgPIE labor costs that go into employment calculation
}
\section{Labor costs for employment variables}{
\tabular{lll}{
   Name \tab Unit \tab Meta \cr
   Labor|Employment|Labor costs linked to employment \tab million US$2017/yr \tab Total labor costs used for employment calculation \cr
}
}

\examples{

  \dontrun{
    x <- reportLaborCostsEmpl(gdx)
  }

}
\author{
Debbora Leip
}
