% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportLaborProductivity.R
\name{reportLaborProductivity}
\alias{reportLaborProductivity}
\title{reportLaborProductivity}
\usage{
reportLaborProductivity(gdx, productAggr = TRUE, level = "regglo")
}
\arguments{
\item{gdx}{GDX file}

\item{productAggr}{Aggregate over products or not (boolean)}

\item{level}{spatial aggregation: "reg", "glo", "regglo", "iso"}
}
\value{
labor productivity as MAgPIE object
}
\description{
reports labor productivity in crop production
}
\examples{

  \dontrun{
    x <- reportLaborProductivity(gdx)
  }

}
\author{
Debbora Leip
}
