% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportLandFootprint.R
\name{reportLandFootprint}
\alias{reportLandFootprint}
\title{reportLandFootprint}
\usage{
reportLandFootprint(gdx)
}
\arguments{
\item{gdx}{GDX file}

\item{detail}{if detail=F, the subcategories of groups are not reported (e.g. "soybean" within "oilcrops")}

\item{level}{spatial aggregation: "reg", "glo", "regglo", "iso"}
}
\value{
land use for food production per capita (million hectares per-capita)
}
\description{
reports land footprint of food production for aggregated per-capita kcal intake including exogenous scenarios
}
\examples{

  \dontrun{
    x <- reportLandFootprint(gdx)
  }

}
\author{
Vartika Singh
}
