% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportLandTransitionMatrix.R
\name{reportLandTransitionMatrix}
\alias{reportLandTransitionMatrix}
\title{reportLandTransitionMatrix}
\usage{
reportLandTransitionMatrix(gdx, level = "regglo")
}
\arguments{
\item{gdx}{GDX file}

\item{level}{The aggregation level to be used ("regglo" by default)}
}
\value{
land transition matrix as MAgPIE object (Mha/yr)
}
\description{
reports land transition matrix with gross land-use change flows
between all land types
}
\section{Total land transition variables}{
\tabular{lll}{
   Name \tab Unit \tab Meta \cr
   Resources|Land Transitions \tab Mha/yr \tab Total gross land-use change (sum of all off-diagonal transitions) \cr
}
}

\section{From Cropland variables}{
\tabular{lll}{
   Name \tab Unit \tab Meta \cr
   Resources|Land Transitions|+|From Cropland \tab Mha/yr \tab Total annual gross land-use change from cropland to other land types \cr
   Resources|Land Transitions|From Cropland|+|To Pastures and Rangelands \tab Mha/yr \tab Annual gross transition from cropland to pastures and rangelands \cr
   Resources|Land Transitions|From Cropland|+|To Urban Area \tab Mha/yr \tab Annual gross transition from cropland to urban area \cr
   Resources|Land Transitions|From Cropland|+|To Other Land \tab Mha/yr \tab Annual gross transition from cropland to other land \cr
   Resources|Land Transitions|From Cropland|+|To Forest \tab Mha/yr \tab Annual gross transition from cropland to forest (sum of planted, primary, and secondary forest) \cr
   Resources|Land Transitions|From Cropland|To Forest|+|To Planted Forest \tab Mha/yr \tab Annual gross transition from cropland to planted forest \cr
   Resources|Land Transitions|From Cropland|To Forest|+|To Primary Forest \tab Mha/yr \tab Annual gross transition from cropland to primary forest (structurally zero) \cr
   Resources|Land Transitions|From Cropland|To Forest|+|To Secondary Forest \tab Mha/yr \tab Annual gross transition from cropland to secondary forest \cr
}
}

\section{From Pastures and Rangelands variables}{
\tabular{lll}{
   Name \tab Unit \tab Meta \cr
   Resources|Land Transitions|+|From Pastures and Rangelands \tab Mha/yr \tab Total annual gross land-use change from pastures and rangelands to other land types \cr
   Resources|Land Transitions|From Pastures and Rangelands|+|To Cropland \tab Mha/yr \tab Annual gross transition from pastures and rangelands to cropland \cr
   Resources|Land Transitions|From Pastures and Rangelands|+|To Urban Area \tab Mha/yr \tab Annual gross transition from pastures and rangelands to urban area \cr
   Resources|Land Transitions|From Pastures and Rangelands|+|To Other Land \tab Mha/yr \tab Annual gross transition from pastures and rangelands to other land \cr
   Resources|Land Transitions|From Pastures and Rangelands|+|To Forest \tab Mha/yr \tab Annual gross transition from pastures and rangelands to forest (sum of planted, primary, and secondary forest) \cr
   Resources|Land Transitions|From Pastures and Rangelands|To Forest|+|To Planted Forest \tab Mha/yr \tab Annual gross transition from pastures and rangelands to planted forest \cr
   Resources|Land Transitions|From Pastures and Rangelands|To Forest|+|To Primary Forest \tab Mha/yr \tab Annual gross transition from pastures and rangelands to primary forest (structurally zero) \cr
   Resources|Land Transitions|From Pastures and Rangelands|To Forest|+|To Secondary Forest \tab Mha/yr \tab Annual gross transition from pastures and rangelands to secondary forest \cr
}
}

\section{From Planted Forest variables}{
\tabular{lll}{
   Name \tab Unit \tab Meta \cr
   Resources|Land Transitions|+|From Planted Forest \tab Mha/yr \tab Total annual gross land-use change from planted forest to other land types \cr
   Resources|Land Transitions|From Planted Forest|+|To Cropland \tab Mha/yr \tab Annual gross transition from planted forest to cropland \cr
   Resources|Land Transitions|From Planted Forest|+|To Pastures and Rangelands \tab Mha/yr \tab Annual gross transition from planted forest to pastures and rangelands \cr
   Resources|Land Transitions|From Planted Forest|+|To Urban Area \tab Mha/yr \tab Annual gross transition from planted forest to urban area \cr
   Resources|Land Transitions|From Planted Forest|+|To Other Land \tab Mha/yr \tab Annual gross transition from planted forest to other land \cr
   Resources|Land Transitions|From Planted Forest|+|To Forest \tab Mha/yr \tab Annual gross transition from planted forest to other forest types (sum of primary and secondary forest) \cr
   Resources|Land Transitions|From Planted Forest|To Forest|+|To Primary Forest \tab Mha/yr \tab Annual gross transition from planted forest to primary forest (structurally zero) \cr
   Resources|Land Transitions|From Planted Forest|To Forest|+|To Secondary Forest \tab Mha/yr \tab Annual gross transition from planted forest to secondary forest \cr
}
}

\section{From Primary Forest variables}{
\tabular{lll}{
   Name \tab Unit \tab Meta \cr
   Resources|Land Transitions|+|From Primary Forest \tab Mha/yr \tab Total annual gross land-use change from primary forest to other land types \cr
   Resources|Land Transitions|From Primary Forest|+|To Cropland \tab Mha/yr \tab Annual gross transition from primary forest to cropland \cr
   Resources|Land Transitions|From Primary Forest|+|To Pastures and Rangelands \tab Mha/yr \tab Annual gross transition from primary forest to pastures and rangelands \cr
   Resources|Land Transitions|From Primary Forest|+|To Urban Area \tab Mha/yr \tab Annual gross transition from primary forest to urban area \cr
   Resources|Land Transitions|From Primary Forest|+|To Other Land \tab Mha/yr \tab Annual gross transition from primary forest to other land \cr
   Resources|Land Transitions|From Primary Forest|+|To Forest \tab Mha/yr \tab Annual gross transition from primary forest to other forest types (sum of planted and secondary forest) \cr
   Resources|Land Transitions|From Primary Forest|To Forest|+|To Planted Forest \tab Mha/yr \tab Annual gross transition from primary forest to planted forest \cr
   Resources|Land Transitions|From Primary Forest|To Forest|+|To Secondary Forest \tab Mha/yr \tab Annual gross transition from primary forest to secondary forest \cr
}
}

\section{From Secondary Forest variables}{
\tabular{lll}{
   Name \tab Unit \tab Meta \cr
   Resources|Land Transitions|+|From Secondary Forest \tab Mha/yr \tab Total annual gross land-use change from secondary forest to other land types \cr
   Resources|Land Transitions|From Secondary Forest|+|To Cropland \tab Mha/yr \tab Annual gross transition from secondary forest to cropland \cr
   Resources|Land Transitions|From Secondary Forest|+|To Pastures and Rangelands \tab Mha/yr \tab Annual gross transition from secondary forest to pastures and rangelands \cr
   Resources|Land Transitions|From Secondary Forest|+|To Urban Area \tab Mha/yr \tab Annual gross transition from secondary forest to urban area \cr
   Resources|Land Transitions|From Secondary Forest|+|To Other Land \tab Mha/yr \tab Annual gross transition from secondary forest to other land \cr
   Resources|Land Transitions|From Secondary Forest|+|To Forest \tab Mha/yr \tab Annual gross transition from secondary forest to other forest types (sum of planted and primary forest) \cr
   Resources|Land Transitions|From Secondary Forest|To Forest|+|To Planted Forest \tab Mha/yr \tab Annual gross transition from secondary forest to planted forest \cr
   Resources|Land Transitions|From Secondary Forest|To Forest|+|To Primary Forest \tab Mha/yr \tab Annual gross transition from secondary forest to primary forest (structurally zero) \cr
}
}

\section{From Urban Area variables}{
\tabular{lll}{
   Name \tab Unit \tab Meta \cr
   Resources|Land Transitions|+|From Urban Area \tab Mha/yr \tab Total annual gross land-use change from urban area to other land types \cr
   Resources|Land Transitions|From Urban Area|+|To Cropland \tab Mha/yr \tab Annual gross transition from urban area to cropland \cr
   Resources|Land Transitions|From Urban Area|+|To Pastures and Rangelands \tab Mha/yr \tab Annual gross transition from urban area to pastures and rangelands \cr
   Resources|Land Transitions|From Urban Area|+|To Other Land \tab Mha/yr \tab Annual gross transition from urban area to other land \cr
   Resources|Land Transitions|From Urban Area|+|To Forest \tab Mha/yr \tab Annual gross transition from urban area to forest (sum of planted, primary, and secondary forest) \cr
   Resources|Land Transitions|From Urban Area|To Forest|+|To Planted Forest \tab Mha/yr \tab Annual gross transition from urban area to planted forest \cr
   Resources|Land Transitions|From Urban Area|To Forest|+|To Primary Forest \tab Mha/yr \tab Annual gross transition from urban area to primary forest (structurally zero) \cr
   Resources|Land Transitions|From Urban Area|To Forest|+|To Secondary Forest \tab Mha/yr \tab Annual gross transition from urban area to secondary forest \cr
}
}

\section{From Other Land variables}{
\tabular{lll}{
   Name \tab Unit \tab Meta \cr
   Resources|Land Transitions|+|From Other Land \tab Mha/yr \tab Total annual gross land-use change from other land to other land types \cr
   Resources|Land Transitions|From Other Land|+|To Cropland \tab Mha/yr \tab Annual gross transition from other land to cropland \cr
   Resources|Land Transitions|From Other Land|+|To Pastures and Rangelands \tab Mha/yr \tab Annual gross transition from other land to pastures and rangelands \cr
   Resources|Land Transitions|From Other Land|+|To Urban Area \tab Mha/yr \tab Annual gross transition from other land to urban area \cr
   Resources|Land Transitions|From Other Land|+|To Forest \tab Mha/yr \tab Annual gross transition from other land to forest (sum of planted, primary, and secondary forest) \cr
   Resources|Land Transitions|From Other Land|To Forest|+|To Planted Forest \tab Mha/yr \tab Annual gross transition from other land to planted forest \cr
   Resources|Land Transitions|From Other Land|To Forest|+|To Primary Forest \tab Mha/yr \tab Annual gross transition from other land to primary forest (structurally zero) \cr
   Resources|Land Transitions|From Other Land|To Forest|+|To Secondary Forest \tab Mha/yr \tab Annual gross transition from other land to secondary forest \cr
}
}

\examples{

  \dontrun{
    x <- reportLandTransitionMatrix(gdx)
  }

}
\author{
Florian Humpenoeder
}
