% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportLandUse.R
\name{reportLandUse}
\alias{reportLandUse}
\title{reportLandUse}
\usage{
reportLandUse(gdx, level = "regglo")
}
\arguments{
\item{gdx}{GDX file}

\item{level}{The aggregation level to be used ("regglo" by default)}
}
\value{
land-use as MAgPIE object (million ha)
}
\description{
reports land-use
}
\section{Total land cover variables}{
\tabular{lll}{
   Name \tab Unit \tab Meta \cr
   Resources|Land Cover \tab million ha \tab Total land cover \cr
   Resources|Land Cover|Agricultural land \tab million ha \tab Agricultural land including cropland and pastures \cr
}
}

\section{Cropland variables}{
\tabular{lll}{
   Name \tab Unit \tab Meta \cr
   Resources|Land Cover|+|Cropland \tab million ha \tab Arable land, i.e. land in bioenergy crop, food, and feed/fodder crops, permanent crops as well as other arable land (physical area) \cr
   Resources|Land Cover|Cropland|+|Croparea \tab million ha \tab Physical cropland area used for crop production \cr
   Resources|Land Cover|Cropland|+|Fallow \tab million ha \tab Fallow cropland \cr
   Resources|Land Cover|Cropland|+|Tree Cover \tab million ha \tab Trees on cropland for agroforestry \cr
}
}

\section{Pasture and urban variables}{
\tabular{lll}{
   Name \tab Unit \tab Meta \cr
   Resources|Land Cover|+|Pastures and Rangelands \tab million ha \tab Pasture and range land based on FAO definition of permanent meadows and pastures \cr
   Resources|Land Cover|+|Urban Area \tab million ha \tab Built-up land associated with human settlements \cr
}
}

\section{Other natural land variables}{
\tabular{lll}{
   Name \tab Unit \tab Meta \cr
   Resources|Land Cover|+|Other Land \tab million ha \tab Non-forest natural land including primary non-forest, restored and recovered natural land \cr
   Resources|Land Cover|Other Land|Initial \tab million ha \tab Primary non-forest natural land without clearly visible indications of human activities \cr
   Resources|Land Cover|Other Land|Recovered \tab million ha \tab Recovered natural land due to the abandonment of agricultural or forestry land without intention for nature/biodiversity conservation \cr
   Resources|Land Cover|Other Land|Restored \tab million ha \tab Intentionally restored natural land for the purpose of nature and/or biodiversity conservation \cr
}
}

\section{Forest variables}{
\tabular{lll}{
   Name \tab Unit \tab Meta \cr
   Resources|Land Cover|+|Forest \tab million ha \tab Sum of primary, secondary and planted forest (FAO definition) \cr
   Resources|Land Cover|Forest|+|Natural Forest \tab million ha \tab Naturally regenerated forest including primary and secondary forest \cr
   Resources|Land Cover|Forest|Natural Forest|+|Primary Forest \tab million ha \tab Naturally regenerated forest of native tree species where there are no clearly visible indications of human activities (FAO definition) \cr
   Resources|Land Cover|Forest|Natural Forest|+|Secondary Forest \tab million ha \tab Forest predominantly composed of trees established through natural regeneration excluding primary forest (based on FAO definition) \cr
   Resources|Land Cover|Forest|Natural Forest|Secondary Forest|Young \tab million ha \tab Young secondary forest \cr
   Resources|Land Cover|Forest|Natural Forest|Secondary Forest|Mature \tab million ha \tab Mature secondary forest \cr
   Resources|Land Cover|Forest|+|Planted Forest \tab million ha \tab Forest predominantly composed of trees established through planting and/or deliberate seeding (FAO definition) \cr
   Resources|Land Cover|Forest|Planted Forest|+|Timber \tab million ha \tab Planted forest established to meet timber demand: intensively managed plantations, one or two species, even-aged, regular spacing (FAO Plantation Forest) \cr
   Resources|Land Cover|Forest|Planted Forest|+|CO2-price AR \tab million ha \tab Afforestation/reforestation established in response to the CO2 price for carbon sequestration \cr
   Resources|Land Cover|Forest|Planted Forest|CO2-price AR|+|Plantation \tab million ha \tab CO2-price AR grown as a monoculture and/or with non-native species (FAO Plantation Forest); non-zero only when s32_aff_plantation=1 \cr
   Resources|Land Cover|Forest|Planted Forest|CO2-price AR|+|Natural \tab million ha \tab CO2-price AR grown with native tree species resembling natural vegetation (FAO Other Planted Forest); non-zero only when s32_aff_plantation=0 \cr
   Resources|Land Cover|Forest|Planted Forest|+|NPI_NDC AR \tab million ha \tab Afforestation/reforestation established to meet national policies and NDC commitments (native species; FAO Other Planted Forest) \cr
   Resources|Land Cover|Forest|Planted Forest|+|Other Planted \tab million ha \tab Planted forest present at model initialisation; the non-plantation component of FAO "Other Planted Forest" \cr
}
}

\examples{

  \dontrun{
    x <- reportLandUse(gdx)
  }

}
\author{
Florian Humpenoeder, Kristine Karstens, Isabelle Weindl
}
