% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportLandUseChange.R
\name{reportLandUseChange}
\alias{reportLandUseChange}
\title{reportLandUseChange}
\usage{
reportLandUseChange(gdx, baseyear = 1995, level = "regglo", annual = FALSE)
}
\arguments{
\item{gdx}{GDX file}

\item{baseyear}{baseyear for calculating land-use change. Ignored if annual is TRUE.}

\item{level}{An aggregation level for the spatial dimension. Can be any level
available via superAggregateX.}

\item{annual}{If FALSE (default), report cumulative land-use change relative to baseyear
(million ha wrt baseyear) for all land types. If TRUE, report the average annual net change
rate per time step (million ha/yr) for primary forest and natural planted forest. The first
time step has no preceding time step and is set to 0.}
}
\value{
land-use change as MAgPIE object (million ha wrt to baseyear, or million ha/yr if
annual is TRUE)
}
\description{
reports land-use change
}
\section{Land-use change variables}{
\tabular{lll}{
   Name \tab Unit \tab Meta \cr
   Resources|Land Cover Change|Cropland \tab million ha wrt baseyear \tab Change in cropland area relative to baseyear \cr
   Resources|Land Cover Change|Pastures and Rangelands \tab million ha wrt baseyear \tab Change in pasture area relative to baseyear \cr
   Resources|Land Cover Change|Forest \tab million ha wrt baseyear \tab Change in forest area relative to baseyear \cr
   Resources|Land Cover Change|Other Land \tab million ha wrt baseyear \tab Change in other land area relative to baseyear \cr
}
}

\section{Annual land-use change variables (annual = TRUE)}{
\tabular{lll}{
   Name \tab Unit \tab Meta \cr
   Resources|Land Cover Change|Forest|Natural Forest|Primary Forest|Annual \tab million ha/yr \tab Average annual net change in primary forest area over the time step \cr
   Resources|Land Cover Change|Forest|Planted Forest|Natural|Annual \tab million ha/yr \tab Average annual net change in natural planted forest area over the time step \cr
}
}

\examples{

  \dontrun{
    x <- reportLandUseChange(gdx)
  }

}
\author{
Florian Humpenoeder, Miodrag Stevanović
}
