% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportLandUseForSEALS.R
\name{reportLandUseForSEALS}
\alias{reportLandUseForSEALS}
\title{reportLandUseForSEALS}
\usage{
reportLandUseForSEALS(
  magCellLand = "cell.land_0.5_share.mz",
  outFile = "cell.land_0.5_SEALS.nc",
  scenName = NULL,
  dir = ".",
  selectyears = c(2020, 2030, 2050)
)
}
\arguments{
\item{magCellLand}{Disaggregated land use (grid-cell land area share) as magclass object or file (.mz) from a MAgPIE run.}

\item{outFile}{a file name the output should be written to using \code{ncdf4::nc_create} and \code{ncdf4::ncvar_put}}

\item{scenName}{Optional scenario name}

\item{dir}{output directory which contains cellular magpie output}

\item{selectyears}{Numeric vector of years to provide data for.}
}
\value{
Proportions of different land use classes per grid sell in a NetCDF format.
}
\description{
Writes MAgPIE land use projections to a specific NetCDF
that can be read by the Spatial Economic Allocation Landscape Simulator (SEALS) model
for generating high resolution land use maps.
}
\examples{
\dontrun{
x <- reportLandUseForSEALS(
  magCellLand = "cell.land_0.5_share.mz",
  outFile = "cell.land_0.5_SEALS.nc",
  selectyears = c(2020, 2030, 2050)
)
}

}
\author{
Patrick v. Jeetze
}
