% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportLivestockShare.R
\name{reportLivestockShare}
\alias{reportLivestockShare}
\title{reportLivestockShare}
\usage{
reportLivestockShare(gdx, level = "regglo")
}
\arguments{
\item{gdx}{GDX file}

\item{level}{An aggregation level for the spatial dimension. Can be any level
available via superAggregateX.}
}
\value{
per-capita calories as MAgPIE object (kcal/cap/day)
}
\description{
reports the share of livestock products (including fish) in total calorie food supply
}
\section{Livestock share variables}{
\tabular{lll}{
   Name \tab Unit \tab Meta \cr
   Nutrition|Dietary Composition|Livestock Share \tab kcal/kcal \tab Share of livestock products (incl. fish) in total calorie supply \cr
}
}

\examples{

  \dontrun{
    x <- reportLivestockShare(gdx)
  }

}
\author{
Benjamin Leon Bodirsky
}
