% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportLocalDemandShares.R
\name{reportLocalDemandShares}
\alias{reportLocalDemandShares}
\title{reportLocalDemandShares}
\usage{
reportLocalDemandShares(gdx, type = "potential", level = "regglo")
}
\arguments{
\item{gdx}{GDX file}

\item{type}{Type of ratio that should be calculated
\itemize{
       \item \code{local}: How much local demand (taking into account industrial/rural split) is satisfied by local consumption
       \item \code{localtotal}: How much total gridded demand is satisfied by local consumption
       \item \code{potential}: How much total gridded demand is potentially
                               satisfied by gridded production
       \item \code{prod}: Not really used, to delete
       }}

\item{level}{spatial aggregation: "reg", "glo", "regglo"}
}
\value{
share of food demand at disaggregated level coming from local production as MAgPIE object
}
\description{
reports local demand and production shares based on local consumption
}
\examples{
\dontrun{
x <- reportLocalDemandShares(gdx)
}

}
\author{
David M Chen
}
