% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportGridAWMS.R, R/reportManure.R
\name{reportManure}
\alias{reportManure}
\title{reportManure}
\usage{
reportManure(gdx, nutrient = "nr")

reportManure(gdx, nutrient = "nr")
}
\arguments{
\item{gdx}{GDX file}

\item{nutrient}{nr, p, c...}
}
\description{
Reports the Nitrogen in Manure of all animals for future MAgPIE projections

Reports the Nitrogen in Manure of all animals for future MAgPIE projections
}
\examples{
  \dontrun{
    x <- reportManure(gdx)
  }

  \dontrun{
    x <- reportManure(gdx)
  }

}
\author{
Benjamin Leon Bodirsky

Benjamin Leon Bodirsky
}
