% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportNitrogenBudgetPasture.R
\name{reportNitrogenBudgetPasture}
\alias{reportNitrogenBudgetPasture}
\title{reportNitrogenBudgetCropland}
\usage{
reportNitrogenBudgetPasture(
  gdx,
  include_emissions = FALSE,
  grid = FALSE,
  dir = "."
)
}
\arguments{
\item{gdx}{GDX file}

\item{include_emissions}{TRUE also divides the N surplus into different emissions}

\item{grid}{if TRUE, disaggregate to grid level}

\item{dir}{for gridded outputs: magpie output directory which contains a mapping file (rds) for disaggregation}
}
\description{
Reports the Nitrogen Budgets of Croplands for future MAgPIE projections
}
\examples{
  \dontrun{
    x <- reportNitrogenBudgetCropland(gdx)
  }

}
\seealso{
\code{\link{NitrogenBudget}}
}
\author{
Benjamin Leon Bodirsky
}
