% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportNitrogenEfficiencies.R
\name{reportNitrogenEfficiencies}
\alias{reportNitrogenEfficiencies}
\title{reportNitrogenBudgetCropland}
\usage{
reportNitrogenEfficiencies(gdx)
}
\arguments{
\item{gdx}{GDX file}
}
\description{
Reports different nitrogen use efficiency indicators
}
\examples{
\dontrun{
x <- reportNitrogenEfficiencies(gdx)
}

}
\seealso{
\code{\link{reportNitrogenBudgetCropland}}
}
\author{
Benjamin Leon Bodirsky
}
