% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportOverallCosts.R
\name{reportOverallCosts}
\alias{reportOverallCosts}
\title{reportOverallCosts}
\usage{
reportOverallCosts(gdx, level = "reg")
}
\arguments{
\item{gdx}{GDX file}

\item{level}{level of aggregation of output}
}
\value{
A MAgPIE object containing the overall costs [million US$05] including total investment costs
}
\description{
Calculates Overall costs [million US$05] including total investment costs
}
\examples{

  \dontrun{
    x <- reportOverallCosts(gdx)
  }

}
\author{
Edna J. Molina Bacca
}
