% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportPeatland.R
\name{reportPeatland}
\alias{reportPeatland}
\title{reportPeatland}
\usage{
reportPeatland(gdx)
}
\arguments{
\item{gdx}{GDX file}
}
\value{
peatland area as magclass object (million ha)
}
\description{
reports peatland area
}
\examples{

  \dontrun{
    x <- reportPeatland(gdx)
  }

}
\author{
Florian Humpenoeder
}
