% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportPopulation.R
\name{reportPopulation}
\alias{reportPopulation}
\title{reportPopulation}
\usage{
reportPopulation(gdx, level = "regglo")
}
\arguments{
\item{gdx}{GDX file}

\item{level}{spatial aggregation: "reg", "glo", "regglo", "iso"}
}
\value{
population as MAgPIE object
}
\description{
reports Population
}
\section{Population variables}{
\tabular{lll}{
   Name \tab Unit \tab Meta \cr
   Population \tab million people \tab Total population \cr
}
}

\examples{

  \dontrun{
    x <- reportPopulation(gdx)
  }


}
\author{
Florian Humpenoeder
}
