% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportPriceAgriculture.R
\name{reportPriceAgriculture}
\alias{reportPriceAgriculture}
\title{reportPriceAgriculture}
\usage{
reportPriceAgriculture(gdx, level = "regglo")
}
\arguments{
\item{gdx}{GDX file}
}
\value{
agricultural commodity prices as MAgPIE object (USD)
}
\description{
reports food commodity prices
}
\section{Agricultural price variables}{
\tabular{lll}{
   Name \tab Unit \tab Meta \cr
   Prices|Agriculture|Crops \tab US$2017/tDM \tab Prices for crop products \cr
   Prices|Agriculture|Livestock products \tab US$2017/tDM \tab Prices for livestock products \cr
}
}

\examples{

  \dontrun{
    x <- reportPriceAgriculture(gdx)
  }

}
\author{
Mishko Stevanovic
}
