% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportPriceBioenergy.R
\name{reportPriceBioenergy}
\alias{reportPriceBioenergy}
\title{reportPriceBioenergy}
\usage{
reportPriceBioenergy(gdx)
}
\arguments{
\item{gdx}{GDX file}
}
\value{
bioenergy price as MAgPIE object Unit: see names
}
\description{
reports bioenergy prices
}
\examples{

  \dontrun{
    x <- reportPriceBioenergy(gdx)
  }

}
\author{
Florian Humpenoeder
}
