% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportPriceElasticities.R
\name{reportPriceElasticities}
\alias{reportPriceElasticities}
\title{reportHunger}
\usage{
reportPriceElasticities(gdx)
}
\arguments{
\item{gdx}{GDX file}
}
\value{
magpie object with hunger (mio people) or hunger share
}
\description{
Calculates the share of people living in hunger.
}
\examples{

  \dontrun{
    x <- reportHunger(gdx)
  }

}
\author{
Benjamin Leon Bodirsky
}
