% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportPriceGHG.R
\name{reportPriceGHG}
\alias{reportPriceGHG}
\title{reportPriceGHG}
\usage{
reportPriceGHG(gdx)
}
\arguments{
\item{gdx}{GDX file}
}
\value{
GHG emission prices as MAgPIE object
}
\description{
reports GHG emission prices
}
\examples{

  \dontrun{
    x <- reportPriceGHG(gdx)
  }

}
\author{
Florian Humpenoeder, Amsalu W. Yalew
}
