% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportPriceLand.R
\name{reportPriceLand}
\alias{reportPriceLand}
\title{reportPriceLand}
\usage{
reportPriceLand(gdx)
}
\arguments{
\item{gdx}{GDX file}
}
\value{
land prices as MAgPIE object Unit: see names
}
\description{
reports land prices (land rent)
}
\examples{

  \dontrun{
    x <- reportPriceLand(gdx)
  }

}
\author{
Florian Humpenoeder
}
