% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportPriceShock.R
\name{reportPriceShock}
\alias{reportPriceShock}
\title{reportPriceShock}
\usage{
reportPriceShock(gdx)
}
\arguments{
\item{gdx}{GDX file}
}
\value{
magpie object
}
\description{
Reports the change in consumption and expenditure due to higher or lower food prices
}
\examples{

  \dontrun{
    x <- reportPriceShock(gdx)
  }

}
\author{
Benjamin Leon Bodirsky
}
