% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportPriceWater.R
\name{reportPriceWater}
\alias{reportPriceWater}
\title{reportPriceWater}
\usage{
reportPriceWater(gdx, level = "regglo")
}
\arguments{
\item{gdx}{GDX file}

\item{level}{aggregation level of returned data ("regglo" by default)}
}
\value{
water usage as MAgPIE object Unit: see names
}
\description{
reports water prices
}
\section{Water price variables}{
\tabular{lll}{
   Name \tab Unit \tab Meta \cr
   Prices|Water|Agriculture \tab Index 2005=100 \tab Agricultural water price index \cr
}
}

\examples{
\dontrun{
x <- reportPriceWater(gdx)
}

}
\author{
Florian Humpenoeder
}
