% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportPriceWoodyBiomass.R
\name{reportPriceWoodyBiomass}
\alias{reportPriceWoodyBiomass}
\title{reportPriceWoodyBiomass}
\usage{
reportPriceWoodyBiomass(gdx)
}
\arguments{
\item{gdx}{GDX file}
}
\value{
land prices as MAgPIE object Unit: see names
}
\description{
reports woody biomass prices (land rent)
}
\examples{

  \dontrun{
    x <- reportPriceWoodyBiomass(gdx)
  }

}
\author{
Abhijeet Mishra
}
