% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportProcessing.R
\name{reportProcessing}
\alias{reportProcessing}
\title{reportProcessing}
\usage{
reportProcessing(gdx, detail = T, indicator = "primary_to_process")
}
\arguments{
\item{gdx}{GDX file}

\item{detail}{if detail=FALSE, the subcategories of groups are not reported (e.g. "soybean" within "oilcrops")}

\item{indicator}{"primary_to_process" for process or "secondary_from_primary" for secondary product output}
}
\value{
processing demand as MAgPIE object (Mt DM)
}
\description{
reportes processing input and output quantities primary-to-process or primary-to-secondary
}
\examples{

  \dontrun{
    x <- reportProcessing(gdx=gdx, detail=T, indicator = "primary_to_process")
  }


}
\author{
David Chen, Benjamin Leon Bodirsky
}
