% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportProducerPriceIndex.R
\name{reportProducerPriceIndex}
\alias{reportProducerPriceIndex}
\title{reportProducerPriceIndex}
\usage{
reportProducerPriceIndex(gdx, prod_groups = FALSE)
}
\arguments{
\item{gdx}{GDX file}

\item{prod_groups}{whether to return only product groups}
}
\value{
Producer price index as MAgPIE object Unit: see names
}
\description{
reports producer price index
}
\examples{

  \dontrun{
    x <- reportProducerPriceIndex(gdx)
  }

}
\author{
Isabelle Weindl, David M CHen
}
