% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportProductionBioenergy.R
\name{reportProductionBioenergy}
\alias{reportProductionBioenergy}
\title{reportProductionBioenergy}
\usage{
reportProductionBioenergy(gdx, detail = FALSE)
}
\arguments{
\item{gdx}{GDX file}

\item{detail}{if detail=FALSE, the subcategories of groups are not reported (e.g. "soybean" within "oilcrops")}
}
\value{
production as MAgPIE object. Unit: see names
}
\description{
reports 2nd gen bioenergy production
}
\section{Bioenergy production variables}{
\tabular{lll}{
   Name \tab Unit \tab Meta \cr
   Production|Bioenergy|2nd generation \tab EJ/yr \tab Second generation bioenergy production (grassy and woody crops) \cr
   Production|Bioenergy|2nd generation|++|Grassy bioenergy crops \tab EJ/yr \tab Production from short rotation grasses \cr
   Production|Bioenergy|2nd generation|++|Woody bioenergy crops \tab EJ/yr \tab Production from short rotation trees \cr
   Production|Bioenergy|2nd generation|Cumulative \tab EJ \tab Cumulative second generation bioenergy production \cr
}
}

\examples{

  \dontrun{
    x <- reportProductionBioenergy(gdx)
  }


}
\author{
Florian Humpenoeder
}
