% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportProductionGrowth.R
\name{reportProductionGrowth}
\alias{reportProductionGrowth}
\title{reportProductionGrowth}
\usage{
reportProductionGrowth(gdx, detail = FALSE)
}
\arguments{
\item{gdx}{GDX file}

\item{detail}{if true, provides results for all commodities, otherwhise aggregates some groups}
}
\value{
Production growth rates (index)
}
\description{
reports production growth rate
}
\examples{

  \dontrun{
    x <- reportProductionGrowth(gdx="fulldata.gdx",detail=TRUE)
  }

}
\author{
Xiaoxi Wang
}
