% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportProductionNr.R
\name{reportProductionNr}
\alias{reportProductionNr}
\title{reportProductionNr}
\usage{
reportProductionNr(gdx, detail = FALSE)
}
\arguments{
\item{gdx}{GDX file}

\item{detail}{if detail = FALSE, the subcategories of groups are not reported (e.g. "soybean" within "oilcrops")}
}
\value{
production as MAgPIE object. Unit: see names
}
\description{
reports production in Nr analogous to reportProduction
}
\section{Nitrogen production variables}{
\tabular{lll}{
   Name \tab Unit \tab Meta \cr
   Production Nr \tab Mt Nr/yr \tab Total nitrogen in agricultural production \cr
   Production Nr|+|Crop products \tab Mt Nr/yr \tab Nitrogen in crop production \cr
   Production Nr|+|Livestock products \tab Mt Nr/yr \tab Nitrogen in livestock production \cr
}
}

\examples{

  \dontrun{
    x <- reportProductionNr(gdx)
  }


}
\author{
Benjamin Leon Bodirsky
}
