% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportProtectedArea.R
\name{reportProtectedArea}
\alias{reportProtectedArea}
\title{reportProtectedArea}
\usage{
reportProtectedArea(gdx)
}
\arguments{
\item{gdx}{GDX file}
}
\value{
protected area in Mha
}
\description{
reports protected areas
}
\examples{

  \dontrun{
    x <- reportProtectedArea(gdx)
  }

}
\author{
Florian Humpenoeder
}
