% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportProtein.R
\name{reportProtein}
\alias{reportProtein}
\title{reportProtein}
\usage{
reportProtein(gdx, detail = FALSE)
}
\arguments{
\item{gdx}{GDX file}

\item{detail}{if detail=F, the subcategories of groups are not reported (e.g. "soybean" within "oilcrops")}
}
\value{
per-capita protein as MAgPIE object (protein/cap/day)
}
\description{
reports per-capita protein food supply (including household waste)
}
\examples{

  \dontrun{
    x <- reportKcal(gdx)
  }

}
\author{
Benjamin Leon Bodirsky, Kristine Karstens, Abhijeet Mishra, Florian Humpenoeder
}
