% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportRelativeHourlyLaborCosts.R
\name{reportRelativeHourlyLaborCosts}
\alias{reportRelativeHourlyLaborCosts}
\title{reportRelativeHourlyLaborCosts}
\usage{
reportRelativeHourlyLaborCosts(gdx)
}
\arguments{
\item{gdx}{GDX file}
}
\value{
labor costs per ag. worker in relation to GDP pc as MAgPIE object
}
\description{
reports labor costs per ag. worker in relation to GDP pc from MAgPIE results
}
\examples{

  \dontrun{
    x <- reportRelativeHourlyLaborCosts(gdx)
  }

}
\author{
Debbora Leip
}
