% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportResidues.R
\name{reportResidues}
\alias{reportResidues}
\title{reportSOM}
\usage{
reportResidues(gdx)
}
\arguments{
\item{gdx}{GDX file}
}
\description{
Report soil organic carbon stock size for future MAgPIE projections
}
\section{Residue variables}{
\tabular{lll}{
   Name \tab Unit \tab Meta \cr
   Resources|Carbon|Cropland|Residues \tab Mt C/yr \tab Total carbon in crop residues \cr
   Resources|Carbon|Cropland|Residues|+|Cereals \tab Mt C/yr \tab Carbon in cereal residues \cr
   Resources|Carbon|Cropland|Residues|+|Oilcrops \tab Mt C/yr \tab Carbon in oilcrop residues \cr
}
}

\examples{
  \dontrun{
    x <- reportSOM(gdx)
  }


}
\author{
Kristine Karstens
}
