% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tryReport.R
\name{reportResult}
\alias{reportResult}
\title{reportResult}
\usage{
reportResult(resultType, message, reportExpr, elapsed, result = NULL)
}
\arguments{
\item{resultType}{Character string for the result type}

\item{message}{Character string with the result message}

\item{reportExpr}{The expression containing the call to the report function}

\item{elapsed}{The elapsed time for the report in seconds}

\item{result}{The result magpie object or NULL}
}
\value{
A named list with report result information
}
\description{
Create a named list representing the outcome of a report
}
\keyword{internal}
