% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportRotationLength.R
\name{reportRotationLength}
\alias{reportRotationLength}
\title{reportRotationLength}
\usage{
reportRotationLength(gdx)
}
\arguments{
\item{gdx}{GDX file}
}
\value{
Forest rotation length
}
\description{
reports Forest rotation length.
}
\examples{

  \dontrun{
    x <- reportRotationLength(gdx)
  }

}
\author{
Abhijeet Mishra
}
