% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportRuralDemandShares.R
\name{reportRuralDemandShares}
\alias{reportRuralDemandShares}
\title{reportRuralDemandShares}
\usage{
reportRuralDemandShares(gdx, type = "tradOnly", level = "regglo")
}
\arguments{
\item{gdx}{GDX file}

\item{type}{Type of ratio that should be calculated
\itemize{
\item \code{all}: How much rural & trad demand as a share of all demand is satisfied locally
\item \code{tradOnly}: How much rural & trad demand as a share of rural & trad demand is satisfied locally
\item \code{potential}: How much total gridded demand is potentially
satisfied by gridded production
}}

\item{level}{spatial aggregation: "reg", "glo", "regglo"}
}
\value{
share of food demand at disaggregated level coming from local production as MAgPIE object
}
\description{
reports rural demand and production shares based on local consumption
}
\section{Rural demand share variables}{
\tabular{lll}{
   Name \tab Unit \tab Meta \cr
   Share of Rural Demand Satisfied by Rural Production|Primary Crop and Livestock Products \tab 0 - 1 \tab Share of rural demand met by local production \cr
   Share of Total Demand Satisfied by Rural Production|Primary Crop and Livestock Products \tab 0 - 1 \tab Share of total demand met by rural production \cr
   Share of Total Demand Potentially Satisfied by Local Production \tab 0 - 1 \tab Potential local production share \cr
}
}

\examples{
\dontrun{
x <- reportruralDemandShares(gdx)
}

}
\author{
David M Chen
}
