% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportSDG1.R
\name{reportSDG1}
\alias{reportSDG1}
\title{reportSDG1}
\usage{
reportSDG1(gdx)
}
\arguments{
\item{gdx}{GDX file}
}
\value{
MAgPIE object
}
\description{
reports all SDG indicators relevant for SDG1 - Poverty
}
\examples{

  \dontrun{
    x <- reportSDG3(gdx)
  }

}
\author{
Benjamin Bodirsky
}
