% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportSDG12.R
\name{reportSDG12}
\alias{reportSDG12}
\title{reportSDG12}
\usage{
reportSDG12(gdx)
}
\arguments{
\item{gdx}{GDX file}
}
\value{
MAgPIE object
}
\description{
reports all SDG indicators relevant for SD12 - Sustainable Production and Consumption
}
\examples{

  \dontrun{
    x <- reportSDG12(gdx)
  }

}
\author{
Benjamin Bodirsky
}
