% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportSDG15.R
\name{reportSDG15}
\alias{reportSDG15}
\title{reportSDG15}
\usage{
reportSDG15(gdx)
}
\arguments{
\item{gdx}{GDX file}
}
\value{
MAgPIE object
}
\description{
reports all SDG indicators relevant for SD15 - Life on Land
}
\examples{

  \dontrun{
    x <- reportSDG15(gdx)
  }

}
\author{
Benjamin Bodirsky, Isabelle Weindl
}
