% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportSDG2.R
\name{reportSDG2}
\alias{reportSDG2}
\title{reportSDG2}
\usage{
reportSDG2(gdx)
}
\arguments{
\item{gdx}{GDX file}
}
\value{
MAgPIE object
}
\description{
reports all SDG indicators relevant for SD2 - Hunger
}
\examples{

  \dontrun{
    x <- reportSDG2(gdx)
  }

}
\author{
Benjamin Bodirsky
}
