% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportSDG3.R
\name{reportSDG3}
\alias{reportSDG3}
\title{reportSDG3}
\usage{
reportSDG3(gdx)
}
\arguments{
\item{gdx}{GDX file}
}
\value{
MAgPIE object
}
\description{
reports all SDG indicators relevant for SDG3 - Health
}
\examples{

  \dontrun{
    x <- reportSDG3(gdx)
  }

}
\author{
Benjamin Bodirsky
}
