% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportSDG6.R
\name{reportSDG6}
\alias{reportSDG6}
\title{reportSDG6}
\usage{
reportSDG6(gdx, level = "regglo", outputdir = ".")
}
\arguments{
\item{gdx}{GDX file}

\item{level}{level of aggregation (cluster: "cell", regional: "regglo")}

\item{outputdir}{output directory}
}
\value{
MAgPIE object
}
\description{
reports all SDG indicators relevant for SDG6 - Access to Water
}
\examples{

  \dontrun{
    x <- reportSDG6(gdx)
  }

}
\author{
Felicitas Beier, Isabelle Weindl
}
