% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportSDG9.R
\name{reportSDG9}
\alias{reportSDG9}
\title{reportSDG9}
\usage{
reportSDG9(gdx)
}
\arguments{
\item{gdx}{GDX file}
}
\value{
MAgPIE object
}
\description{
reports all SDG indicators relevant for SD9 - Industrial innovation and infrastructure
}
\examples{

  \dontrun{
    x <- reportSDG9(gdx)
  }

}
\author{
Felicitas Beier
}
