% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportTau2.R
\name{reportTau2}
\alias{reportTau2}
\title{reportTau2}
\usage{
reportTau2(gdx)
}
\arguments{
\item{gdx}{GDX file}
}
\value{
tau values as MAgPIE object for managed pastures (Index)
}
\description{
reports Tau for managed pastures and croplands
}
\examples{
\dontrun{
x <- reportTau2(gdx)
}

}
\author{
Marcos Alves
}
