% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportTauGrass.R
\name{reportTauGrass}
\alias{reportTauGrass}
\title{reportTauGrass}
\usage{
reportTauGrass(gdx)
}
\arguments{
\item{gdx}{GDX file}
}
\value{
tau values as MAgPIE object for managed pastures (Index)
}
\description{
reports Tau
}
\examples{
\dontrun{
x <- reportTauGrass(gdx)
}

}
\author{
Marcos
}
