% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportTc.R
\name{reportTc}
\alias{reportTc}
\title{reportTc}
\usage{
reportTc(gdx)
}
\arguments{
\item{gdx}{GDX file}
}
\value{
tc values as MAgPIE object (%/yr)
}
\description{
reports Tc
}
\examples{

  \dontrun{
    x <- reportTc(gdx)
  }

}
\author{
Florian Humpenoeder
}
