% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportTimber.R
\name{reportTimber}
\alias{reportTimber}
\title{reportTimber}
\usage{
reportTimber(gdx)
}
\arguments{
\item{gdx}{GDX file}
}
\value{
Timber demand
}
\description{
reports MAgPIE demand for timber.
}
\examples{

  \dontrun{
    x <- reportTimber(gdx)
  }

}
\author{
Abhijeet Mishra
}
