% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportTimberDemand.R
\name{reportTimberDemand}
\alias{reportTimberDemand}
\title{reportTimberDemand}
\usage{
reportTimberDemand(gdx)
}
\arguments{
\item{gdx}{GDX file}
}
\value{
Timber demand
}
\description{
reports MAgPIE demand for timber.
}
\examples{

  \dontrun{
    x <- reportTimberDemand(gdx)
  }

}
\author{
Abhijeet Mishra
}
