% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportTimberDemandVolumetric.R
\name{reportTimberDemandVolumetric}
\alias{reportTimberDemandVolumetric}
\title{reportTimberDemandVolumetric}
\usage{
reportTimberDemandVolumetric(gdx)
}
\arguments{
\item{gdx}{GDX file}
}
\value{
Timber demand
}
\description{
reports MAgPIE demand for timber.
}
\examples{

  \dontrun{
    x <- reportTimberDemandVolumetric(gdx)
  }

}
\author{
Abhijeet Mishra
}
