% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportTimberProductionVolumetric.R
\name{reportTimberProductionVolumetric}
\alias{reportTimberProductionVolumetric}
\title{reportTimberProductionVolumetric}
\usage{
reportTimberProductionVolumetric(gdx)
}
\arguments{
\item{gdx}{GDX file}
}
\value{
Timber demand
}
\description{
reports MAgPIE production for timber.
}
\examples{

  \dontrun{
    x <- reportTimberProductionVolumetric(gdx)
  }

}
\author{
Abhijeet Mishra
}
