% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportTotalHoursWorked.R
\name{reportTotalHoursWorked}
\alias{reportTotalHoursWorked}
\title{reportTotalHoursWorked}
\usage{
reportTotalHoursWorked(gdx, level = "regglo")
}
\arguments{
\item{gdx}{GDX file}

\item{level}{spatial aggregation: "reg", "glo", "regglo"}
}
\value{
total hours worked as MAgPIE object
}
\description{
reports total hours worked in crop+livestock production (and maccs) from MAgPIE results
}
\section{Total hours worked variables}{
\tabular{lll}{
   Name \tab Unit \tab Meta \cr
   Labor|Total Hours Worked \tab mio h \tab Total hours worked in crop and livestock production \cr
}
}

\examples{

  \dontrun{
    x <- reportTotalHoursWorked(gdx)
  }


}
\author{
Debbora Leip
}
