% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportTradeGrowth.R
\name{reportTradeGrowth}
\alias{reportTradeGrowth}
\title{reportTradeGrowth}
\usage{
reportTradeGrowth(gdx, detail = FALSE)
}
\arguments{
\item{gdx}{GDX file}

\item{detail}{if true, provides results for all commodities, otherwhise aggregates some groups}
}
\value{
Trade growth rates (index)
}
\description{
reports trade growth rate
}
\section{Trade growth variables}{
\tabular{lll}{
   Name \tab Unit \tab Meta \cr
   Trade|Trade Growth Rate \tab Index \tab Trade volume index relative to base year \cr
   Trade|Trade Growth Rate|+|Crop products \tab Index \tab Crop trade growth index \cr
   Trade|Trade Growth Rate|+|Livestock products \tab Index \tab Livestock trade growth index \cr
}
}

\examples{

  \dontrun{
    x <- reportTradeGrowth(gdx="fulldata.gdx",detail=TRUE)
  }


}
\author{
Xiaoxi Wang
}
