% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportValueMaterialDemand.R
\name{reportValueMaterialDemand}
\alias{reportValueMaterialDemand}
\title{reportValueMaterialDemand}
\usage{
reportValueMaterialDemand(gdx, level = "regglo")
}
\arguments{
\item{gdx}{GDX file}

\item{level}{aggregation level of returned data ("regglo" by default)}
}
\value{
magpie object
}
\description{
reports value of material demand
}
\section{Material demand value variables}{
\tabular{lll}{
   Name \tab Unit \tab Meta \cr
   Value|Bioeconomy Demand \tab million US$2017/yr \tab Total value of bioeconomy material demand \cr
   Value|Bioeconomy Demand|+|Crops \tab million US$2017/yr \tab Value of crop products for material demand \cr
   Value|Bioeconomy Demand|+|Residues \tab million US$2017/yr \tab Value of residues for material demand \cr
}
}

\examples{
\dontrun{
x <- reportValueMaterialDemand(gdx)
}


}
\author{
David Chen
}
