% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportValueTrade.R
\name{reportValueTrade}
\alias{reportValueTrade}
\title{reportValueTrade}
\usage{
reportValueTrade(gdx, detail = FALSE)
}
\arguments{
\item{gdx}{GDX file}

\item{detail}{if true, provides estimates for all commodities, otherwhise aggregates some groups}
}
\value{
trade value as MAgPIE object Unit: see names
}
\description{
reports trade value
}
\examples{

  \dontrun{
    x <- reportValueTrade(gdx)
  }

}
\author{
Florian Humpenoeder
}
