% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportWaterAvailability.R
\name{reportWaterAvailability}
\alias{reportWaterAvailability}
\title{reportWaterAvailability}
\usage{
reportWaterAvailability(gdx)
}
\arguments{
\item{gdx}{GDX file}
}
\value{
water availability as MAgPIE object Unit: see names
}
\description{
reports water availability
}
\examples{

  \dontrun{
    x <- reportWaterAvailability(gdx)
  }

}
\author{
Felicitas Beier
}
