% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportWaterIndicatorsGlobal.R
\name{reportWaterIndicatorsGlobal}
\alias{reportWaterIndicatorsGlobal}
\title{reportWaterIndicatorsGlobal}
\usage{
reportWaterIndicatorsGlobal(gdx, outputdir = ".")
}
\arguments{
\item{gdx}{GDX file}

\item{outputdir}{output directory}
}
\value{
MAgPIE object
}
\description{
reports a set of water indicators at global level
}
\examples{
\dontrun{
x <- reportWaterIndicatorsGlobal(gdx)
}

}
\author{
Felicitas Beier
}
