% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportWaterUsage.R
\name{reportWaterUsage}
\alias{reportWaterUsage}
\title{reportWaterUsage}
\usage{
reportWaterUsage(gdx)
}
\arguments{
\item{gdx}{GDX file}
}
\value{
water usage as MAgPIE object Unit: see names
}
\description{
reports water usage
}
\examples{

  \dontrun{
    x <- reportWaterUsage(gdx)
  }

}
\author{
Florian Humpenoeder
}
