% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportWaterUsageCrops.R
\name{reportWaterUsageCrops}
\alias{reportWaterUsageCrops}
\title{reportWaterUsageCrops}
\usage{
reportWaterUsageCrops(gdx)
}
\arguments{
\item{gdx}{GDX file}
}
\value{
water usage as MAgPIE object Unit: see names
}
\description{
reports water usage for crops only
}
\examples{

  \dontrun{
    x <- reportWaterUsageCrops(gdx)
  }

}
\author{
Florian Humpenoeder
}
