% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportWaterUsageLivestock.R
\name{reportWaterUsageLivestock}
\alias{reportWaterUsageLivestock}
\title{reportWaterUsageLivestock}
\usage{
reportWaterUsageLivestock(gdx)
}
\arguments{
\item{gdx}{GDX file}
}
\value{
water usage as MAgPIE object Unit: see names
}
\description{
reports water usage for livestock only
}
\examples{

  \dontrun{
    x <- reportWaterUsageLivestock(gdx)
  }

}
\author{
Florian Humpenoeder
}
