% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportWorkingAgePopulation.R
\name{reportWorkingAgePopulation}
\alias{reportWorkingAgePopulation}
\title{reportWorkingAgePopulation}
\usage{
reportWorkingAgePopulation(gdx, level = "regglo")
}
\arguments{
\item{gdx}{GDX file}

\item{level}{spatial aggregation: "reg", "glo", "regglo", "iso"}
}
\value{
working age population as MAgPIE object
}
\description{
reports working age population
}
\section{Working age population variables}{
\tabular{lll}{
   Name \tab Unit \tab Meta \cr
   Working age population \tab million people \tab Population aged 15-64 years \cr
}
}

\examples{

  \dontrun{
    x <- reportWorkingAgePopulation(gdx)
  }


}
\author{
Debbora Leip
}
