% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportYieldsCropRaw.R
\name{reportYieldsCropRaw}
\alias{reportYieldsCropRaw}
\title{reportYieldsCropRaw}
\usage{
reportYieldsCropRaw(gdx, detail = FALSE, level = "regglo")
}
\arguments{
\item{gdx}{GDX file}

\item{detail}{if detail=FALSE, the subcategories of groups are not reported (e.g. "soybean" within "oilcrops")}

\item{level}{aggregation level of returned data ("regglo" by default)}
}
\value{
yield as MAgPIE object (t DM/ha)
}
\description{
reports potential yields before calibration
}
\details{
Uncalibrated potential yields as they come from the crop model, before management
calibration and technological change. Weighted with the 1995 cropping pattern held fixed.
}
\section{Uncalibrated input-data yield variables}{
\tabular{lll}{
   Name \tab Unit \tab Meta \cr
   Productivity|Yields|Input data|Uncalibrated \tab t DM/ha \tab Potential crop yields before calibration \cr
   Productivity|Yields|Input data|Uncalibrated|+|Cereals \tab t DM/ha \tab Uncalibrated cereal yields \cr
}
}

\examples{
\dontrun{
x <- reportYieldsCropRaw(gdx)
}

}
\author{
Edna J. Molina Bacca
}
