% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportYieldShifter.R
\name{reportYieldShifter}
\alias{reportYieldShifter}
\title{reportYieldShifter}
\usage{
reportYieldShifter(
  gdx,
  file = NULL,
  level = "reg",
  baseyear = "y2000",
  relative = TRUE
)
}
\arguments{
\item{gdx}{GDX file}

\item{file}{a file name the output should be written to using write.magpie}

\item{level}{Level of regional aggregation; "reg" (regional), "glo" (global), "regglo" (regional and global) or any other aggregation level defined in superAggregate}

\item{baseyear}{baseyear for the yield shifter. Also fixes land patterns for aggregation to baseyear.}

\item{relative}{relative or absolute changes to baseyear}
}
\value{
crop yield as MAgPIE object (unit depends on attributes)
}
\description{
Reports the Crop model input yield shifter
}
\section{Yield shifter variables}{
\tabular{lll}{
   Name \tab Unit \tab Meta \cr
   Productivity|Climate Change Yield Shifter \tab Index baseyear=1 \tab Climate-induced yield changes relative to base year \cr
   Productivity|Climate Change Yield Shifter|Cereals|Rainfed \tab Index baseyear=1 \tab Yield shifter for rainfed cereals \cr
   Productivity|Climate Change Yield Shifter|Cereals|Irrigated \tab Index baseyear=1 \tab Yield shifter for irrigated cereals \cr
}
}

\examples{

  \dontrun{
    x <- reportYieldShifter(gdx)
  }


}
\seealso{
\code{\link{reportYieldShifter}}
}
\author{
Benjamin Leon Bodirsky
}
