% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Residues.R
\name{Residues}
\alias{Residues}
\title{Residues}
\usage{
Residues(
  gdx,
  level = "regglo",
  products = "kres",
  waterAggr = TRUE,
  output = "all"
)
}
\arguments{
\item{gdx}{GDX file}

\item{level}{Level of regional aggregation; "reg" (regional), "glo" (global), "regglo" (regional and global)}

\item{products}{Selection of products (either "kcr" or "kres")}

\item{waterAggr}{Aggregate irrigated and non-irriagted production or not (boolean).}

\item{output}{Switch between different outputs: "biomass", "fieldBalance", "resDemand", all}
}
\value{
production as MAgPIE object (unit depends on attributes)
}
\description{
reads various crop residue (carbon) outputs out of a MAgPIE gdx file
}
\examples{

  \dontrun{
    x <- Residues(gdx)
  }
  @importFrom madrat toolAggregate
}
\seealso{
\code{\link{ResidueBiomass}}
}
\author{
Kristine Karstens, Michael Crawford
}
