% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_spamfile.R
\name{retrieve_spamfile}
\alias{retrieve_spamfile}
\title{retrieve_spamfile}
\usage{
retrieve_spamfile(gdx, dir)
}
\arguments{
\item{gdx}{gdx file}

\item{dir}{directory with spamfiles or clustermap*.rds}
}
\value{
dataframe
}
\description{
Provides a mapping with regions, cells and grids
}
\examples{

\dontrun{ spamfile(dir) }
}
\author{
Benjamin Leon Bodirsky, Edna J. Molina Bacca
}
