% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RotationLength.R
\name{RotationLength}
\alias{RotationLength}
\title{RotationLength}
\usage{
RotationLength(gdx, file = NULL, level = "regglo")
}
\arguments{
\item{gdx}{GDX file}

\item{file}{a file name the output should be written to using write.magpie}

\item{level}{Level of regional aggregation; "cell", "reg" (regional), "glo" (global), "regglo" (regional and global) or any secdforest aggregation level defined in superAggregate}
}
\value{
Forest rotation length
}
\description{
reads rotation length out of a MAgPIE gdx file
}
\details{
Forest rotation length
}
\examples{

  \dontrun{
    x <- RotationLength(gdx)
  }
}
\author{
Abhijeet Mishra
}
