% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ruralDemandShares.R
\name{ruralDemandShares}
\alias{ruralDemandShares}
\title{ruralDemandShares}
\usage{
ruralDemandShares(
  gdx,
  type = "tradOnly",
  level = "reg",
  product_aggr = TRUE,
  file = NULL
)
}
\arguments{
\item{gdx}{GDX file}

\item{type}{Type of ratio that should be calculated
\itemize{
       \item \code{all}: How much rural & trad demand as a share of all demand is satisfied locally
       \item \code{tradOnly}: How much rural & trad demand as a share of rural & trad demand is satisfied locally
       \item \code{potential}: How much total gridded demand is potentially
                               satisfied by gridded production
       }}

\item{level}{spatial aggregation to report employment ("reg", "glo" or "regglo")}

\item{product_aggr}{sum over products if TRUE}

\item{file}{a file name the output should be written to using write.magpie}
}
\value{
share of food consumed locally
}
\description{
reports rural demand shares based on local consumption
}
\examples{
\dontrun{
x <- localDemandShares(gdx)
}
}
\author{
David M Chen
}
