% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdgIndicator.R
\name{sdgIndicator}
\alias{sdgIndicator}
\title{sdgIndicator}
\usage{
sdgIndicator(indicatorName, unit, data)
}
\arguments{
\item{indicatorName}{The name of the SDG indicator variable (e.g., "SDG|SDG02|Food availability")}

\item{unit}{The name of the unit to be used (e.g., "kcal/cap/day")}

\item{data}{A magpie object containing the actual indicator data}
}
\value{
Returns a magpie object with the correct name
}
\description{
Helper function that renames SDG data according to a fixed naming scheme.
}
\keyword{internal}
