% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seed.R
\name{Seed}
\alias{Seed}
\title{Seed}
\usage{
Seed(gdx, level = "reg", attributes = "dm", dir = ".", spamfiledirectory = "")
}
\arguments{
\item{gdx}{GDX file}

\item{level}{Level of regional aggregation ("reg", "glo", "regglo")}

\item{attributes}{dry matter: Mt ("dm"), gross energy: PJ ("ge"), reactive nitrogen: Mt ("nr"), phosphor: Mt ("p"), potash: Mt ("k"), wet matter: Mt ("wm"). Can also be a vector.}

\item{dir}{for gridded outputs: magpie output directory which contains a mapping file (rds) for disaggregation}

\item{spamfiledirectory}{deprecated. please use \code{dir} instead}
}
\value{
demand as MAgPIE object (Unit depends on attributes)
}
\description{
Calculates MAgPIE demand for Seed out of a gdx file
}
\details{
Demand definitions are equivalent to FAO CBS categories
}
\examples{

  \dontrun{
    x <- demand(level="regglo", products="kcr")
  }

}
\author{
Benjamin Leon Bodirsky
}
