% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superAggregateX.R
\name{superAggregateX}
\alias{superAggregateX}
\title{superAggregateX}
\usage{
superAggregateX(
  data,
  aggr_type,
  level = "reg",
  weight = NULL,
  crop_aggr = FALSE
)
}
\arguments{
\item{data}{A MAgPIE object}

\item{aggr_type}{Aggregation Type. Can be any function for one or two
objects (data and weight) of the same size. Currently pre-supported
functions: "sum","mean","weighted_mean".}

\item{level}{Either a level or the name of a mapping file.
Allowed level types are global "glo", regional "reg" and "regglo".}

\item{weight}{Currently only used for weighted_mean}

\item{crop_aggr}{determines whether output should be crop-specific (FALSE)
or aggregated over all crops (TRUE). The method used for aggregation is set
by aggr_type}
}
\value{
returns a MAgPIE
object.
}
\description{
drop-in replacement for superAggregate based on toolAggregate
}
\author{
Jan Philipp Dietrich
}
