% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surplusChange.R
\name{surplusChange}
\alias{surplusChange}
\title{surplusChange}
\usage{
surplusChange(gdx1, gdx2, file = NULL, level = "reg", type = "consumer")
}
\arguments{
\item{gdx1}{GDX file from benchmark scenario}

\item{gdx2}{GDX file from the analyzed scenario}

\item{file}{a file name the output should be written to using write.magpie}

\item{level}{Level of regional aggregation; "reg" (regional), "glo" (global), "regglo" (regional and global) or any other aggregation level defined in superAggregate}

\item{type}{Economic surplus type: "consumer" (default), "producer" or "welfare"}
}
\value{
A MAgPIE object containing aggregate changes in producer surplus, consumer surplus and aggregate 
economic welfare between an analyzed scenario and a benchmark scenario, in million $US.
}
\description{
calculates aggregate change in economic surplus in mio.US$ based on a MAgPIE gdx files from two different scenarios.
}
\examples{

  \dontrun{
    x <- surplusChange(gdx1, gdx2)
  }

}
\author{
Miodrag Stevanovic
}
