% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tau.R
\name{tau}
\alias{tau}
\title{tau}
\usage{
tau(
  gdx,
  file = NULL,
  level = "reg",
  start_value = FALSE,
  digits = 4,
  prev_year = "y1985",
  type = "crop"
)
}
\arguments{
\item{gdx}{GDX file}

\item{file}{a file name the output should be written to using write.magpie}

\item{level}{Level of regional aggregation; "reg" (regional), "glo" (global),
"regglo" (regional and global) or any other aggregation level defined in superAggregate}

\item{start_value}{If TRUE, the initial values are added under the year \code{prev_year}}

\item{digits}{The result will be rounded to this number of digits}

\item{prev_year}{Year to store the initialization tau information in}

\item{type}{currently only "crop"}
}
\value{
A MAgPIE object containing tau values (index)
}
\description{
Calculates Landuse intensity indicator tau based on a MAgPIE gdx file
}
\examples{
\dontrun{
x <- tau(gdx)
}

}
\author{
Jan Philipp Dietrich, Patrick v. Jeetze
}
