% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tc.R
\name{tc}
\alias{tc}
\title{tc}
\usage{
tc(
  gdx,
  file = NULL,
  level = "reg",
  annual = TRUE,
  avrg = FALSE,
  baseyear = 1995,
  type = "crop"
)
}
\arguments{
\item{gdx}{GDX file}

\item{file}{a file name the output should be written to using write.magpie}

\item{level}{Level of regional aggregation; "reg" (regional), "glo" (global), "regglo" (regional
and global) or any other aggregation level defined in superAggregate}

\item{annual}{If TRUE, annual values are reported. If FALSE, the values for the whole timestep are
reported. If FALSE, avrg has no effect}

\item{avrg}{If FALSE the annual tc rates of the current period are returned, otherwise the average annual
tc rate for the period tbase to tn is returned. tbase defaults to the first timestep (see baseyear)}

\item{baseyear}{Determines the base year timestep for annual tc calculation. Average tc rates for later
timesteps are calculated with respect to baseyear. No tc rates for timesteps before baseyear are returned)}

\item{type}{currently only 'crop'}
}
\value{
A MAgPIE object containing tc rates. Annual ones if annual=TRUE, for the whole timestep if annual=FALSE.
}
\description{
Calculates TC rates based on a MAgPIE gdx file
}
\examples{
\dontrun{
x <- tc(gdx)
}

}
\author{
Jan Philipp Dietrich
}
