% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Timber.R
\name{Timber}
\alias{Timber}
\title{Timber}
\usage{
Timber(gdx, file = NULL, level = "regglo")
}
\arguments{
\item{gdx}{GDX file}

\item{file}{a file name the output should be written to using write.magpie}

\item{level}{Level of regional aggregation; "cell", "reg" (regional), "regglo" (regional and global), custom region aggregation
or any secdforest aggregation level defined in superAggregateX}
}
\value{
Forest demandfor timber production
}
\description{
reads timber demand out of a MAgPIE gdx file
}
\details{
Forest demandfor timber production
}
\examples{

  \dontrun{
    x <- Timber(gdx)
  }
}
\author{
Abhijeet Mishra
}
