% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimberProductionVolumetric.R
\name{TimberProductionVolumetric}
\alias{TimberProductionVolumetric}
\title{TimberProductionVolumetric}
\usage{
TimberProductionVolumetric(gdx, file = NULL, level = "regglo")
}
\arguments{
\item{gdx}{GDX file}

\item{file}{a file name the output should be written to using write.magpie}

\item{level}{Level of regional aggregation; "cell", "reg" (regional), "glo" (global), "regglo" (regional and global) or any secdforest aggregation level defined in superAggregate}
}
\value{
Forest demand for timber production
}
\description{
reads timber production out of a MAgPIE gdx file
}
\details{
Forest demand for timber production
}
\examples{

  \dontrun{
    x <- TimberProductionVolumetric(gdx)
  }
}
\author{
Abhijeet Mishra
}
