% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timePeriods.R
\name{timePeriods}
\alias{timePeriods}
\title{timePeriods}
\usage{
timePeriods(gdx)
}
\arguments{
\item{gdx}{GDX file}
}
\value{
magpie time periods as MAgPIE object as a number of years
}
\description{
Calculates MAgPIE time period lengths between each two timesteps
}
\examples{

  \dontrun{
    x <- timePeriods(gdx=gdx)
  }

}
\author{
Mishko Stevanovic
}
