% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TotalHoursWorked.R
\name{totalHoursWorked}
\alias{totalHoursWorked}
\title{totalHoursWorked}
\usage{
totalHoursWorked(gdx, level = "reg", file = NULL)
}
\arguments{
\item{gdx}{GDX file}

\item{level}{spatial aggregation to report employment ("reg", "glo", or "regglo")}

\item{file}{a file name the output should be written to using write.magpie}
}
\value{
total hours worked in agriculture per year
}
\description{
returns total hours worked per year in crop+livestock production from MAgPIE results
}
\examples{
\dontrun{
x <- totalHoursWorked(gdx)
}
}
\author{
Debbora Leip
}
