% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trade.R
\name{trade}
\alias{trade}
\title{trade}
\usage{
trade(
  gdx,
  file = NULL,
  level = "reg",
  products = "k_trade",
  productAggr = FALSE,
  attributes = "dm",
  weight = FALSE,
  relative = FALSE,
  type = "net-exports"
)
}
\arguments{
\item{gdx}{GDX file}

\item{file}{a file name the output should be written to using write.magpie}

\item{level}{Level of regional aggregation ("reg", "glo", "regglo", or name of custom mapping)}

\item{products}{Selection of products (either by naming products, e.g. "tece", or naming a set,e.g."kcr")}

\item{productAggr}{aggregate over products or not (boolean)}

\item{attributes}{dry matter: Mt ("dm"), gross energy: PJ ("ge"),
reactive nitrogen: Mt ("nr"), phosphor: Mt ("p"), potash: Mt ("k"), wet matter: Mt ("wm"). Can also be a vector.}

\item{weight}{in case relative=T also the weighting for the self sufficiencies
is provided as it is an intensive parameter}

\item{relative}{if relative=TRUE, self sufficiencies are reported,
so the amount of production divided by domestic demand}

\item{type}{exports-imports ("net-exports"), gross imports ("imports") or
gross exports ("exports"); only valid if relative=FALSE}
}
\value{
trade (production-demand) as MAgPIE object; unit depends on attributes
}
\description{
Calculates MAgPIE trade or self-sufficiencies out of a gdx file
}
\details{
Trade definitions are equivalent to FAO CBS categories
}
\examples{

  \dontrun{
    x <- trade(gdx="fulldata.gdx", level="regglo", products="kcr")
  }

}
\author{
Benjamin Leon Bodirsky, Florian Humpenoeder, Mishko Stevanovic
}
