% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tradeKastner.R
\name{tradeKastner}
\alias{tradeKastner}
\title{tradeKastner}
\usage{
tradeKastner(
  gdx,
  trade,
  level = "reg",
  products = "kall",
  attributes = "dm",
  selfselfTrade = TRUE
)
}
\arguments{
\item{gdx}{GDX file to read from (for production data)}

\item{trade}{Bilateral trade data (ov21_trade or other)}

\item{level}{Level of regional aggregation ("reg", "glo", "regglo", or custom mapping)}

\item{products}{Selection of products (e.g. "kall", "kcr", "kli")}

\item{attributes}{dry matter: Mt ("dm"), gross energy: PJ ("ge"),
reactive nitrogen: Mt ("nr"), phosphor: Mt ("p"), potash: Mt ("k"), wet matter: Mt ("wm")}

\item{selfselfTrade}{if TRUE makes the self.self trade diagonal the amount from trade matrix,
as opposed to the domestic consumption diagonal of Kastner matrix}
}
\value{
MAgPIE object with bilateral trade adjusted using Kastner method.
  Dimensions: (importer_exporter, year, product)
  Values represent apparent consumption of bilateral trade (production + imports - exports)
}
\description{
Implements Kastner et al. 2011 (DOI: 10.1016/j.ecolecon.2011.01.012) on the MAgPIE
bilateral trade matrix. Adjusts the trade matrix based on a assumption of proportionality such
that intermediate trade partners (importing to re-export) are removed from the matrix.
}
\author{
David M Chen
}
