% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tradeValue.R
\name{tradeValue}
\alias{tradeValue}
\title{tradeValue}
\usage{
tradeValue(
  gdx,
  file = NULL,
  level = "reg",
  products = "k_trade",
  product_aggr = FALSE,
  type = "net-exports",
  glo_weight = "export",
  relative = FALSE
)
}
\arguments{
\item{gdx}{GDX file}

\item{file}{a file name the output should be written to using write.magpie}

\item{level}{Level of regional aggregation ("reg", "glo", "regglo")}

\item{products}{Selection of products (either by naming products, e.g. "tece", or naming a set,e.g."kcr")}

\item{product_aggr}{aggregate over products or not (boolean)}

\item{type}{exports-imports ("net-exports"), gross imports ("imports") or gross exports ("exports"); only valid if relative=FALSE}

\item{glo_weight}{Decides the calculation of global prices. Weighting schemes are applied for estimation of global producer price.
If \code{"export"} prices are calculated as average of regional exporters' prices, weighted by the export volumes. If \code{"production"} (default),
prices are calculated as average of regional prices weighted by regional production. Alternatively, if \code{"free_trade"},
 the global prices are directly taken from the shadow prices of the global trade constraint, and no averaging is performed.
 Alternatively, if \code{"constant_prices_initial"} constant 1995 global prices for each commodity are used as weight.}

\item{relative}{if relative=TRUE, self sufficiencies are reported (the amount of production divided by domestic demand)}
}
\value{
A MAgPIE object containing the value of trade flows in Million of US dollars
}
\description{
Calculates the value of traded goods based on a gdx file
}
\examples{
\dontrun{
x <- tradeValue(gdx)
}

}
\author{
Misko Stevanovic, Florian Humpenoeder, Edna J. Molina Bacca
}
