% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tryList.R
\name{tryList}
\alias{tryList}
\title{tryList}
\usage{
tryList(..., gdx, level = "regglo")
}
\arguments{
\item{...}{report function to be run}

\item{gdx}{gdx file to report from}

\item{level}{spatial level (either "regglo" for region+global or "iso" for ISO countries)}
}
\value{
A list of magpie objects (successful reports) or NULL (failed reports)
}
\description{
Internal support function to run a list of reportings in
a \code{\link{tryReport}} environment.
}
\seealso{
\code{\link{tryReport}}, \code{\link{reportResult}}
}
\author{
Jan Philipp Dietrich
}
