% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tryReport.R
\name{tryReport}
\alias{tryReport}
\title{tryReport}
\usage{
tryReport(report, width, gdx)
}
\arguments{
\item{report}{report function to be run}

\item{width}{max number of characters per line}

\item{gdx}{gdx file to report from}
}
\description{
Internal support function to run a reporting in a try environment
and properly report problems if something goes wrong without stopping the
further processing in case of an error
}
\author{
Jan Philipp Dietrich
}
