% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validation}
\alias{validation}
\title{validation}
\usage{
validation(
  gdx,
  hist,
  file = "validation.pdf",
  runinfo = NULL,
  clusterinfo = NULL,
  debug = FALSE,
  reportfile = NULL,
  scenario = NULL,
  getReport = NULL,
  ...
)
}
\arguments{
\item{gdx}{GDX file}

\item{hist}{Validation data. All formats allowed which can be converted to quitte (including characters containing the path to a mif file)}

\item{file}{a file name the output pdf}

\item{runinfo}{(optional) Rdata object with run information}

\item{clusterinfo}{(optional) RDS file or vector containing mapping information on 0.5degree between regions and cluster}

\item{debug}{Switch to activate or deactivate debug mode}

\item{reportfile}{file name to which a backup of the magpie reporting should be written (file ending should be ".mif"). No report written if set to NULL or
if report is already provided via getReport!}

\item{scenario}{scenario name used inside reportfile. Not used if reportfile is NULL.}

\item{getReport}{the return value of the \code{getReport} function. Can be provided if available to reduce
overall runtime.}

\item{...}{additional arguments supplied to the validationpdf function}
}
\description{
Create Validation pdf from MAgPIE output and corresponding validation.mif
}
\examples{

  \dontrun{
    validation("fulldata.gdx","validation.mif",filter="Yield")
  }

}
\author{
Jan Philipp Dietrich
}
