% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ValueMaterialDemand.R
\name{ValueMaterialDemand}
\alias{ValueMaterialDemand}
\title{ValueMaterialDemand}
\usage{
ValueMaterialDemand(gdx, file = NULL, level = "regglo")
}
\arguments{
\item{gdx}{GDX file}

\item{file}{a file name the output should be written to using write.magpie}

\item{level}{aggregation level, reg, glo or regglo}
}
\description{
calculates agricultural costs without taxes and incentives (i.e. GHG taxes and BII incentives)
}
\examples{
\dontrun{
x <- ValueMaterialDemand(gdx)
}

}
\author{
David M Chen
}
