% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VegfruitShare.R
\name{VegfruitShare}
\alias{VegfruitShare}
\title{VegfruitShare}
\usage{
VegfruitShare(
  gdx,
  file = NULL,
  level = "reg",
  after_shock = TRUE,
  calibrated = TRUE,
  attributes = "kcal"
)
}
\arguments{
\item{gdx}{GDX file}

\item{file}{a file name the output should be written to using write.magpie}

\item{level}{Level of regional aggregation; "iso" ISO country codes, "reg" (regional), "glo" (global)}

\item{after_shock}{FALSE is using the exogenous real income and the prices before a shock, TRUE is using the endogeenous real income that takes into account food price change on real income}

\item{calibrated}{if FALSE, the true regression outputs are used, if TRUE the values calibrated to the start years are used}

\item{attributes}{unit: kilocalories per day ("kcal"), g protein per day ("protein"). Mt reactive nitrogen ("nr").}
}
\value{
magpie object with the livestock share in a region or country. Unit is dimensionsless, but value depends on chosen attribute
}
\description{
Calculates the share of fruits, vegetables and nuts in total food supply from the food demand model
}
\examples{

  \dontrun{
    x <- VegfruitShare(gdx)
  }

}
\author{
Benjamin Bodirsky
}
