% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wageRent.R
\name{wageRent}
\alias{wageRent}
\title{wageRent}
\usage{
wageRent(gdx, file = NULL, level = "regglo")
}
\arguments{
\item{gdx}{GDX file}

\item{file}{a file name the output should be written to using write.magpie}

\item{level}{aggregation level, reg, glo or regglo}
}
\description{
calculates wage rent for exogenous wage scenarios
}
\examples{
\dontrun{
x <- wageRent(gdx)
}

}
\author{
Debbora Leip
}
