% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/water_AAI.R
\name{water_AAI}
\alias{water_AAI}
\title{water_AAI}
\usage{
water_AAI(gdx, file = NULL, level = "reg", dir = ".")
}
\arguments{
\item{gdx}{GDX file}

\item{file}{a file name the output should be written to using \code{\link[magclass]{write.magpie}}.
See \code{\link[magclass]{write.magpie}} for supported file types}

\item{level}{spatial level of aggregation: "cell" (cellular),
"reg" (regional), "glo" (global), "regglo" (regional and global)
or any other aggregation level defined in superAggregate}

\item{dir}{for gridded outputs: magpie output directory which contains
a mapping file (rds)}
}
\value{
A MAgPIE object containing the area actually irrigated (Mha)
}
\description{
reads area actually irrigated from a MAgPIE gdx file
}
\examples{
\dontrun{
x <- water_AEI(gdx)
}
}
\author{
Stephen Wirth, Anne Biewald, Felicitas Beier
}
