% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/water_AEI.R
\name{water_AEI}
\alias{water_AEI}
\title{water_AEI}
\usage{
water_AEI(gdx, file = NULL, level = "reg")
}
\arguments{
\item{gdx}{GDX file}

\item{file}{a file name the output should be written to using write.magpie}

\item{level}{spatial level of aggregation: "cell" (cellular), "reg" (regional), "glo" (global), "regglo" (regional and global) or any other aggregation level defined in superAggregate}
}
\value{
A MAgPIE object containing the area equipped for irrigation (Mha)
}
\description{
reads area equipped for irrigation from a MAgPIE gdx file
}
\examples{

  \dontrun{
    x <- water_AEI(gdx)
  }

}
\author{
Markus Bonsch
}
