% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/water_avail.R
\name{water_avail}
\alias{water_avail}
\title{water_avail}
\usage{
water_avail(
  gdx,
  file = NULL,
  level = "reg",
  dir = ".",
  sources = NULL,
  sum = TRUE,
  digits = 4
)
}
\arguments{
\item{gdx}{GDX file}

\item{file}{a file name the output should be written to using write.magpie}

\item{level}{spatial level of aggregation: "cell" (cellular),
"reg" (regional), "glo" (global), "regglo" (regional and global)
or any other aggregation level defined in superAggregate}

\item{dir}{directory for files necessary for disaggregation}

\item{sources}{Vector of water sources that shall be obtained. NULL for all sources}

\item{sum}{Sum the contribution of different sources (TRUE) or display them individually (FALSE)}

\item{digits}{integer. For rounding of the return values}
}
\value{
A MAgPIE object containing the available water (km^3)
}
\description{
reads available water from a MAgPIE gdx file
}
\examples{
\dontrun{
x <- water_avail(gdx)
}
}
\author{
Markus Bonsch, Felicitas Beier
}
