% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/water_EFexceedance.R
\name{water_EFexceedance}
\alias{water_EFexceedance}
\title{water_EFexceedance}
\usage{
water_EFexceedance(gdx, level = "cell", users = "all", outputdir = ".")
}
\arguments{
\item{gdx}{GDX file}

\item{level}{Level of aggregation; "cell" (cluster), "reg" (regional), "glo" (global), "regglo" (regional and global)}

\item{users}{all sectors used in MAgPIE ("all") or only agricultural sector ("agr")}

\item{outputdir}{output directory}
}
\value{
A MAgPIE object containing environmental flow exceedance
}
\description{
calculation of volume of environmental flow exceedance from MAgPIE gdx file
}
\examples{

  \dontrun{
    x <- water_EFexceedance(gdx)
  }

}
\author{
Felicitas Beier
}
