% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/water_efficiency.R
\name{water_efficiency}
\alias{water_efficiency}
\title{water_efficiency}
\usage{
water_efficiency(gdx, file = NULL, level = "reg")
}
\arguments{
\item{gdx}{GDX file}

\item{file}{a file name the output should be written to using write.magpie}

\item{level}{spatial level of aggregation: "cell" (cellular), "reg" (regional), "glo" (global), "regglo" (regional and global) or any other aggregation level defined in superAggregate}
}
\value{
A MAgPIE object containing the irrigation efficiency on the requested aggregation level
}
\description{
reads Irrigation efficiency from a MAgPIE gdx file
}
\examples{

  \dontrun{
    x <- water_efficiency(gdx)
  }

}
\author{
Markus Bonsch
}
