% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/water_price.R
\name{water_price}
\alias{water_price}
\title{water_price}
\usage{
water_price(
  gdx,
  file = NULL,
  level = "reg",
  weight = "value",
  index = FALSE,
  index_baseyear = 2005,
  digits = 4
)
}
\arguments{
\item{gdx}{GDX file}

\item{file}{a file name the output should be written to using write.magpie}

\item{level}{spatial level of aggregation: "cell" (cellular), "reg" (regional), "glo" (global), "regglo" (regional and global) or any other aggregation level defined in superAggregate}

\item{weight}{For determining weights to use for generating water prices at levels beyond 'cellular'. Takes "value" and "quantity". "value" sums regional weights by value of water per cluster, "quantity" sums regional weight by qty of water per cluster}

\item{index}{FALSE (default) or TRUE}

\item{index_baseyear}{baseyear to use for index calculation (only used if index=TRUE)}

\item{digits}{integer. For rounding of the return values}
}
\value{
A MAgPIE object containing the water shadow prices (US Dollar/cubic metre).
}
\description{
reads water prices from a MAgPIE gdx file
}
\examples{

  \dontrun{
    x <- water_price(gdx)
  }

}
\author{
Markus Bonsch, Vartika Singh, Miodrag Stevanovic
}
