% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/water_usage.R
\name{water_usage}
\alias{water_usage}
\title{water_usage}
\usage{
water_usage(
  gdx,
  file = NULL,
  level = "reg",
  users = NULL,
  sum = FALSE,
  seasonality = "total",
  abstractiontype = "withdrawal",
  digits = 4
)
}
\arguments{
\item{gdx}{GDX file}

\item{file}{a file name the output should be written to using write.magpie}

\item{level}{spatial level of aggregation: "grid" (grid-cell) "cell" (cellular),
"reg" (regional), "glo" (global), "regglo" (regional and global) or
any other aggregation level defined in gdxAggregate}

\item{users}{NULL or "sectors" or "kcr" or "kli".
If NULL, all sectors including crop-wise water use and livestock will be obtained.
If sectors, will only report for high-level sectors - agriculture, industry, electricity, domestic, ecosystem.
Sum is applicable only in the case of sectors}

\item{sum}{determines whether output should be sector specific (FALSE) or aggregated over all sectors (TRUE)}

\item{seasonality}{water usage time of the year. options: "grper" (growing period) or "total" (entire year).
Note: currently only implemented for non-agricultural water usage.}

\item{abstractiontype}{water usage abstraction type: "withdrawal" or "consumption"}

\item{digits}{integer. For rounding of the return values}
}
\value{
A MAgPIE object containing the water usage (km^3/yr)
}
\description{
reads area usage from a MAgPIE gdx file
}
\examples{
\dontrun{
x <- water_usage(gdx)
}

}
\author{
Markus Bonsch, Vartika Singh, Felicitas Beier
}
