% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waterEFR.R
\name{waterEFR}
\alias{waterEFR}
\title{waterEFR}
\usage{
waterEFR(gdx, file = NULL, level = "cell", digits = 4)
}
\arguments{
\item{gdx}{GDX file}

\item{file}{a file name the output should be written to using write.magpie}

\item{level}{spatial level of aggregation: "cell" (cellular), "reg" (regional),
"glo" (global), "regglo" (regional and global)}

\item{digits}{integer. For rounding of the return values}
}
\value{
A MAgPIE object containing environmental flow requirements (km^3)
}
\description{
reads environmental flow requirements (as they enter MAgPIE)
             from a MAgPIE gdx file
}
\examples{
\dontrun{
x <- waterEFR(gdx)
}

}
\author{
Felicitas Beier
}
