% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waterEFVarea.R
\name{waterEFVarea}
\alias{waterEFVarea}
\title{waterEFVarea}
\usage{
waterEFVarea(gdx, file = NULL, level = "reg", digits = 4)
}
\arguments{
\item{gdx}{GDX file}

\item{file}{a file name the output should be written to using write.magpie}

\item{level}{spatial level of aggregation: "cell" (cellular), "reg" (regional),
"glo" (global), "regglo" (regional and global), or
"grid" (for disaggregated output using cropland as weight)}

\item{digits}{integer. For rounding of the return values}
}
\value{
A MAgPIE object containing the area under environmental flow violations (Mha)
}
\description{
calculates area that falls into cluster experiencing
             environmental flow violations
             from MAgPIE outputs
}
\examples{
\dontrun{
x <- waterEFVarea(gdx)
}
}
\author{
Felicitas Beier
}
