% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waterEFViolation.R
\name{waterEFViolation}
\alias{waterEFViolation}
\title{waterEFViolation}
\usage{
waterEFViolation(gdx, file = NULL, level = "reg", digits = 4)
}
\arguments{
\item{gdx}{GDX file}

\item{file}{a file name the output should be written to using write.magpie}

\item{level}{spatial level of aggregation: "cell" (cellular), "reg" (regional),
"glo" (global), "regglo" (regional and global), or
"grid" (for disaggregated output using cropland as weight)}

\item{digits}{integer. For rounding of the return values}
}
\value{
A MAgPIE object containing the volume of environmental flow violations (km^3)
}
\description{
calculates environmental flow violation volume
             from MAgPIE outputs
}
\examples{
\dontrun{
x <- waterEFViolation(gdx)
}

}
\author{
Felicitas Beier
}
