% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waterEFVratio.R
\name{waterEFVratio}
\alias{waterEFVratio}
\title{waterEFVratio}
\usage{
waterEFVratio(gdx, file = NULL, level = "cell")
}
\arguments{
\item{gdx}{GDX file}

\item{file}{a file name the output should be written to using write.magpie}

\item{level}{spatial level of aggregation: "cell" (cellular), "reg" (regional),
"glo" (global), "regglo" (regional and global) or
"grid" (grid cell)}
}
\value{
MAgPIE object
}
\description{
calculates ratio of environmental flow violation volume (EFV) to
             water environmental flow requirements (EFR) in MAgPIE.
}
\examples{

  \dontrun{
    x <- waterEFVratio(gdx)
  }

}
\author{
Felicitas Beier
}
