% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WaterEFVvolume.R
\name{WaterEFVvolume}
\alias{WaterEFVvolume}
\title{WaterEFVvolume}
\usage{
WaterEFVvolume(gdx, cfg, spam, input_folder, years)
}
\arguments{
\item{gdx}{GDX file}

\item{cfg}{config file}

\item{spam}{spam matrix}

\item{input_folder}{input folder}

\item{years}{years}
}
\value{
A MAgPIE object containing the volume of environmental flow violations
}
\description{
calculation of water environmental flow violation volume
}
\examples{

  \dontrun{
    x <- WaterEFVvolume(gdx)
  }

}
\author{
Markus Bonsch, Florian Humpenoeder
}
