% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waterStress.R
\name{waterStress}
\alias{waterStress}
\title{waterStress}
\usage{
waterStress(gdx, stressRatio = 0.4, file = NULL, level = "cell")
}
\arguments{
\item{gdx}{GDX file}

\item{stressRatio}{threshold defining level of water stress
(e.g. 0.2 for medium water stress,
0.4 for high water stress)}

\item{file}{a file name the output should be written to using write.magpie}

\item{level}{spatial level of aggregation: "cell" (cellular), "reg" (regional),
"glo" (global), "regglo" (regional and global)}
}
\value{
MAgPIE object indicating whether location is water stressed (1) or not (0)
}
\description{
calculates which areas are affected by water stress from water availability and water demand in MAgPIE.
             Water stress is calculated based on the proportion of water withdrawals to water availability.
             Thresholds based on World Resources Institute
             definition (https://www.wri.org/data/water-stress-country):
             Low stress: <10%
             Low-to-medium stress: 10-20%
             Medium to high stress: 20-40%
             High stress: 40-80%
             Extremely high stress: >80%
}
\examples{
\dontrun{
x <- waterStress(gdx)
}

}
\author{
Felicitas Beier
}
