% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waterStressedPopulation.R
\name{waterStressedPopulation}
\alias{waterStressedPopulation}
\title{waterStressedPopulation}
\usage{
waterStressedPopulation(
  gdx,
  file = NULL,
  level = "cell",
  dir = ".",
  absolute = TRUE
)
}
\arguments{
\item{gdx}{GDX file}

\item{file}{a file name the output should be written to using write.magpie}

\item{level}{spatial level of aggregation: "cell" (cellular), "reg" (regional),
"glo" (global), "regglo" (regional and global) or
"grid" (grid cell)}

\item{dir}{for gridded outputs:
magpie output directory which contains a mapping file (rds or spam) disaggregation}

\item{absolute}{TRUE: reports people living in water stressed region in million,
FALSE: returns share of population}
}
\value{
MAgPIE object
}
\description{
People living in water stressed region
}
\examples{

  \dontrun{
    x <- waterStressRatio(gdx)
  }

}
\author{
Felicitas Beier
}
