% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waterStressRatio.R
\name{waterStressRatio}
\alias{waterStressRatio}
\title{waterStressRatio}
\usage{
waterStressRatio(gdx, file = NULL, level = "cell")
}
\arguments{
\item{gdx}{GDX file}

\item{file}{a file name the output should be written to using write.magpie}

\item{level}{spatial level of aggregation: "cell" (cellular), "reg" (regional),
"glo" (global), "regglo" (regional and global) or
"grid" (grid cell)}
}
\value{
MAgPIE object
}
\description{
calculates water stress ratio from water availability and water demand in MAgPIE.
             Water stress ratio is the ratio of water withdrawals (in the growing period)
             to water availability (in the growing period)
}
\examples{

  \dontrun{
    x <- waterStressRatio(gdx)
  }

}
\author{
Felicitas Beier
}
