% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/woodHarvestArea.R
\name{woodHarvestArea}
\alias{woodHarvestArea}
\title{woodHarvestArea}
\usage{
woodHarvestArea(gdx)
}
\arguments{
\item{gdx}{A fulldata.gdx of a magpie run, usually with endogenous forestry enabled}
}
\value{
A magpie object with the following dimensions: region, id, year, source, ageClass
}
\description{
Reads wood harvest area separated by source (primforest,
secdforest, forestry, other) and age classes from a gdx.
The data is on cluster level and the unit is Mha per year.
}
\author{
Pascal Sauer
}
