% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/woodProduction.R
\name{woodProduction}
\alias{woodProduction}
\title{woodProduction}
\usage{
woodProduction(gdx)
}
\arguments{
\item{gdx}{A fulldata.gdx of a magpie run, usually with endogenous forestry enabled}
}
\value{
A magpie object with the following dimensions: region, id, year, source, woodType
}
\description{
Reads roundwood and fuelwood production/harvest data separated by source (primforest,
secdforest, forestry, other) from a gdx. The data is on cluster level and
the unit is Petagram (= mio. t) dry matter per year (Pg DM yr-1).
}
\author{
Pascal Sauer
}
