% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YieldsCropCalib.R
\name{YieldsCropCalib}
\alias{YieldsCropCalib}
\title{YieldsCropCalib}
\usage{
YieldsCropCalib(gdx, file = NULL, level = "cell", tau = FALSE)
}
\arguments{
\item{gdx}{GDX file}

\item{file}{a file name the output should be written to using write.magpie}

\item{level}{Level of regional aggregation}

\item{tau}{if TRUE add effect of TAU to the yield}
}
\value{
A MAgPIE object containing values of potential yields after the calibration routines.
With \code{tau = TRUE} the result equals \code{vm_yld}.
}
\description{
Reads potential yields after calibration
}
\examples{
\dontrun{
x <- YieldsCropCalib(gdx)
}

}
\author{
Edna Molina Bacca
}
