% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AgriResearchIntensity.R
\name{AgriResearchIntensity}
\alias{AgriResearchIntensity}
\title{AgriResearchIntensity}
\usage{
AgriResearchIntensity(gdx, file = NULL, level = "reg", debug = FALSE)
}
\arguments{
\item{gdx}{GDX file}

\item{file}{a file name the output should be written to using write.magpie}

\item{level}{aggregation level, reg, glo or regglo, cell or grid}

\item{debug}{debug mode TRUE makes some consistency checks between estimates for different resolutions.}
}
\description{
calculates Agricultural Research Intensity (Investment in AgR&D/Total GDP)
from a MAgPIE gdx file
}
\examples{
\dontrun{
x <- AgriResearchIntensity(gdx)
}

}
\author{
David M Chen
}
