% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BII.R
\name{BII}
\alias{BII}
\title{Biodiversity intactness index}
\usage{
BII(
  gdx,
  file = NULL,
  level = "glo",
  mode = "auto",
  landClass = "sum",
  spatialWeight = NULL,
  adjusted = FALSE,
  bii_coeff = NULL,
  side_layers = NULL
)
}
\arguments{
\item{gdx}{GDX file}

\item{file}{a file name the output should be written to using write.magpie}

\item{level}{level of regional aggregation; "cell" (magpie cluster level),
"reg" (regional), "glo" (global), "regglo" (regional and global),
"iso" (country level), "grid" (0.5 degree grid cell level).}

\item{mode}{"auto" (default), "from_grid", "MAgPIE" or "postprocessing".
\itemize{
               \item "MAgPIE" reports the BV based on values from the MAgPIE biodiversity module.
               \item "postprocessing" calculates the BV based on land information from MAgPIE (for versions where biodiversity module was not available yet).
               \item "auto" uses "MAgPIE" if available and falls back to "postprocessing" otherwise.
               \item "from_grid" calculates BII values from BII output and returns aggregated values at the aggregation level specified.
}}

\item{landClass}{"all" returns average BII values for all land classes of ov_bv,
"sum" returns the weighted BII over all land classes of ov44_bv_weighted.}

\item{spatialWeight}{Spatial weight for aggregating BII values. Only relevant if mode is "from_grid", adjusted is TRUE,
or level is either "grid" or "iso".}

\item{adjusted}{if "TRUE", function returns adjusted BII values (results have been adjusted for primary and secondary other land).}

\item{bii_coeff}{file containing BII coefficients. Only needed for mode = "postprocessing". NULL tries to automatically detected the file.}

\item{side_layers}{file containing LUH2 side layers.
NULL tries to automatically detected the file.}
}
\value{
Biodiversity intactness index (unitless)
}
\description{
calculates the area weighted biodiversity intactness index (BII) out of a MAgPIE gdx file
}
\details{
Calculates global, regional and cluster-level biodiversity intactness index (BII)
}
\examples{
\dontrun{
x <- BII(gdx)
}

}
\author{
Patrick v. Jeetze, Florian Humpenoeder, Felicitas Beier
}
