% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changelogYears.R
\name{changelogYears}
\alias{changelogYears}
\title{changelogYears}
\usage{
changelogYears()
}
\description{
Returns the years to put into the data changelog, see \code{\link{addToDataChangelog}}.
}
\author{
Pascal Sauer
}
