% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CropareaDiversityIndex.R
\name{CropareaDiversityIndex}
\alias{CropareaDiversityIndex}
\title{CropareaDiversityIndex}
\usage{
CropareaDiversityIndex(
  gdx,
  index = "shannon",
  level = "reg",
  measurelevel = "cell",
  groupdiv = "agg1",
  dir = "."
)
}
\arguments{
\item{gdx}{GDX file}

\item{index}{can be "shannon", "gini" or "invsimpson" for different types of diversitiy indices}

\item{level}{Level of regional aggregation; "reg" (regional), "glo" (global),
"regglo" (regional and global) or
any other aggregation level defined in superAggregate}

\item{measurelevel}{level at which diversity is measured. "cell" means diversity}

\item{groupdiv}{should crop groups be split up into several individual items or not? Choose either FALSE or different (dis)aggregation methods "agg1", "agg2"}

\item{dir}{for gridded outputs: magpie output directory which contains a mapping file (rds) for disaggregation
is measured at cellular level, even if lateron average diversity is aggregated
to regional level.}
}
\value{
MAgPIE object (unit depends on attributes)
}
\description{
calculates an index that measures the croparea diversity
}
\examples{
\dontrun{
x <- CropareaDiversityIndex(gdx)
}

}
\seealso{
\code{\link{CropareaDiversityIndex}}
}
\author{
Benjamin Leon Bodirsky
}
