% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emisSOC.R
\name{emisSOC}
\alias{emisSOC}
\title{emisSOC}
\usage{
emisSOC(gdx, file = NULL, sumLand = FALSE)
}
\arguments{
\item{gdx}{GDX file}

\item{file}{a file name the output should be written to using write.magpie}

\item{sumLand}{TRUE (default) or FALSE. Sum over land types (TRUE)
or report land-type specific emissions (FALSE).}
}
\value{
CO2 emissions as MAgPIE object
}
\description{
reads detailed CO2 soil emissions out of a MAgPIE gdx file
}
\examples{
\dontrun{
x <- emisSOC(gdx)
}
}
\author{
Krstine Karstens
}
