% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorCosts.R
\name{factorCosts}
\alias{factorCosts}
\title{factorCosts}
\usage{
factorCosts(gdx, products = "kli", file = NULL, level = "regglo")
}
\arguments{
\item{gdx}{GDX file}

\item{products}{products for which factor costs should be reported ("kcr",
"kli", "kres", "fish", or "pasture")}

\item{file}{a file name the output should be written to using write.magpie}

\item{level}{Level of regional aggregation ("reg", "glo", "regglo")}
}
\value{
MAgPIE object containing factor costs [million US$17]
}
\description{
reads factor costs for crops, livestock, residues or pasture
entering the objective function from a MAgPIE gdx file. Depending on the
product and the MAgPIE version (and factor cost realization), factor costs
are either already split into labor and capital, will be split in this
function, or are kept as the aggregate
}
\examples{
\dontrun{
x <- factorCosts(gdx)
}

}
\author{
Debbora Leip
}
