% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorCostShares.R
\name{factorCostShares}
\alias{factorCostShares}
\title{factorCostShares}
\usage{
factorCostShares(
  gdx,
  type = "optimization",
  products = "kcr",
  level = "reg",
  file = NULL
)
}
\arguments{
\item{gdx}{GDX file}

\item{type}{\itemize{
  \item "requirements": shares from factor requirements
  \item "optimization": cost shares between labor and capital costs in optimization
  \item "accounting": cost shares based on accounting of labor and capital costs
}}

\item{products}{products for which cost shares should be reported, kcr or kli}

\item{level}{spatial aggregation to report employment ("reg", "glo" or "regglo")}

\item{file}{a file name the output should be written to using write.magpie}
}
\value{
labor and capital cost share out of factor costs
}
\description{
returns labor and capital cost share out of factor costs (i.e. labor + capital)
}
\examples{
\dontrun{
x <- factorCostShares(gdx)
}
}
\author{
Debbora Leip
}
