% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fallow.R
\name{Fallow}
\alias{Fallow}
\title{Fallow}
\usage{
Fallow(gdx, level = "reg", dir = ".", debug = FALSE)
}
\arguments{
\item{gdx}{GDX file}

\item{level}{aggregation level, reg, glo or regglo, cell or grid}

\item{dir}{for gridded outputs: magpie output directory which contains a mapping file (rds or spam) disaggregation}

\item{debug}{debug mode TRUE makes some consistency checks between estimates for different resolutions.}
}
\description{
calculates fallow land (Mha) from a MAgPIE gdx file
}
\examples{
\dontrun{
x <- Fallow(gdx)
}

}
\author{
Benjamin Leon Bodirsky
}
