% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fallow.R
\name{fallow}
\alias{fallow}
\title{fallow}
\usage{
fallow(gdx, level = "reg", debug = FALSE)
}
\arguments{
\item{gdx}{GDX file}

\item{level}{aggregation level, reg, glo or regglo, cell or grid}

\item{debug}{debug mode TRUE makes some consistency checks
between estimates for different resolutions}
}
\description{
calculates fallow land (Mha) from a MAgPIE gdx file
}
\examples{
\dontrun{
x <- fallow(gdx)
}

}
\author{
Benjamin Leon Bodirsky
}
