% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ForestYield.R
\name{ForestYield}
\alias{ForestYield}
\title{ForestYield}
\usage{
ForestYield(gdx, file = NULL, level = "cell")
}
\arguments{
\item{gdx}{GDX file}

\item{file}{a file name the output should be written to using write.magpie}

\item{level}{Level of regional aggregation; "cell", "reg" (regional), "glo" (global), "regglo" (regional and global) or any secdforest aggregation level defined in superAggregate}
}
\value{
Forest yield for timber production in tDM per ha per year
}
\description{
calculates timber yield out of a MAgPIE gdx file
}
\details{
Forest yield for timber production
}
\examples{

  \dontrun{
    x <- ForestYield(gdx)
  }
}
\author{
Abhijeet Mishra, Florian Humpenoeder
}
