% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getReportFSECCropDiversityGrid.R
\name{getReportFSECCropDiversityGrid}
\alias{getReportFSECCropDiversityGrid}
\title{getReportFSECCropDiversityGrid}
\usage{
getReportFSECCropDiversityGrid(gdx, reportOutputDir = NULL, scenario = NULL)
}
\arguments{
\item{gdx}{a GDX file}

\item{reportOutputDir}{a folder name for the output to be written to. If NULL the report is not saved to
disk, and only returned to the calling function.}

\item{scenario}{the name of the scenario used. If NULL the report is not saved to disk, and only returned to the
calling function.}
}
\value{
A list of MAgPIE objects containing the reports
}
\description{
Reports grid cell level crop diversity for the FSEC project
}
\examples{
\dontrun{
x <- getReportFSECCropDiversityGrid(gdx)
}

}
\author{
Patrick v. Jeetze
}
