% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getReportFSECPollution.R
\name{getReportFSECPollution}
\alias{getReportFSECPollution}
\title{getReportFSECPollution}
\usage{
getReportFSECPollution(
  gdx,
  reportOutputDir = NULL,
  magpieOutputDir,
  scenario = NULL
)
}
\arguments{
\item{gdx}{a GDX file}

\item{reportOutputDir}{a folder name for the output to be written to. If NULL the report is not saved to
disk, and only returned to the calling function.}

\item{magpieOutputDir}{a magpie output directory which contains a mapping file (clustermap*.rds) for the
disaggregation of grid output}

\item{scenario}{the name of the scenario used. If NULL the report is not saved to disk, and only returned to the
calling function.}
}
\value{
A list of MAgPIE objects containing the reports
}
\description{
Reports nutrient surplus indicators for the FSEC project
}
\examples{

  \dontrun{
    x <- getReportFSECPollution(gdx, magpieOutputDir)
  }

}
\author{
Michael Crawford
}
