% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getReportFSECStevenLord.R
\name{getReportFSECStevenLord}
\alias{getReportFSECStevenLord}
\title{getReportFSECStevenLord}
\usage{
getReportFSECStevenLord(
  magpieOutputDir,
  reportOutputDir = NULL,
  scenario = NULL
)
}
\arguments{
\item{magpieOutputDir}{a magpie output directory which contains a mapping file (clustermap*.rds) for the
disaggregation of grid output}

\item{reportOutputDir}{a folder name for the output to be written to. If NULL the report is not saved to
disk, and only returned to the calling function.}

\item{scenario}{the name of the scenario used. If NULL the report is not saved to disk, and only returned to the
calling function.}
}
\value{
A list of MAgPIE objects containing the reports
}
\description{
Collects reports for Steven Lord's Cost of Action / Cost of Inaction analysis.
Specifically:
- Total land (Mh)
- Nutrient surplus (Mt N)
- Biodiversity (BII)
- Population (Millions people)
- Global Surface Temperature (deg C)
- GDP (PPP)
- GHG emissions (CO2, CH4, NO2, as well as totalled and cumulative in CO2eq)
- Food system costs
- Tau
}
\examples{

  \dontrun{
    x <- getReportFSECStevenLord(magpieOutputDir)
  }

}
\author{
Michael Crawford
}
