% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getReportFSECStevenLord.R
\name{getReportFSECStevenLord}
\alias{getReportFSECStevenLord}
\title{getReportFSECStevenLord}
\usage{
getReportFSECStevenLord(magpieOutputDir, reportOutputDir, scenario)
}
\arguments{
\item{magpieOutputDir}{a magpie output directory which contains a mapping file (clustermap*.rds) for the
disaggregation of grid output}

\item{reportOutputDir}{a folder name for the output to be written to. If NULL the report is not saved to
disk, and only returned to the calling function.}

\item{scenario}{the name of the scenario used. If NULL the report is not saved to disk, and only returned to the
calling function.}
}
\description{
Collects reports for Steven Lord's cost of action / cost of inaction analysis.
}
\examples{

  \dontrun{
    x <- getReportFSECStevenLord(magpieOutputDir)
  }

}
\author{
Michael Crawford
}
