% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getReportFSECStevenLord.R
\name{getReportFSECStevenLord}
\alias{getReportFSECStevenLord}
\title{getReportFSECStevenLord}
\usage{
getReportFSECStevenLord(
  gdx,
  reportOutputDir = NULL,
  magpieOutputDir,
  scenario = NULL
)
}
\arguments{
\item{gdx}{a GDX file}

\item{reportOutputDir}{a folder name for the output to be written to. If NULL the report is not saved to
disk, and only returned to the calling function.}

\item{magpieOutputDir}{a magpie output directory which contains a mapping file (clustermap*.rds) for the
disaggregation of grid output}

\item{scenario}{the name of the scenario used. If NULL the report is not saved to disk, and only returned to the
calling function.}
}
\value{
A list of MAgPIE objects containing the reports
}
\description{
Reports cropland, pasture, and nitrogen pollution indicators for the FSEC project (specifically Steven Lord)
}
\examples{

  \dontrun{
    x <- getReportFSECStevenLord(gdx, magpieOutputDir)
  }

}
\author{
Michael Crawford
}
