% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getReportGridNitrogenPollution.R
\name{getReportGridNitrogenPollution}
\alias{getReportGridNitrogenPollution}
\title{getReportGridNitrogenPollution}
\usage{
getReportGridNitrogenPollution(
  magpieOutputDir,
  reportOutputDir = NULL,
  scenario = NULL
)
}
\arguments{
\item{magpieOutputDir}{a magpie output directory which contains a mapping file (clustermap*.rds) for the
disaggregation of grid output}

\item{reportOutputDir}{a folder name for the output to be written to. If NULL the report is not saved to
disk, and only returned to the calling function.}

\item{scenario}{the name of the scenario used. If NULL the report is not saved to disk, and only returned to the
calling function.}
}
\value{
A list of MAgPIE objects containing the reports
}
\description{
Reports nutrient surplus indicators as well as exceedance of the critical nitrogen surplus at the
grid level
}
\examples{

  \dontrun{
    x <- getReportGridNitrogenPollution(magpieOutputDir)
  }

}
\author{
Michael Crawford
}
