% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harvested_area_timber.R
\name{harvested_area_timber}
\alias{harvested_area_timber}
\title{harvested_area_timber}
\usage{
harvested_area_timber(gdx, file = NULL, level = "cell")
}
\arguments{
\item{gdx}{GDX file}

\item{file}{a file name the output should be written to using write.magpie}

\item{level}{Level of regional aggregation; "cell", "reg" (regional), "glo" (global), "regglo" (regional and global) or any secdforest aggregation level defined in superAggregate}
}
\value{
Area harvested for timber production in Mha per year
}
\description{
reads harvested area for timber production out of a MAgPIE gdx file
}
\details{
Area harvested for timber production: timber plantations, primary forest, secondary forest and non-forest land (woodfuel only)
}
\examples{
\dontrun{
x <- harvested_area_timber(gdx)
}

}
\author{
Abhijeet Mishra
}
