% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IntakeDetailedProtein.R
\name{IntakeDetailedProtein}
\alias{IntakeDetailedProtein}
\title{IntakeDetailedProtein}
\usage{
IntakeDetailedProtein(gdx, file = NULL)
}
\arguments{
\item{gdx}{GDX file}
}
\value{
Product disaggregated Protein intake as MAgPIE object at regional level (unit: grams/cap/day)
}
\description{
Calculates detailed per-capita (protein in grams) intake from magpie results at regional level
}
\examples{

  \dontrun{
    x <- IntakeDetailedProtein(gdx)
  }

}
\author{
Vartika Singh, Isabelle Weindl
}
