% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landTransitionMatrix.R
\name{landTransitionMatrix}
\alias{landTransitionMatrix}
\title{landTransitionMatrix}
\usage{
landTransitionMatrix(gdx, file = NULL, level = "reg")
}
\arguments{
\item{gdx}{GDX file}

\item{file}{a file name the output should be written to using write.magpie}

\item{level}{Level of regional aggregation; "reg" (regional), "glo" (global), "regglo" (regional and global)
or any other aggregation level defined in gdxAggregate}
}
\value{
land transition matrix as MAgPIE object (Mha/yr). The full 7x7 matrix
including diagonal (persistence) is returned.
}
\description{
reads the land transition matrix out of a MAgPIE gdx file
}
\examples{
\dontrun{
x <- landTransitionMatrix(gdx)
}

}
\seealso{
\code{\link{reportLandTransitionMatrix}}
}
\author{
Florian Humpenoeder
}
