% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lastIter.R
\name{lastIter}
\alias{lastIter}
\title{lastIter}
\usage{
lastIter(gdx, param, MagpieOutput = TRUE)
}
\arguments{
\item{gdx}{GDX file}

\item{param}{Parameter to be returned}

\item{MagpieOutput}{In inelastic model runs, the food demand model is run once
in iter1 based on exo prices prescribed in iter0, then the magpie model in
iter1. The magpie results therefore appear in iter1.
In elastic model runs, iter2 is started by rerunning the food demand model,
but if convergence is reached, MAgPIE does not run again. MAgPIE results
are therefore still in iter1, even if the iteration is in iter2.}
}
\value{
magpie object
}
\description{
Returns the value of a parameter in the last iteration
}
\examples{
\dontrun{
x <- lastIter(gdx)
}

}
\author{
Benjamin Leon Bodirsky
}
