% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localDemandShares.R
\name{localDemandShares}
\alias{localDemandShares}
\title{localDemandShares}
\usage{
localDemandShares(
  gdx,
  type = "prod",
  level = "reg",
  product_aggr = TRUE,
  urb_aggr = TRUE,
  fvc_aggr = TRUE,
  file = NULL
)
}
\arguments{
\item{gdx}{GDX file}

\item{type}{"prod" or "dem". the former indicates the share of production consumed in its own cluster
while the latter indicates how much cluster-level demand is satisfied by local consumption}

\item{level}{spatial aggregation to report employment ("reg", "glo" or "regglo")}

\item{product_aggr}{sum over products if TRUE}

\item{urb_aggr}{sum over products if TRUE}

\item{fvc_aggr}{sum over products if TRUE}

\item{file}{a file name the output should be written to using write.magpie}
}
\value{
share or level of food consumed locally, in disaggregated transport module
}
\description{
reports local demand and production shares based on local consumption
}
\examples{
\dontrun{
x <- localDemandShares(gdx)
}
}
\author{
David M Chen
}
