% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magpie4-package.R
\docType{package}
\name{magpie4-package}
\alias{magpie4-package}
\alias{magpie4}
\title{MAgPIE outputs R library for MAgPIE version 4.x}
\description{
Common output routines for extracting results from the MAgPIE framework (versions 4.x).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/pik-piam/magpie4}
  \item \doi{10.5281/zenodo.1158582}
  \item Report bugs at \url{https://github.com/pik-piam/magpie4/issues}
}

}
\author{
Maintainer: Jan Philipp Dietrich <dietrich@pik-potsdam.de>
}
