% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ManureExcretion.R
\name{ManureExcretion}
\alias{ManureExcretion}
\title{ManureExcretion}
\usage{
ManureExcretion(
  gdx,
  level = "reg",
  products = "kli",
  awms = c("grazing", "stubble_grazing", "fuel", "confinement"),
  agg = TRUE,
  disagg_lvst = "landbased"
)
}
\arguments{
\item{gdx}{GDX file}

\item{level}{aggregation level: glo, reg, cell, grid, iso}

\item{products}{livestock products}

\item{awms}{large animal waste management categories: "grazing","stubble_grazing","fuel","confinement"),}

\item{agg}{aggregation over "awms" or over "products".}

\item{disagg_lvst}{Livestock grid-level disaggregation method: "landbased" (default) splits
ruminant manure by awms category (grazing/fuel weighted by pasture production, stubble_grazing/
confinement weighted by cropland production) and monogastric manure by development state
(urban vs cropland weighted); "glw" disaggregates using the gridded livestock distribution file
(f71_livestock_distribution_0.5.mz, produced by mrland::calcLivestockDistribution; must be
present next to gdx).}
}
\value{
MAgPIE object
}
\description{
downscales Manure Excretion
}
\examples{

  \dontrun{
    x <- ManureExcretion(gdx)
  }

}
\author{
Benjamin Leon Bodirsky, Bin Lin
}
