% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ManureExcretion.R
\name{ManureExcretion}
\alias{ManureExcretion}
\title{ManureExcretion}
\usage{
ManureExcretion(
  gdx,
  level = "reg",
  products = "kli",
  awms = c("grazing", "stubble_grazing", "fuel", "confinement"),
  agg = TRUE
)
}
\arguments{
\item{gdx}{GDX file}

\item{level}{aggregation level: glo, reg, cell, grid, iso}

\item{products}{livestock products}

\item{awms}{large animal waste management categories: "grazing","stubble_grazing","fuel","confinement"),}

\item{agg}{aggregation over "awms" or over "products".}
}
\value{
MAgPIE object
}
\description{
downscales Manure Excretion
}
\examples{

  \dontrun{
    x <- ManureExcretion(gdx)
  }

}
\author{
Benjamin Leon Bodirsky
}
