% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NitrogenBudget.R
\name{NitrogenBudget}
\alias{NitrogenBudget}
\title{NitrogenBudget}
\usage{
NitrogenBudget(
  gdx,
  include_emissions = FALSE,
  level = "reg",
  dir = ".",
  spamfiledirectory = "",
  debug = FALSE
)
}
\arguments{
\item{gdx}{GDX file}

\item{include_emissions}{TRUE also divides the N surplus into different emissions}

\item{level}{aggregation level, reg, glo or regglo, cell or grid}

\item{dir}{for gridded outputs: magpie output directory which contains a mapping file (rds or spam) disaggregation}

\item{spamfiledirectory}{deprecated. please use \code{dir} instead}

\item{debug}{debug mode TRUE makes some consistency checks between estimates for different resolutions.}
}
\description{
calculates projections of Nitrogen Budgets for Croplands from a MAgPIE gdx file
}
\examples{

  \dontrun{
    x <- NitrogenBudget(gdx)
  }

}
\author{
Benjamin Leon Bodirsky, Michael Crawford
}
