% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NitrogenBudget.R
\name{NitrogenBudget}
\alias{NitrogenBudget}
\title{NitrogenBudget}
\usage{
NitrogenBudget(
  gdx,
  include_emissions = FALSE,
  level = "reg",
  dir = ".",
  debug = FALSE,
  cropTypes = FALSE
)
}
\arguments{
\item{gdx}{GDX file}

\item{include_emissions}{TRUE also divides the N surplus into different emissions}

\item{level}{aggregation level, reg, glo or regglo, cell, iso or grid}

\item{dir}{for gridded outputs: magpie output directory which contains a mapping file (rds) for disaggregation}

\item{debug}{debug mode TRUE makes some consistency checks between estimates for different resolutions.}

\item{cropTypes}{FALSE for aggregate results; TRUE for crop-specific results}
}
\description{
calculates projections of Nitrogen Budgets for Croplands (Tg Nr per) from a MAgPIE gdx file
}
\examples{
\dontrun{
x <- NitrogenBudget(gdx)
}

}
\author{
Benjamin Leon Bodirsky, Michael Crawford, Edna J. Molina Bacca, Florian Humpenoeder
}
