% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population.R
\name{population}
\alias{population}
\title{population}
\usage{
population(
  gdx,
  file = NULL,
  level = "reg",
  age = FALSE,
  sex = FALSE,
  bmi_groups = FALSE,
  magpie_input = "auto",
  dir = ".",
  spamfiledirectory = ""
)
}
\arguments{
\item{gdx}{GDX file}

\item{file}{a file name the output should be written to using write.magpie}

\item{level}{Level of regional aggregation; "reg" (regional),
"glo" (global), "regglo" (regional and global)
or any other aggregation level defined in superAggregate}

\item{age}{if TRUE, population is split up by age groups}

\item{sex}{if TRUE, population is split up by sex}

\item{bmi_groups}{if TRUE, the population will be split up in body-mass-index groups.}

\item{magpie_input}{Available modes are "auto" (default), TRUE or FALSE.
This setting is only activated if argument "bmi_groups" is set to TRUE.
If set to TRUE, BMI distribution is estimated ex-post
such that it corresponds to the per-capita kcal intake values
finally driving MAgPIE dynamics.
In cases where exogenous diet scenarios (e.g. EAT Lancet diets)
are simulated, these ex-post BMI distribution can diverge from the (calibrated)
regression outputs from the food demand model.
If set to FALSE, the BMI distribution as calculated in the
food demand model is used, which might not be consistent with
the intake and calorie supply used in a MAgPIE simulation in
the case of exogenous diet scenarios (e.g. EAT Lancet diets).
The default setting "auto" detects automatically whether
an exogenous scenario for per-capita kcal intake is simulated by MAgPIE,
and uses the respective settings:
1) ex-post estimate in case of exogenous scenarios and
2) estimates from the food demand model in case of endogenous scenarios.}

\item{dir}{for gridded outputs: magpie output directory which contains
a mapping file (rds or spam) disaggregation}

\item{spamfiledirectory}{deprecated. please use \code{dir} instead}
}
\value{
population as MAgPIE object (million people)
}
\description{
reads population out of a MAgPIE gdx file
}
\examples{
\dontrun{
x <- population(gdx)
}

}
\seealso{
\code{\link{reportPopulation}}
}
\author{
Florian Humpenoeder, Benjamin Bodirsky, Isabelle Weindl
}
