% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{processing}
\alias{processing}
\title{processing}
\usage{
processing(
  gdx,
  file = NULL,
  level = "reg",
  products = readGDX(gdx, "kall"),
  product_aggr = FALSE,
  attributes = "dm",
  type = NULL,
  indicator = "secondary_from_primary"
)
}
\arguments{
\item{gdx}{GDX file}

\item{file}{a file name the output should be written to using write.magpie}

\item{level}{Level of regional aggregation ("reg", "glo", "regglo")}

\item{products}{Selection of products (either by naming products, e.g. "tece", or naming a set,e.g."kcr")}

\item{product_aggr}{aggregate over products or not (boolean)}

\item{attributes}{dry matter: Mt ("dm"), gross energy: PJ ("ge"), reactive nitrogen: Mt ("nr"), phosphor: Mt ("p"), potash: Mt ("k"), wet matter: Mt ("wm"). Can also be a vector.}

\item{type}{Demand type(s): "Food", "Feed", "Processing", "Material", "Bioenergy", "Seed", "Supply chain loss", "Domestic Balanceflow"; NULL returns all types}

\item{indicator}{process or secondary product output}
}
\value{
processing as MAgPIE object (Unit depends on attributes)
}
\description{
Calculates MAgPIE disaggregated processing out of a gdx file
}
\details{
Demand definitions are equivalent to FAO CBS categories
}
\examples{

  \dontrun{
    x <- processing(gdx = gdx, level="regglo", products="kcr", indicator="primary_to_process")
  }

}
\author{
David Chen, Benjamin Leon Bodirsky
}
