% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{processing}
\alias{processing}
\title{processing}
\usage{
processing(gdx, level = "reg", indicator = "secondary_from_primary")
}
\arguments{
\item{gdx}{GDX file}

\item{level}{Level of regional aggregation ("reg", "glo", "regglo")}

\item{indicator}{process or secondary product output}
}
\value{
processing as MAgPIE object (Unit depends on attributes)
}
\description{
Calculates MAgPIE disaggregated processing out of a gdx file
}
\details{
Demand definitions are equivalent to FAO CBS categories
}
\examples{

  \dontrun{
    x <- processing(gdx = gdx, level="regglo", products="kcr", indicator="primary_to_process")
  }

}
\author{
David Chen, Benjamin Leon Bodirsky
}
