% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportAAI.R
\name{reportAAI}
\alias{reportAAI}
\title{reportAAI}
\usage{
reportAAI(gdx)
}
\arguments{
\item{gdx}{GDX file}
}
\value{
Area actually irrigated as MAgPIE object. Unit: see names
}
\description{
reports Area actually irrigated
}
\examples{

  \dontrun{
    x <- reportAEI(gdx)
  }

}
\author{
Stephen Wirth, Anne Biewald
}
