% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportAEI.R
\name{reportAEI}
\alias{reportAEI}
\title{reportAEI}
\usage{
reportAEI(gdx, level = "regglo")
}
\arguments{
\item{gdx}{GDX file}
}
\value{
Area equipped for Irrigation as MAgPIE object. Unit: see names
}
\description{
reports Area equipped for Irrigation
}
\section{Area equipped for irrigation variables}{
\tabular{lll}{
   Name \tab Unit \tab Meta \cr
   Resources|Land Cover|Cropland|Area equipped for irrigation \tab million ha \tab Cropland area equipped with irrigation infrastructure \cr
}
}

\examples{

  \dontrun{
    x <- reportAEI(gdx)
  }

}
\author{
Stephen Wirth
}
