% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportAgEmployment.R
\name{reportAgEmployment}
\alias{reportAgEmployment}
\title{reportAgEmployment}
\usage{
reportAgEmployment(gdx, type = "absolute", level = "regglo")
}
\arguments{
\item{gdx}{GDX file}

\item{type}{"absolute" for total number of people employed, "share" for share out of working age population}

\item{level}{spatial aggregation: "reg", "glo", "regglo", "iso"}
}
\value{
agricultural employment as MAgPIE object
}
\description{
reports employment in crop+livestock production from MAgPIE results
}
\examples{

  \dontrun{
    x <- reportAgEmployment(gdx)
  }

}
\author{
Debbora Leip
}
