% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportAgEmployment.R
\name{reportAgEmployment}
\alias{reportAgEmployment}
\title{reportAgEmployment}
\usage{
reportAgEmployment(gdx, type = "absolute", detail = FALSE, level = "regglo")
}
\arguments{
\item{gdx}{GDX file}

\item{type}{"absolute" for total number of people employed, "share" for share out of working age population}

\item{detail}{if TRUE, employment is disaggregated to crop and livestock production, if FALSE only aggregated
employment is reported}

\item{level}{spatial aggregation: "reg", "glo", "regglo", "iso"}
}
\value{
agricultural employment as MAgPIE object
}
\description{
reports employment in crop+livestock production from MAgPIE results
}
\section{Agricultural employment variables}{
\tabular{lll}{
   Name \tab Unit \tab Meta \cr
   Labor|Employment|Agricultural employment \tab mio people \tab Total agricultural employment \cr
   Labor|Employment|Agricultural employment|+|Crop products \tab mio people \tab Employment in crop production \cr
   Labor|Employment|Agricultural employment|+|Livestock products \tab mio people \tab Employment in livestock production \cr
   Labor|Employment|Share of working age population employed in agriculture \tab \% \tab Agricultural employment as share of working age population \cr
}
}

\examples{

  \dontrun{
    x <- reportAgEmployment(gdx)
  }


}
\author{
Debbora Leip
}
