% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportAgGDP.R
\name{reportAgGDP}
\alias{reportAgGDP}
\title{reportAgGDP}
\usage{
reportAgGDP(gdx, level = "regglo")
}
\arguments{
\item{gdx}{GDX file}

\item{level}{aggregation level of returned data ("regglo" by default)}
}
\value{
Magpie object
}
\description{
reports MAgPIE Agricultural GDP Mio. USD05 MER
}
\section{Agricultural GDP variables}{
\tabular{lll}{
   Name \tab Unit \tab Meta \cr
   Value|Agriculture GDP \tab million US$2017/yr \tab Agricultural value added (GDP from agriculture sector) \cr
}
}

\examples{
\dontrun{
x <- reportAgGDP(gdx)
}

}
\author{
Edna J. Molina Bacca
}
