% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportAgriResearchIntensity.R
\name{reportAgriResearchIntensity}
\alias{reportAgriResearchIntensity}
\title{reportAgriResearchIntensity}
\usage{
reportAgriResearchIntensity(gdx)
}
\arguments{
\item{gdx}{GDX file}
}
\value{
magpie object
}
\description{
reports Agricultural Research Intensity as \% of total GDP
}
\section{Agricultural research intensity variables}{
\tabular{lll}{
   Name \tab Unit \tab Meta \cr
   Agricultural Research Intensity \tab \% of Total GDP \tab Share of GDP spent on agricultural research \cr
}
}

\examples{
\dontrun{
x <- reportAgriResearchIntensity(gdx)
}


}
\author{
David Chen
}
