% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportAnthropometrics.R
\name{reportAnthropometrics}
\alias{reportAnthropometrics}
\title{reportAnthropometrics}
\usage{
reportAnthropometrics(gdx, level = "regglo")
}
\arguments{
\item{gdx}{GDX file}

\item{level}{spatial aggregation: "reg", "glo", "regglo", "iso"}
}
\value{
Magpie object
}
\description{
reports Underweight, Normalweight, Overweight and Obesity as well as body height for males and females
}
\examples{
\dontrun{
x <- reportBodyweight(gdx)
}
}
\author{
Benjamin Leon Bodirsky
}
