% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportCarbonstock.R
\name{reportCarbonstock}
\alias{reportCarbonstock}
\title{reportCarbonstock}
\usage{
reportCarbonstock(gdx)
}
\arguments{
\item{gdx}{GDX file}
}
\description{
Reports the carbon stocks for future MAgPIE projections
}
\examples{
  \dontrun{
    x <- reportSOM(gdx)
  }
}
\author{
Kristine Karstens
}
