% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportCattle.R
\name{reportCattle}
\alias{reportCattle}
\title{reportCattle}
\usage{
reportCattle(gdx)
}
\arguments{
\item{gdx}{GDX file}
}
\value{
Cattle values as magpie objetc
}
\description{
reports cattle related numbers
}
\examples{
\dontrun{
x <- reportCattle(gdx)
}


}
\author{
Marcos Alves
}
