% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportConsumVal.R
\name{reportConsumVal}
\alias{reportConsumVal}
\title{reportConsumVal}
\usage{
reportConsumVal(gdx)
}
\arguments{
\item{gdx}{GDX file}
}
\value{
Magpie object associated with the consumption value
}
\description{
reports MAgPIE consumption value
}
\examples{
\dontrun{
x <- reportConsumVal(gdx)
}
}
\author{
Edna J. Molina Bacca
}
