% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportCostCapitalInvestment.R
\name{reportCostCapitalInvestment}
\alias{reportCostCapitalInvestment}
\title{reportCostCapitalInvestment}
\usage{
reportCostCapitalInvestment(gdx)
}
\arguments{
\item{gdx}{GDX file}
}
\value{
Magpie object associated with overall costs and value of production
}
\description{
reports MAgPIE capital investments
}
\examples{
\dontrun{
x <- reportCostCapitalInvestment(gdx)
}
}
\author{
Edna J. Molina Bacca
}
