% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportCostCapitalStocks.R
\name{reportCostCapitalStocks}
\alias{reportCostCapitalStocks}
\title{reportCostCapitalStocks}
\usage{
reportCostCapitalStocks(gdx)
}
\arguments{
\item{gdx}{GDX file}
}
\value{
Magpie object associated with overall costs and value of production
}
\description{
reports MAgPIE capital stocks
}
\examples{
\dontrun{
x <- reportCostCapitalStocks(gdx)
}
}
\author{
Edna J. Molina Bacca
}
