% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportCapitalStocks.R
\name{reportCostCapitalStocks}
\alias{reportCostCapitalStocks}
\title{reportCostCapitalStocks}
\usage{
reportCostCapitalStocks(gdx)
}
\arguments{
\item{gdx}{GDX file}
}
\value{
Magpie object associated with overall costs and value of production
}
\description{
reports MAgPIE capital stocks
}
\section{Capital stock variables}{
\tabular{lll}{
   Name \tab Unit \tab Meta \cr
   Capital Stocks| Arable farm capital \tab million US$2017 \tab Capital stocks used in cropland (sticky cost implementation) \cr
}
}

\examples{
\dontrun{
x <- reportCostCapitalStocks(gdx)
}
}
\author{
Edna J. Molina Bacca
}
