% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportCostOverall.R
\name{reportCostOverall}
\alias{reportCostOverall}
\title{reportCostOverall}
\usage{
reportCostOverall(gdx)
}
\arguments{
\item{gdx}{GDX file}
}
\value{
Magpie object associated with overall costs and value of production
}
\description{
reports MAgPIE costs
}
\examples{

  \dontrun{
    x <- reportCostOverall(gdx)
  }
}
\author{
Edna J. Molina Bacca
}
