% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportCostsAccounting.R
\name{reportCostsAccounting}
\alias{reportCostsAccounting}
\title{reportCostsAccounting}
\usage{
reportCostsAccounting(gdx, level = "regglo")
}
\arguments{
\item{gdx}{GDX file}

\item{level}{An aggregation level for the spatial dimension. Can be any level
available via superAggregateX.}
}
\value{
Costs accounting including total investments
}
\description{
reports MAgPIE costs including total investments
}
\section{Cost accounting variables}{
\tabular{lll}{
   Name \tab Unit \tab Meta \cr
   Costs Accounting \tab million US$2017/yr \tab Total cost accounting including investments \cr
   Costs Accounting|+|Land Conversion \tab million US$2017/yr \tab Investment costs for land conversion \cr
   Costs Accounting|+|Transport \tab million US$2017/yr \tab Transport cost investments \cr
   Costs Accounting|+|TC \tab million US$2017/yr \tab Technological change investment costs \cr
}
}

\examples{
\dontrun{
x <- reportCostsAccounting(gdx)
}

}
\author{
Edna J. Molina Bacca
}
