% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportCostsAEI.R
\name{reportCostsAEI}
\alias{reportCostsAEI}
\title{reportCostsAEI}
\usage{
reportCostsAEI(gdx)
}
\arguments{
\item{gdx}{GDX file}
}
\value{
magpie object containing AEI costs
}
\description{
reports MAgPIE AEI costs
}
\examples{

  \dontrun{
    x <- reportCostsAEI(gdx)
  }

}
\author{
Felicitas Beier
}
