% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportCostsInputFactors.R
\name{reportCostsInputFactors}
\alias{reportCostsInputFactors}
\title{reportFactorCosts}
\usage{
reportCostsInputFactors(gdx)
}
\arguments{
\item{gdx}{GDX file}
}
\value{
magpie object with factor costs
}
\description{
reports MAgPIE factor costs (split into labor and capital for sticky realization)
}
\examples{
\dontrun{
x <- reportCostsInputFactors(gdx)
}

}
\author{
Debbora Leip
}
