% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportCostsPresolve.R
\name{reportCostsPresolve}
\alias{reportCostsPresolve}
\title{reportCostsPresolve}
\usage{
reportCostsPresolve(gdx)
}
\arguments{
\item{gdx}{GDX file}
}
\value{
consumption value as MAgPIE object Unit: see names
}
\description{
reports MAgPIE costs
}
\examples{

  \dontrun{
    x <- reportCostsPresolve(gdx)
  }

}
\author{
Florian Humpenoeder
}
