% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportCostsWholesale.R
\name{reportCostsWholesale}
\alias{reportCostsWholesale}
\title{reportCostsWholesale}
\usage{
reportCostsWholesale(gdx, level = "regglo")
}
\arguments{
\item{gdx}{GDX file}

\item{level}{Level of regional aggregation; "reg" (regional), "glo" (global),
"regglo" (regional and global) or any other aggregation level defined in superAggregate}
}
\value{
A MAgPIE object containing values related with costs wholesale trade [million US$17/tDM]
}
\description{
Reads data to calculate wholesale costs
}
\examples{
\dontrun{
x <- reportCostsWholesale(gdx)
}

}
\author{
David M Chen
}
