% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportCostsWithoutIncentives.R
\name{reportCostsWithoutIncentives}
\alias{reportCostsWithoutIncentives}
\title{reportCostsWithoutIncentives}
\usage{
reportCostsWithoutIncentives(gdx, level = "regglo")
}
\arguments{
\item{gdx}{GDX file}

\item{level}{spatial aggregation: "reg", "glo", "regglo"}
}
\value{
magpie object
}
\description{
reports Costs Without Incentives
}
\section{Costs without incentives variables}{
\tabular{lll}{
   Name \tab Unit \tab Meta \cr
   Costs Accounting|Costs without incentives \tab million US$2017/yr \tab Total costs excluding policy incentives \cr
}
}

\examples{
\dontrun{
x <- reportCostsWithoutIncentives(gdx)
}


}
\author{
David Chen
}
