% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportCostsWithoutIncentives.R
\name{reportCostsWithoutIncentives}
\alias{reportCostsWithoutIncentives}
\title{reportCostsWithoutIncentives}
\usage{
reportCostsWithoutIncentives(gdx)
}
\arguments{
\item{gdx}{GDX file}
}
\value{
magpie object
}
\description{
reports Costs Without Incentives
}
\examples{
\dontrun{
x <- reportCostsWithoutIncentives(gdx)
}


}
\author{
David Chen
}
