% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportCostTC.R
\name{reportCostTC}
\alias{reportCostTC}
\title{reportCostTC}
\usage{
reportCostTC(gdx)
}
\arguments{
\item{gdx}{GDX file}
}
\value{
magpie object with TC costs
}
\description{
reports MAgPIE TC costs
}
\section{TC cost variables}{
\tabular{lll}{
   Name \tab Unit \tab Meta \cr
   Costs|TC \tab million US$2017/yr \tab Technological change investment costs \cr
}
}

\examples{
\dontrun{
x <- reportCostTC(gdx)
}


}
\author{
David Chen
}
