% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportCostTransport.R
\name{reportCostTransport}
\alias{reportCostTransport}
\title{reportCostTransport}
\usage{
reportCostTransport(gdx, level = "regglo")
}
\arguments{
\item{gdx}{GDX file}

\item{level}{An aggregation level for the spatial dimension. Can be any level
available via superAggregateX.}
}
\value{
consumption value as MAgPIE object Unit: see names
}
\description{
reports MAgPIE costs
}
\section{Transport cost variables}{
\tabular{lll}{
   Name \tab Unit \tab Meta \cr
   Costs|Transport \tab million US$2017/yr \tab Total transport costs for agricultural commodities \cr
   Costs|Transport|+|Crops \tab million US$2017/yr \tab Transport costs for crop products \cr
   Costs|Transport|+|Livestock products \tab million US$2017/yr \tab Transport costs for livestock products \cr
}
}

\examples{
  \dontrun{
    x <- reportCostTransport(gdx)
  }

}
\author{
David Chen
}
