% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportCropareaGrid.R
\name{reportCropareaGrid}
\alias{reportCropareaGrid}
\title{reportCropareaGrid}
\usage{
reportCropareaGrid(gdx, dir = ".", spamfiledirectory = "")
}
\arguments{
\item{gdx}{GDX file}

\item{dir}{for gridded outputs: magpie output directory which contains a mapping file (rds or spam) disaggregation}

\item{spamfiledirectory}{deprecated. please use \code{dir} instead}
}
\value{
Croparea as MAgPIE object (million ha/yr)
}
\description{
reports croparea
}
\examples{
\dontrun{
x <- reportCropareaGrid(gdx)
}

}
\author{
Benjamin Bodirsky
}
