% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportCropareaGrid.R
\name{reportCropareaGrid}
\alias{reportCropareaGrid}
\title{reportCropareaGrid}
\usage{
reportCropareaGrid(gdx)
}
\arguments{
\item{gdx}{GDX file}
}
\value{
Croparea as MAgPIE object (million ha/yr)
}
\description{
reports croparea
}
\section{Grid-level croparea}{

This function produces grid-level (0.5 degree) croparea data for individual crops.
Variable names follow the reportingnames mapping (e.g., Cereals, Oilcrops, etc.).
}

\examples{
\dontrun{
x <- reportCropareaGrid(gdx)
}

}
\author{
Benjamin Bodirsky
}
