% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportCropDiversity.R
\name{reportCropDiversity}
\alias{reportCropDiversity}
\title{reportCropDiversity}
\usage{
reportCropDiversity(gdx, grid = FALSE)
}
\arguments{
\item{gdx}{GDX file}

\item{grid}{Set to TRUE, if outputs should be reported on 0.5 degree grid level}
}
\value{
Crop diversity as MAgPIE object
}
\description{
reports crop diversity
}
\examples{
\dontrun{
x <- reportCropDiversity(gdx)
}

}
\author{
Patrick v. Jeetze
}
