% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportDemand.R
\name{reportDemand}
\alias{reportDemand}
\title{reportDemand}
\usage{
reportDemand(gdx, detail = FALSE, agmip = FALSE)
}
\arguments{
\item{gdx}{GDX file}

\item{detail}{if detail=F, the subcategories of groups are not reported (e.g. "soybean" within "oilcrops")}

\item{agmip}{if agmip=T, additional sector aggregates required for agmip are reported (e.g. "AGR")}
}
\value{
demand as MAgPIE object (Mt DM)
}
\description{
reports Demand for Food, Feed, Processing, Material, Bioenergy, Seed and Supply Chain Loss
}
\examples{

  \dontrun{
    x <- reportDemand()
  }

}
\author{
Benjamin Leon Bodirsky, Isabelle Weindl
}
