% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportDemandBioenergy.R
\name{reportDemandBioenergy}
\alias{reportDemandBioenergy}
\title{reportDemandBioenergy}
\usage{
reportDemandBioenergy(gdx, detail = FALSE)
}
\arguments{
\item{gdx}{GDX file}

\item{detail}{if detail=F, the subcategories of groups are not reported (e.g. "soybean" within "oilcrops")}
}
\value{
Bioenergy demand as MAgPIE object (EJ/yr)
}
\description{
reports Bioenergy Demand in EJ/yr
}
\examples{

  \dontrun{
    x <- reportDemandBioenergy()
  }
}
\author{
Florian Humpenoeder, Kristine Karstens
}
