% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportDemandNr.R
\name{reportDemandNr}
\alias{reportDemandNr}
\title{reportDemandNr}
\usage{
reportDemandNr(gdx, detail = FALSE)
}
\arguments{
\item{gdx}{GDX file}

\item{detail}{if detail=F, the subcategories of groups are not reported (e.g. "soybean" within "oilcrops")}
}
\value{
demand as MAgPIE object (Mt DM)
}
\description{
Similar to reportDemand, but for nitrogen. reports Demand for Food, Feed, Processing, Material, Bioenergy, Seed and Supply Chain Loss
}
\section{Nitrogen demand variables}{
\tabular{lll}{
   Name \tab Unit \tab Meta \cr
   Demand|Food \tab Mt Nr/yr \tab Nitrogen demand for food \cr
   Demand|Feed \tab Mt Nr/yr \tab Nitrogen demand for feed \cr
   Demand|Processing \tab Mt Nr/yr \tab Nitrogen demand for processing \cr
   Demand|Bioenergy \tab Mt Nr/yr \tab Nitrogen demand for bioenergy \cr
   Demand|Material \tab Mt Nr/yr \tab Nitrogen demand for materials \cr
}
}

\examples{

  \dontrun{
    x <- reportDemand()
  }


}
\author{
Benjamin Leon Bodirsky
}
