% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportEmissions.R
\name{reportEmissions}
\alias{reportEmissions}
\title{reportEmissions}
\usage{
reportEmissions(gdx, storage = TRUE)
}
\arguments{
\item{gdx}{GDX file}

\item{storage}{Accounting for long term carbon storage in wood products. Default is TRUE}
}
\value{
GHG emissions as MAgPIE object (Unit: Mt CO2/yr, Mt N2O/yr and Mt CH4/yr)
}
\description{
reports GHG emissions
}
\examples{

\dontrun{
x <- reportEmissions(gdx)
}

}
\author{
Florian Humpenoeder, Benjamin Leon Bodirsky
}
