% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportExpenditureFoodIndex.R
\name{reportExpenditureFoodIndex}
\alias{reportExpenditureFoodIndex}
\title{reportExpenditureFoodIndex}
\usage{
reportExpenditureFoodIndex(
  gdx,
  baseyear = "y2010",
  basketyear = "y2010",
  level = "regglo"
)
}
\arguments{
\item{gdx}{GDX file}

\item{baseyear}{Baseyear of the price index}

\item{basketyear}{Year of reference food basket
(should be in the past for comparison of different runs to
have identical and comparable food basket)}

\item{level}{aggregation level of returned data ("regglo" by default)}
}
\value{
Food expenditure index as MAgPIE object
}
\description{
reports food expenditure index and food expenditure index
corrected for emission costs
}
\section{Food expenditure index variables}{
\tabular{lll}{
   Name \tab Unit \tab Meta \cr
   Prices|Food Expenditure Index \tab Index 2010=100 \tab Food expenditure index (all food products) \cr
   Prices|Food Expenditure Index|Plant-based food products \tab Index 2010=100 \tab Food expenditure index for plant-based products \cr
   Prices|Food Expenditure Index|Livestock food products \tab Index 2010=100 \tab Food expenditure index for livestock products \cr
   Prices|Food Expenditure Index corrected for ghg costs \tab Index 2010=100 \tab Food expenditure index corrected for GHG costs \cr
   Prices|Agricultural Primary Products Expenditure Index \tab Index 2010=100 \tab Agricultural primary products expenditure index \cr
}
}

\examples{
\dontrun{
x <- reportPriceFoodIndex(gdx)
}

}
\author{
Felicitas Beier
}
