% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportExtraResidueEmissions.R
\name{reportExtraResidueEmissions}
\alias{reportExtraResidueEmissions}
\title{reportExtraResidueEmissions.R}
\usage{
reportExtraResidueEmissions(gdx, level = "reg")
}
\arguments{
\item{gdx}{GDX file}

\item{level}{Level of regional aggregation; "reg" (regional), "glo" (global), "regglo" (regional and global)}
}
\value{
emissions as MAgPIE object (unit: Mt X/yr, plus cumulative Mt X/yr)
}
\description{
reads residue biomass from a MAgPIE gdx file and multiplies DM by emission factors from GFED
}
\examples{
  \dontrun{
    x <- reportExtraResidueEmissions(gdx, level = "glo")
  }
}
\author{
Michael Crawford
}
