% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportFactorCostShares.R
\name{reportFactorCostShares}
\alias{reportFactorCostShares}
\title{reportFactorCostShares}
\usage{
reportFactorCostShares(gdx, level = "regglo")
}
\arguments{
\item{gdx}{GDX file}

\item{level}{spatial aggregation: "reg", "glo", "regglo"}
}
\value{
labor and capital cost shares as MAgPIE object
}
\description{
reports labor and capital cost share out of factor costs from MAgPIE results
}
\examples{
\dontrun{
x <- reportFactorCostShares(gdx)
}

}
\author{
Debbora Leip
}
