% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportFactorCostShares.R
\name{reportFactorCostShares}
\alias{reportFactorCostShares}
\title{reportFactorCostShares}
\usage{
reportFactorCostShares(gdx, type = "optimization", level = "regglo")
}
\arguments{
\item{gdx}{GDX file}

\item{type}{\itemize{
\item "requirements": shares from factor requirements
\item "optimization": cost shares between labor and capital costs in optimization
\item "accounting": cost shares based on accounting of labor and capital costs
}}

\item{level}{spatial aggregation: "reg", "glo", "regglo"}
}
\value{
labor and capital cost shares as MAgPIE object
}
\description{
reports labor and capital cost share out of factor costs from MAgPIE results
}
\section{Factor cost share variables}{
\tabular{lll}{
   Name \tab Unit \tab Meta \cr
   Factor cost shares optimization|Crop products|+|Labor cost share \tab \% \tab Labor cost share in crop production \cr
   Factor cost shares optimization|Crop products|+|Capital cost share \tab \% \tab Capital cost share in crop production \cr
   Factor cost shares optimization|Livestock products|+|Labor cost share \tab \% \tab Labor cost share in livestock production \cr
   Factor cost shares optimization|Livestock products|+|Capital cost share \tab \% \tab Capital cost share in livestock production \cr
}
}

\examples{
\dontrun{
x <- reportFactorCostShares(gdx)
}

}
\author{
Debbora Leip
}
