% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportFeed.R
\name{reportFeed}
\alias{reportFeed}
\title{reportFeed}
\usage{
reportFeed(gdx, detail = TRUE, level = "regglo")
}
\arguments{
\item{gdx}{GDX file}

\item{detail}{if detail=F, the subcategories of groups are not reported (e.g. "soybean" within "oilcrops")}

\item{level}{An aggregation level for the spatial dimension. Can be any level
available via superAggregateX.}
}
\value{
feed demand as MAgPIE object (Mt DM)
}
\description{
reportes feed demand by animal type
}
\section{Feed demand variables}{
\tabular{lll}{
   Name \tab Unit \tab Meta \cr
   Demand|Feed|++|Feed for Ruminant meat \tab Mt DM/yr \tab Feed demand for ruminant meat production \cr
   Demand|Feed|++|Feed for Dairy \tab Mt DM/yr \tab Feed demand for dairy production \cr
   Demand|Feed|++|Feed for Pig meat \tab Mt DM/yr \tab Feed demand for pig meat production \cr
   Demand|Feed|++|Feed for Poultry meat \tab Mt DM/yr \tab Feed demand for poultry meat production \cr
   Demand|Feed|++|Feed for Eggs \tab Mt DM/yr \tab Feed demand for egg production \cr
   Demand|Feed|++|Feed for Aquaculture \tab Mt DM/yr \tab Feed demand for aquaculture \cr
}
}

\examples{

  \dontrun{
    x <- reportFeed()
  }

}
\author{
Isabelle Weindl
}
