% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportFeedConversion.R
\name{reportFeedConversion}
\alias{reportFeedConversion}
\title{reportFeedConversion}
\usage{
reportFeedConversion(gdx, livestockSystem = TRUE, balanceflow = FALSE)
}
\arguments{
\item{gdx}{GDX file}

\item{livestockSystem}{if TRUE, ruminant products and poultry products are aggregated}

\item{balanceflow}{If true, feed includes the calibration balanceflow}
}
\value{
feed demand as MAgPIE object (Mt DM)
}
\description{
reportes feed demand by animal type
}
\section{Feed conversion variables}{
\tabular{lll}{
   Name \tab Unit \tab Meta \cr
   Productivity|Feed conversion \tab GE per GE \tab Overall feed conversion efficiency (gross energy) \cr
   Productivity|Feed conversion|Ruminant meat and dairy \tab GE per GE \tab Feed conversion for ruminant products \cr
   Productivity|Feed conversion|Poultry meat and eggs \tab GE per GE \tab Feed conversion for poultry products \cr
   Productivity|Feed conversion|Pig meat \tab GE per GE \tab Feed conversion for pig meat \cr
   Productivity|Roughage share|Ruminant meat and dairy \tab GE per GE \tab Share of roughage in ruminant feed \cr
   Productivity|Pasture share|Ruminant meat and dairy \tab GE per GE \tab Share of pasture in ruminant feed \cr
}
}

\examples{

  \dontrun{
    x <- reportFeed()
  }


}
\author{
Benjamin Bodirsky
}
