% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportFertilizerNitrogen.R
\name{reportFertilizerNitrogen}
\alias{reportFertilizerNitrogen}
\title{reportFertilizerNitrogen}
\usage{
reportFertilizerNitrogen(gdx, level = "regglo")
}
\arguments{
\item{gdx}{GDX file}

\item{level}{level of output}
}
\description{
Reports inorganic nitrogen application on crops
}
\section{Fertilizer nitrogen variables}{
\tabular{lll}{
   Name \tab Unit \tab Meta \cr
   Resources|Nitrogen|Inorganic Fertilizer Application \tab Mt Nr/yr \tab Total inorganic nitrogen fertilizer application \cr
   Resources|Nitrogen|Inorganic Fertilizer Application|+|Cereals \tab Mt Nr/yr \tab Nitrogen fertilizer application on cereals \cr
   Resources|Nitrogen|Inorganic Fertilizer Application|+|Oilcrops \tab Mt Nr/yr \tab Nitrogen fertilizer application on oilcrops \cr
}
}

\examples{
\dontrun{
x <- reportFertilizerNitrogen(gdx)
}

}
\seealso{
\code{\link{NitrogenBudget}}
}
\author{
David M Chen
}
