% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportFit.R
\name{reportFit}
\alias{reportFit}
\title{reportFit}
\usage{
reportFit(gdx, type = "MPE")
}
\arguments{
\item{gdx}{GDX file}

\item{type}{type of indicator. Options: R2, MAE,  MPE (mean percentage error - bias), MAPE (mean absolute percentage error)}
}
\value{
Selected error indicator
}
\description{
reports fit and error indicators compared to initial values
}
\examples{

  \dontrun{
    x <- reportFit(gdx,type)
  }

}
\author{
Edna Molina Bacca
}
