% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportFit.R
\name{reportFit}
\alias{reportFit}
\title{reportFit}
\usage{
reportFit(gdx, type = "MAPE", level = "cell")
}
\arguments{
\item{gdx}{GDX file}

\item{type}{type of indicator. Options: R2, MAE,  MPE (mean percentage error - bias), MAPE (mean absolute percentage error)}

\item{level}{level at which the regional and global bias should be reported. Options "cell" or "grid"}
}
\value{
Selected error indicator
}
\description{
reports fit and error indicators compared to initial values
}
\section{Fit indicator variables}{
\tabular{lll}{
   Name \tab Unit \tab Meta \cr
   Fit|Cluster|Land Cover \tab MAPE/MAE/MPE \tab Fit indicator for overall land cover \cr
   Fit|Cluster|Land Cover|Cropland \tab MAPE/MAE/MPE \tab Fit indicator for cropland \cr
   Fit|Cluster|Land Cover|Pasture \tab MAPE/MAE/MPE \tab Fit indicator for pasture \cr
   Fit|Cluster|Land Cover|Forest|Natural Forest|Primary Forest \tab MAPE/MAE/MPE \tab Fit indicator for primary forest \cr
   Fit|Cluster|Land Cover|Forest|Natural Forest|Secondary Forest \tab MAPE/MAE/MPE \tab Fit indicator for secondary forest \cr
}
}

\examples{

  \dontrun{
    x <- reportFit(gdx,type)
  }


}
\author{
Edna Molina Bacca, Patrick v. Jeetze
}
