% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportFoodExpenditure.R
\name{reportFoodExpenditure}
\alias{reportFoodExpenditure}
\title{reportFoodExpenditure}
\usage{
reportFoodExpenditure(gdx, detail = FALSE, level = "regglo")
}
\arguments{
\item{gdx}{GDX file}

\item{detail}{if detail=F, the subcategories of groups are not reported (e.g. "soybean" within "oilcrops")}

\item{level}{spatial aggregation: "reg", "glo", "regglo", "iso"}
}
\value{
per-capita expenditure as magpie object in reporting format
}
\description{
reports per-capita calories food supply (including household waste)
}
\examples{

  \dontrun{
    x <- reportFoodExpenditure(gdx)
  }

}
\author{
Benjamin Leon Bodirsky
}
