% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportGraslandSoilCarbon.R
\name{reportGraslandSoilCarbon}
\alias{reportGraslandSoilCarbon}
\title{reportGraslandSoilCarbon}
\usage{
reportGraslandSoilCarbon(gdx, dir = ".", spamfiledirectory = "")
}
\arguments{
\item{gdx}{GDX file}

\item{dir}{dir}

\item{spamfiledirectory}{old spamfiledirectory}
}
\value{
Cattle values as magpie objetc
}
\description{
reports cattle related numbers
}
\examples{
\dontrun{
x <- reportGraslandSoilCarbon(gdx)
}
}
\author{
Marcos Alves
}
