% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportGrasslandManagement.R
\name{reportGrasslandManagement}
\alias{reportGrasslandManagement}
\title{reportGrasslandManagement}
\usage{
reportGrasslandManagement(gdx)
}
\arguments{
\item{gdx}{GDX file}
}
\value{
Cattle values as magpie objetc
}
\description{
reports cattle related numbers
}
\examples{
\dontrun{
x <- reportGrasslandManagement(gdx)
}


}
\author{
Marcos Alves
}
