% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportGrasslandYields.R
\name{reportGrasslandYields}
\alias{reportGrasslandYields}
\title{reportGrasslandYields}
\usage{
reportGrasslandYields(gdx)
}
\arguments{
\item{gdx}{GDX file}
}
\value{
yield as MAgPIE object (Mt DM/ha)
}
\description{
reportGrasslandYields
}
\examples{

  \dontrun{
    x <- reportGrasslandYields(gdx)
  }

}
\author{
Marcos Alves
}
