% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportGrassStats.R
\name{reportGrassStats}
\alias{reportGrassStats}
\title{reportGrassStats}
\usage{
reportGrassStats(gdx)
}
\arguments{
\item{gdx}{GDX file}
}
\value{
magpie object
}
\description{
report evaluation values for pasture management implementation
}
\section{Grass statistics variables}{
\tabular{lll}{
   Name \tab Unit \tab Meta \cr
   Expansion costs|+|Pasture \tab mio USD05MER \tab Pasture expansion costs \cr
   Expansion costs|+|Range \tab mio USD05MER \tab Range expansion costs \cr
   Calibration factor|+|Pasture \tab 1 \tab Grassland calibration factor for pasture \cr
   Lambda|+|Pasture \tab 1 \tab Lambda parameter for pasture management \cr
   Productivity|Yield (before calibration)|+|Pasture \tab tDM/ha \tab Uncalibrated pasture yield \cr
}
}

\examples{
\dontrun{
x <- reportGrassStats(gdx)
}


}
\author{
Marcos Alves
}
