% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportGridCroparea.R
\name{reportGridCroparea}
\alias{reportGridCroparea}
\title{reportGridCroparea}
\usage{
reportGridCroparea(gdx)
}
\arguments{
\item{gdx}{GDX file}
}
\value{
area of cropland as MAgPIE object (million ha)
}
\description{
reports Croparea from gridded (disaggregated) output
}
\examples{

  \dontrun{
    x <- reportGridCroparea(gdx)
  }

}
\author{
Jannes Breier
}
