% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportGridCroparea.R
\name{reportGridCroparea}
\alias{reportGridCroparea}
\title{reportGridCroparea}
\usage{
reportGridCroparea(gdx)
}
\arguments{
\item{gdx}{GDX file}
}
\value{
area of cropland as MAgPIE object (million ha)
}
\description{
reports Croparea from gridded (disaggregated) output
}
\section{Grid-level croparea by irrigation}{

This function produces grid-level (0.5 degree) croparea data by crop type and irrigation system.
Variable names follow the reportingnames mapping (e.g., Cereals.Rainfed, Oilcrops.Irrigated).
}

\examples{

  \dontrun{
    x <- reportGridCroparea(gdx)
  }


}
\author{
Jannes Breier
}
