% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportGridLand.R
\name{reportGridLand}
\alias{reportGridLand}
\title{reportGridLand}
\usage{
reportGridLand(gdx)
}
\arguments{
\item{gdx}{GDX file}
}
\value{
land-use as MAgPIE object (million ha)
}
\description{
reports land-use from gridded (disaggregated) output
}
\section{Grid-level land use}{

This function produces grid-level (0.5 degree) land use data for land cover categories
(cropland, pasture, forest, urban, other land). Variable names follow the reportingnames mapping.
}

\examples{

  \dontrun{
    x <- reportGridLand(gdx)
  }


}
\author{
Jannes Breier
}
