% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportGridManureExcretion.R
\name{reportGridManureExcretion}
\alias{reportGridManureExcretion}
\title{reportGridManureExcretion}
\usage{
reportGridManureExcretion(gdx)
}
\arguments{
\item{gdx}{GDX file}
}
\value{
MAgPIE object
}
\description{
reports Manure with reprting names on grid level.
}
\examples{
\dontrun{
x <- reportGridManureExcretion(gdx)
}

}
\author{
Benjamin Leon Bodirsky
}
