% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportGridManureExcretion.R
\name{reportGridManureExcretion}
\alias{reportGridManureExcretion}
\title{reportGridManureExcretion}
\usage{
reportGridManureExcretion(gdx, dir = ".", spamfiledirectory = "")
}
\arguments{
\item{gdx}{GDX file}

\item{dir}{for gridded outputs: magpie output directory which contains a mapping file (rds or spam) disaggregation}

\item{spamfiledirectory}{deprecated. please use \code{dir} instead}
}
\value{
MAgPIE object
}
\description{
reports Manure with reprting names on grid level.
}
\examples{
\dontrun{
x <- reportGridManureExcretion(gdx)
}

}
\author{
Benjamin Leon Bodirsky
}
