% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportharvested_area_timber.R
\name{reportharvested_area_timber}
\alias{reportharvested_area_timber}
\title{reportharvested_area_timber}
\usage{
reportharvested_area_timber(gdx)
}
\arguments{
\item{gdx}{GDX file}
}
\value{
Area harvested for timber production
}
\description{
reports MAgPIE harvested area for timber.
}
\section{Timber harvested area variables}{
\tabular{lll}{
   Name \tab Unit \tab Meta \cr
   Resources|Timber operations|Harvested area for timber|Forestry \tab mha per yr \tab Area harvested from managed forests \cr
   Resources|Timber operations|Harvested area for timber|Primary forest \tab mha per yr \tab Area harvested from primary forests \cr
   Resources|Timber operations|Harvested area for timber|Secondary forest \tab mha per yr \tab Area harvested from secondary forests \cr
}
}

\examples{

  \dontrun{
    x <- reportharvested_area_timber(gdx)
  }


}
\author{
Abhijeet Mishra
}
