% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportHourlyLaborCosts.R
\name{reportHourlyLaborCosts}
\alias{reportHourlyLaborCosts}
\title{reportHourlyLaborCosts}
\usage{
reportHourlyLaborCosts(gdx)
}
\arguments{
\item{gdx}{GDX file}
}
\value{
hourly labor costs as MAgPIE object
}
\description{
reports hourly labor costs in agriculture from MAgPIE results
}
\examples{

  \dontrun{
    x <- reportHourlyLaborCosts(gdx)
  }

}
\author{
Debbora Leip
}
