% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportIncome.R
\name{reportIncome}
\alias{reportIncome}
\title{reportIncome}
\usage{
reportIncome(gdx, type = "ppp", level = "regglo")
}
\arguments{
\item{gdx}{GDX file}

\item{type}{ppp for purchase power parity, mer for market exchange rate}

\item{level}{spatial aggregation: "reg", "glo", "regglo", "iso"}
}
\value{
Annual per capita and total income as MAgPIE object (US$2005 MER/cap/yr and million US$05 PPP/yr)
}
\description{
reports income
}
\examples{
\dontrun{
x <- reportIncome(gdx)
}

}
\author{
Florian Humpenoeder, Isabelle Weindl, Felicitas Beier
}
