% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportIntakeDetailed.R
\name{reportIntakeDetailed}
\alias{reportIntakeDetailed}
\title{reportIntakeDetailed}
\usage{
reportIntakeDetailed(gdx, detail = TRUE, level = "regglo")
}
\arguments{
\item{gdx}{GDX file}

\item{detail}{if detail=F, the subcategories of groups are not reported (e.g. "soybean" within "oilcrops")}

\item{level}{spatial aggregation: "reg", "glo", "regglo", "iso"}
}
\value{
per-capita calorie intake as MAgPIE object (kcal/cap/day)
}
\description{
reports detailed or aggregated per-capita kcal intake including exogenous scenarios
}
\section{Calorie intake variables}{
\tabular{lll}{
   Name \tab Unit \tab Meta \cr
   Nutrition|Calorie Intake \tab kcal/capita/day \tab Total per-capita calorie intake (excluding waste) \cr
   Nutrition|Calorie Intake|+|Crops \tab kcal/capita/day \tab Calorie intake from crops \cr
   Nutrition|Calorie Intake|+|Livestock products \tab kcal/capita/day \tab Calorie intake from livestock products \cr
   Nutrition|Calorie Intake|+|Secondary products \tab kcal/capita/day \tab Calorie intake from secondary/processed products \cr
}
}

\examples{

  \dontrun{
    x <- reportIntakeDetailed(gdx)
  }


}
\author{
Isabelle Weindl
}
