% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportKcal.R
\name{reportKcal}
\alias{reportKcal}
\title{reportKcal}
\usage{
reportKcal(gdx, detail = FALSE, level = "regglo")
}
\arguments{
\item{gdx}{GDX file}

\item{detail}{if detail=F, the subcategories of groups are not reported (e.g. "soybean" within "oilcrops")}

\item{level}{spatial aggregation: "reg", "glo", "regglo", "iso"}
}
\value{
per-capita calories as MAgPIE object (kcal/cap/day)
}
\description{
reports per-capita calories food supply (including household waste)
}
\section{Calorie supply variables}{
\tabular{lll}{
   Name \tab Unit \tab Meta \cr
   Nutrition|Calorie Supply \tab kcal/capita/day \tab Total per-capita calorie supply (including household waste) \cr
   Nutrition|Calorie Supply|+|Crops \tab kcal/capita/day \tab Calorie supply from crops \cr
   Nutrition|Calorie Supply|+|Livestock products \tab kcal/capita/day \tab Calorie supply from livestock products \cr
   Nutrition|Calorie Supply|+|Secondary products \tab kcal/capita/day \tab Calorie supply from secondary/processed products \cr
}
}

\examples{

  \dontrun{
    x <- reportKcal(gdx)
  }


}
\author{
Benjamin Leon Bodirsky, Kristine karstens, Abhijeet Mishra
}
