% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportLaborCostsEmpl.R
\name{reportLaborCostsEmpl}
\alias{reportLaborCostsEmpl}
\title{reportLaborCostsEmpl}
\usage{
reportLaborCostsEmpl(gdx)
}
\arguments{
\item{gdx}{GDX file}
}
\value{
magpie object with labor costs
}
\description{
reports MAgPIE labor costs that go into employment calculation
}
\examples{

  \dontrun{
    x <- reportLaborCostsEmpl(gdx)
  }

}
\author{
Debbora Leip
}
