% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportLaborProductivity.R
\name{reportLaborProductivity}
\alias{reportLaborProductivity}
\title{reportLaborProductivity}
\usage{
reportLaborProductivity(
  gdx,
  productAggr = TRUE,
  type = "physical",
  level = "regglo"
)
}
\arguments{
\item{gdx}{GDX file}

\item{productAggr}{Aggregate over products or not (boolean)}

\item{type}{type of labor productivity, so far only physical (kg DM / h)}

\item{level}{spatial aggregation: "reg", "glo", "regglo", "iso"}
}
\value{
labor productivity as MAgPIE object
}
\description{
reports labor productivity in crop production
}
\section{Labor productivity variables}{
\tabular{lll}{
   Name \tab Unit \tab Meta \cr
   Labor|Productivity|Physical labor productivity|Crop products \tab kg DM per hour \tab Physical labor productivity for crops \cr
}
}

\examples{

  \dontrun{
    x <- reportLaborProductivity(gdx)
  }


}
\author{
Debbora Leip
}
