% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportLandConservation.R
\name{reportLandConservation}
\alias{reportLandConservation}
\title{reportLandConservation}
\usage{
reportLandConservation(gdx)
}
\arguments{
\item{gdx}{GDX file}
}
\value{
land conservation area in Mha
}
\description{
reports land conservation areas
}
\examples{
\dontrun{
x <- reportLandConservation(gdx)
}

}
\author{
Patrick v. Jeetze, Florian Humpenoeder
}
