% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportLandFootprint.R
\name{reportLandFootprint}
\alias{reportLandFootprint}
\title{reportLandFootprint}
\usage{
reportLandFootprint(gdx, level = "regglo")
}
\arguments{
\item{gdx}{GDX file}

\item{level}{spatial aggregation: "reg", "glo", "regglo", "iso"}
}
\value{
land use for food production per capita (million hectares per-capita)
}
\description{
reports land footprint of food production for the population in a region
accounting for traded food assuming perfect substitution between domestic production
and imports
}
\examples{

  \dontrun{
    x <- reportLandFootprint(gdx)
  }

}
\author{
Vartika Singh
}
