% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportLandTransitionMatrix.R
\name{reportLandTransitionMatrix}
\alias{reportLandTransitionMatrix}
\title{reportLandTransitionMatrix}
\usage{
reportLandTransitionMatrix(gdx, level = "regglo")
}
\arguments{
\item{gdx}{GDX file}

\item{level}{The aggregation level to be used ("regglo" by default)}
}
\value{
land transition matrix as MAgPIE object (Mha/yr)
}
\description{
reports land transition matrix with gross land-use change flows
between all land types
}
\examples{

  \dontrun{
    x <- reportLandTransitionMatrix(gdx)
  }

}
\author{
Florian Humpenoeder
}
