% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportLandUse.R
\name{reportLandUse}
\alias{reportLandUse}
\title{reportLandUse}
\usage{
reportLandUse(gdx, level = "regglo")
}
\arguments{
\item{gdx}{GDX file}

\item{level}{The aggregation level to be used ("regglo" by default)}
}
\value{
land-use as MAgPIE object (million ha)
}
\description{
reports land-use
}
\examples{

  \dontrun{
    x <- reportLandUse(gdx)
  }

}
\author{
Florian Humpenoeder, Kristine Karstens, Isabelle Weindl
}
