% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportLandUseChange.R
\name{reportLandUseChange}
\alias{reportLandUseChange}
\title{reportLandUseChange}
\usage{
reportLandUseChange(gdx, baseyear = 1995)
}
\arguments{
\item{gdx}{GDX file}

\item{baseyear}{baseyear for calculating land-use change}
}
\value{
land-use change as MAgPIE object (million ha wrt to baseyear)
}
\description{
reports land-use change
}
\examples{

  \dontrun{
    x <- reportLandUseChange(gdx)
  }

}
\author{
Florian Humpenoeder
}
