% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportLivestockDemStructure.R
\name{reportLivestockDemStructure}
\alias{reportLivestockDemStructure}
\title{reportLivestockDemStructure}
\usage{
reportLivestockDemStructure(gdx)
}
\arguments{
\item{gdx}{GDX file}
}
\value{
livestock demand structure as MAgPIE object (kcal/kcal)
}
\description{
reports the share of different livestock products (excluding fish) in total livestock calorie food supply
}
\examples{

  \dontrun{
    x <- reportLivestockDemStructure(gdx)
  }

}
\author{
Isabelle Weindl
}
