% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportLivestockShare.R
\name{reportLivestockShare}
\alias{reportLivestockShare}
\title{reportLivestockShare}
\usage{
reportLivestockShare(gdx)
}
\arguments{
\item{gdx}{GDX file}
}
\value{
per-capita calories as MAgPIE object (kcal/cap/day)
}
\description{
reports the share of livestock products (including fish) in total calorie food supply
}
\examples{

  \dontrun{
    x <- reportLivestockShare(gdx)
  }

}
\author{
Benjamin Leon Bodirsky
}
