% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportLocalDemandShares.R
\name{reportLocalDemandShares}
\alias{reportLocalDemandShares}
\title{reportLocalDemandShares}
\usage{
reportLocalDemandShares(gdx, type = "prod", level = "regglo")
}
\arguments{
\item{gdx}{GDX file}

\item{type}{"prod" or "dem". the former indicates the share of production consumed in its own cluster
while the latter indicates how much cluster-level demand is satisfied by local consumption}

\item{level}{spatial aggregation: "reg", "glo", "regglo"}
}
\value{
share of food demand at disaggregated level coming from local production as MAgPIE object
}
\description{
reports local demand and production shares based on local consumption
}
\examples{
\dontrun{
x <- reportLocalDemandShares(gdx)
}

}
\author{
David M Chen
}
