% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportNetForestChange.R
\name{reportNetForestChange}
\alias{reportNetForestChange}
\title{reportNetForestChange}
\usage{
reportNetForestChange(gdx)
}
\arguments{
\item{gdx}{GDX file}
}
\value{
NetForestChange as magclass object (Mha per year)
}
\description{
reports Net Forest Change
}
\examples{
\dontrun{
x <- reportNetForestChange(gdx)
}

}
\author{
Florian Humpenoeder
}
