% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportNitrogenBudgetCropland.R
\name{reportNitrogenBudgetCropland}
\alias{reportNitrogenBudgetCropland}
\title{reportNitrogenBudgetCropland}
\usage{
reportNitrogenBudgetCropland(gdx, include_emissions = FALSE, grid = FALSE)
}
\arguments{
\item{gdx}{GDX file}

\item{include_emissions}{TRUE also divides the N surplus into different emissions}

\item{grid}{grid provides outputs on grid level of 0.5 degree}
}
\description{
Reports the Nitrogen Budgets of Croplands for future MAgPIE projections
}
\examples{
\dontrun{
x <- reportNitrogenBudgetCropland(gdx)
}

}
\seealso{
\code{\link{NitrogenBudget}}
}
\author{
Benjamin Leon Bodirsky
}
