% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportNitrogenBudgetPasture.R
\name{reportNitrogenBudgetPasture}
\alias{reportNitrogenBudgetPasture}
\title{reportNitrogenBudgetCropland}
\usage{
reportNitrogenBudgetPasture(gdx, include_emissions = FALSE, grid = FALSE)
}
\arguments{
\item{gdx}{GDX file}

\item{include_emissions}{TRUE also divides the N surplus into different emissions}

\item{grid}{if TRUE, disaggregate to grid level}
}
\description{
Reports the Nitrogen Budgets of Croplands for future MAgPIE projections
}
\section{Nitrogen pasture budget variables}{
\tabular{lll}{
   Name \tab Unit \tab Meta \cr
   Resources|Nitrogen|Pasture Budget|Inputs \tab Mt Nr/yr \tab Total nitrogen inputs to pastures \cr
   Resources|Nitrogen|Pasture Budget|Inputs|+|Manure \tab Mt Nr/yr \tab Manure nitrogen deposited on pastures \cr
   Resources|Nitrogen|Pasture Budget|Inputs|+|Atmospheric deposition \tab Mt Nr/yr \tab Atmospheric nitrogen deposition on pastures \cr
   Resources|Nitrogen|Pasture Budget|Withdrawals \tab Mt Nr/yr \tab Total nitrogen withdrawals from pastures \cr
   Resources|Nitrogen|Pasture Budget|Balance \tab Mt Nr/yr \tab Nitrogen balance on pastures \cr
   Resources|Nitrogen|Pasture Budget|Balance|+|Surplus \tab Mt Nr/yr \tab Nitrogen surplus on pastures \cr
}
}

\examples{
  \dontrun{
    x <- reportNitrogenBudgetCropland(gdx)
  }


}
\seealso{
\code{\link{NitrogenBudget}}
}
\author{
Benjamin Leon Bodirsky
}
