% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportNitrogenEfficiencies.R
\name{reportNitrogenEfficiencies}
\alias{reportNitrogenEfficiencies}
\title{reportNitrogenEfficiencies}
\usage{
reportNitrogenEfficiencies(gdx, level = "regglo")
}
\arguments{
\item{gdx}{GDX file}

\item{level}{aggregation level of returned data ("regglo" by default)}
}
\description{
Reports different nitrogen use efficiency indicators
}
\section{Nitrogen efficiency variables}{
\tabular{lll}{
   Name \tab Unit \tab Meta \cr
   Resources|Nitrogen|Cropland Budget|Nitrogen Use Efficiency complete \tab Mt Nr/Mt Nr \tab Complete nitrogen use efficiency on cropland \cr
   Resources|Nitrogen|Cropland Budget|Nitrogen Use Efficiency basic \tab Mt Nr/Mt Nr \tab Basic nitrogen use efficiency (harvest/inputs) \cr
   Resources|Nitrogen|Cropland Budget|Soil Nitrogen Uptake Efficiency \tab Mt Nr/Mt Nr \tab Soil nitrogen uptake efficiency \cr
   Resources|Nitrogen|Pasture Budget|Nitrogen Use Efficiency complete \tab Mt Nr/Mt Nr \tab Complete nitrogen use efficiency on pastures \cr
}
}

\examples{
\dontrun{
x <- reportNitrogenEfficiencies(gdx)
}

}
\seealso{
\code{\link{reportNitrogenEfficiencies}}
}
\author{
Benjamin Leon Bodirsky
}
