% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportNitrogenPollution.R
\name{reportNitrogenPollution}
\alias{reportNitrogenPollution}
\title{reportNitrogenPollution}
\usage{
reportNitrogenPollution(gdx, dir = ".")
}
\arguments{
\item{gdx}{GDX file}

\item{dir}{magpie output directory that contains gridded Nitrogen Data}
}
\description{
Reports total Nitrogen Pollution as the sum of surplus from cropland, pasture, awms, consumption and non-agricutlural land
}
\examples{
\dontrun{
x <- reportNitrogenPollution(gdx)
}

}
\seealso{
\code{\link{NitrogenBudget}}
}
\author{
Benjamin Leon Bodirsky, Michael Crawford
}
