% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportOutputPerWorker.R
\name{reportOutputPerWorker}
\alias{reportOutputPerWorker}
\title{reportOutputPerWorker}
\usage{
reportOutputPerWorker(gdx)
}
\arguments{
\item{gdx}{GDX file}
}
\value{
output per worker as MAgPIE object
}
\description{
reports output per worker in crop+livestock production from MAgPIE results
}
\examples{

  \dontrun{
    x <- reportOutputPerWorker(gdx)
  }

}
\author{
Debbora Leip
}
