% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportPastSoilCarbon.R
\name{reportPastSoilCarbon}
\alias{reportPastSoilCarbon}
\title{reportPastSoilCarbon}
\usage{
reportPastSoilCarbon(gdx)
}
\arguments{
\item{gdx}{GDX file}
}
\value{
Soil carbon values as magpie objetc
}
\description{
reports pasture soil carbon
}
\examples{
\dontrun{
x <- reportPastSoilCarbon(gdx)
}

}
\author{
Marcos Alves
}
