% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportPBbiosphere.R
\name{reportPBbiosphere}
\alias{reportPBbiosphere}
\title{reportPBbiosphere}
\usage{
reportPBbiosphere(gdx, level = "regglo", dir = ".")
}
\arguments{
\item{gdx}{GDX file}

\item{level}{level of aggregation (regglo: regions and global)}

\item{dir}{directory with required spatial data}
}
\value{
MAgPIE object
}
\description{
reports biosphere planetary boundary:
             Share of intact land relative to total land area (unitless)
             Share of intact land covered by areas within Global Safety Net (unitless)
             Share of land area that satisfies landscape target (unitless)
}
\examples{
\dontrun{
x <- reportPBbiosphere(gdx)
}

}
\author{
Patrick von Jeetze, Felicitas Beier
}
