% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportPBbiosphere.R
\name{reportPBbiosphere}
\alias{reportPBbiosphere}
\title{reportPBbiosphere}
\usage{
reportPBbiosphere(gdx, level = "regglo", dir = ".")
}
\arguments{
\item{gdx}{GDX file}

\item{level}{level of aggregation (regglo: regions and global)}

\item{dir}{directory with required spatial data}
}
\value{
MAgPIE object
}
\description{
reports biosphere planetary boundary:
             global land surface covered by largely intact natural areas
}
\examples{

  \dontrun{
    x <- reportPBbiosphere(gdx)
  }

}
\author{
Patrick von Jeetze, Felicitas Beier
}
