% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportPBland.R
\name{reportPBland}
\alias{reportPBland}
\title{reportPBland}
\usage{
reportPBland(gdx, level = "GLO", dir = ".")
}
\arguments{
\item{gdx}{GDX file}

\item{level}{level of aggregation (regglo: regions and global)}

\item{dir}{directory with required spatial data}
}
\value{
MAgPIE object
}
\description{
reports land planetary boundary: forest area as percentage of original
             forest area
}
\examples{

  \dontrun{
    x <- reportPBland(gdx)
  }

}
\author{
Felicitas Beier, Patrick von Jeetze
}
