% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportPBland.R
\name{reportPBland}
\alias{reportPBland}
\title{reportPBland}
\usage{
reportPBland(gdx, level = "regglo", dir = ".", foresttype = "all")
}
\arguments{
\item{gdx}{GDX file}

\item{level}{level of aggregation (regglo: regions and global)}

\item{dir}{directory with required spatial data}

\item{foresttype}{managed forest types that are included in the calculation of the
forest area (all: all managed forests,
             noTimber: timber plantations are not counted)}
}
\value{
MAgPIE object
}
\description{
reports land planetary boundary: forest area as percentage of original
             forest area
}
\examples{

  \dontrun{
    x <- reportPBland(gdx)
  }

}
\author{
Felicitas Beier, Patrick von Jeetze
}
