% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportPBnitrogen.R
\name{reportPBnitrogen}
\alias{reportPBnitrogen}
\title{reportPBnitrogen}
\usage{
reportPBnitrogen(gdx, level = "regglo")
}
\arguments{
\item{gdx}{GDX file}

\item{level}{level of aggregation (regglo: regions and global)}
}
\value{
MAgPIE object
}
\description{
reports nitrogen planetary boundary
}
\section{Nitrogen planetary boundary variables}{
\tabular{lll}{
   Name \tab Unit \tab Meta \cr
   Planetary Boundary|Nitrogen|Agricultural Nitrogen surplus \tab Mt N/yr \tab Total nitrogen surplus from cropland and pasture \cr
}
}

\examples{

  \dontrun{
    x <- reportPBnitrogen(gdx)
  }


}
\author{
Felicitas Beier, Mike Crawford
}
