% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportPBnitrogen.R
\name{reportPBnitrogen}
\alias{reportPBnitrogen}
\title{reportPBnitrogen}
\usage{
reportPBnitrogen(gdx, level = "regglo", dir = ".")
}
\arguments{
\item{gdx}{GDX file}

\item{level}{level of aggregation (regglo: regions and global)}

\item{dir}{directory with required spatial data}
}
\value{
MAgPIE object
}
\description{
reports nitrogen planetary boundary
}
\examples{

  \dontrun{
    x <- reportPBnitrogen(gdx)
  }

}
\author{
Mike Crawford, Felicitas Beier
}
