% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportPeatland.R
\name{reportPeatland}
\alias{reportPeatland}
\title{reportPeatland}
\usage{
reportPeatland(gdx)
}
\arguments{
\item{gdx}{GDX file}
}
\value{
peatland area as magclass object (million ha)
}
\description{
reports peatland area
}
\section{Peatland variables}{
\tabular{lll}{
   Name \tab Unit \tab Meta \cr
   Resources|Peatland \tab million ha \tab Area of peatlands \cr
   Resources|Peatland|+|Intact \tab million ha \tab Area of intact peatlands \cr
   Resources|Peatland|+|Degraded \tab million ha \tab Area of drained peatlands \cr
   Resources|Peatland|+|Rewetted \tab million ha \tab Area of rewetted peatlands \cr
}
}

\examples{

  \dontrun{
    x <- reportPeatland(gdx)
  }


}
\author{
Florian Humpenoeder
}
