% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportPlantationEstablishment.R
\name{reportPlantationEstablishment}
\alias{reportPlantationEstablishment}
\title{reportPlantationEstablishment}
\usage{
reportPlantationEstablishment(gdx)
}
\arguments{
\item{gdx}{GDX file}
}
\value{
Area harvested for timber production
}
\description{
reports MAgPIE harvested area for timber.
}
\examples{

  \dontrun{
    x <- reportPlantationEstablishment(gdx)
  }

}
\author{
Abhijeet Mishra
}
