% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportPopulation.R
\name{reportPopulation}
\alias{reportPopulation}
\title{reportPopulation}
\usage{
reportPopulation(gdx)
}
\arguments{
\item{gdx}{GDX file}
}
\value{
population as MAgPIE object
}
\description{
reports Population
}
\examples{

  \dontrun{
    x <- reportPopulation(gdx)
  }

}
\author{
Florian Humpenoeder
}
