% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportPriceAgriculture.R
\name{reportPriceAgriculture}
\alias{reportPriceAgriculture}
\title{reportPriceAgriculture}
\usage{
reportPriceAgriculture(gdx)
}
\arguments{
\item{gdx}{GDX file}
}
\value{
agricultural commodity prices as MAgPIE object (US\$)
}
\description{
reports food commodity prices
}
\examples{

  \dontrun{
    x <- reportPriceAgriculture(gdx)
  }

}
\author{
Mishko Stevanovic
}
