% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportPriceBioenergy.R
\name{reportPriceBioenergy}
\alias{reportPriceBioenergy}
\title{reportPriceBioenergy}
\usage{
reportPriceBioenergy(gdx, level = "regglo")
}
\arguments{
\item{gdx}{GDX file}
}
\value{
bioenergy price as MAgPIE object Unit: see names
}
\description{
reports bioenergy prices
}
\section{Bioenergy price variables}{
\tabular{lll}{
   Name \tab Unit \tab Meta \cr
   Prices|Bioenergy \tab US$2017/GJ \tab Bioenergy price \cr
}
}

\examples{

  \dontrun{
    x <- reportPriceBioenergy(gdx)
  }


}
\author{
Florian Humpenoeder
}
