% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportPriceFoodIndex.R
\name{reportPriceFoodIndex}
\alias{reportPriceFoodIndex}
\title{reportPriceFoodIndex}
\usage{
reportPriceFoodIndex(gdx, baseyear = "y2020")
}
\arguments{
\item{gdx}{GDX file}

\item{baseyear}{baseyear of the price index}
}
\value{
Food price index as MAgPIE object Unit: see names
}
\description{
reports food price index
}
\examples{

  \dontrun{
    x <- reportPriceFoodIndex(gdx)
  }

}
\author{
Florian Humpenoeder, Felicitas Beier
}
