% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportPriceGHG.R
\name{reportPriceGHG}
\alias{reportPriceGHG}
\title{reportPriceGHG}
\usage{
reportPriceGHG(gdx, level = "regglo")
}
\arguments{
\item{gdx}{GDX file}

\item{level}{An aggregation level for the spatial dimension. Can be any level
available via superAggregateX.}
}
\value{
GHG emission prices as MAgPIE object
}
\description{
reports GHG emission prices
}
\section{GHG price variables}{
\tabular{lll}{
   Name \tab Unit \tab Meta \cr
   Prices|GHG Emission|CO2 \tab US$2017/tCO2 \tab Carbon dioxide emission price \cr
   Prices|GHG Emission|N2O \tab US$2017/tN2O \tab Nitrous oxide emission price \cr
   Prices|GHG Emission|CH4 \tab US$2017/tCH4 \tab Methane emission price \cr
}
}

\examples{

  \dontrun{
    x <- reportPriceGHG(gdx)
  }


}
\author{
Florian Humpenoeder, Amsalu W. Yalew
}
