% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportPriceWater.R
\name{reportPriceWater}
\alias{reportPriceWater}
\title{reportPriceWater}
\usage{
reportPriceWater(gdx)
}
\arguments{
\item{gdx}{GDX file}
}
\value{
water usage as MAgPIE object Unit: see names
}
\description{
reports water prices
}
\examples{

  \dontrun{
    x <- reportPriceWater(gdx)
  }

}
\author{
Florian Humpenoeder
}
