% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportPriceWoodyBiomass.R
\name{reportPriceWoodyBiomass}
\alias{reportPriceWoodyBiomass}
\title{reportPriceWoodyBiomass}
\usage{
reportPriceWoodyBiomass(gdx, level = "regglo")
}
\arguments{
\item{gdx}{GDX file}

\item{level}{aggregation level of returned data ("regglo" by default)}
}
\value{
land prices as MAgPIE object Unit: see names
}
\description{
reports woody biomass prices (land rent)
}
\section{Woody biomass price variables}{
\tabular{lll}{
   Name \tab Unit \tab Meta \cr
   Prices|Wood \tab US$2017/tDM \tab Wood price \cr
   Prices|Woodfuel \tab US$2017/tDM \tab Woodfuel price \cr
}
}

\examples{

  \dontrun{
    x <- reportPriceWoodyBiomass(gdx)
  }


}
\author{
Abhijeet Mishra
}
