% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportProcessing.R
\name{reportProcessing}
\alias{reportProcessing}
\title{reportProcessing}
\usage{
reportProcessing(
  gdx,
  detail = TRUE,
  indicator = "primary_to_process",
  level = "regglo"
)
}
\arguments{
\item{gdx}{GDX file}

\item{detail}{if detail=FALSE, the subcategories of groups are not reported (e.g. "soybean" within "oilcrops")}

\item{indicator}{"primary_to_process" for process or "secondary_from_primary" for secondary product output}
}
\value{
processing demand as MAgPIE object (Mt DM)
}
\description{
reportes processing input and output quantities primary-to-process or primary-to-secondary
}
\section{Processing demand variables}{
\tabular{lll}{
   Name \tab Unit \tab Meta \cr
   Demand|Processing|++|Crops \tab Mt DM/yr \tab Primary crop products processed into secondary products \cr
   Demand|Processing|++|Livestock products \tab Mt DM/yr \tab Primary livestock products processed into secondary products \cr
   Processing|Raw material|Processed into Secondary products \tab Mt DM/yr \tab Total raw materials processed into secondary products \cr
}
}

\examples{
\dontrun{
x <- reportProcessing(gdx = gdx, detail = TRUE, indicator = "primary_to_process")
}


}
\author{
David Chen, Benjamin Leon Bodirsky
}
