% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportProducerPriceIndex.R
\name{reportProducerPriceIndex}
\alias{reportProducerPriceIndex}
\title{reportProducerPriceIndex}
\usage{
reportProducerPriceIndex(gdx, prod_groups = FALSE, level = "regglo")
}
\arguments{
\item{gdx}{GDX file}

\item{prod_groups}{whether to return only product groups}
}
\value{
Producer price index as MAgPIE object Unit: see names
}
\description{
reports producer price index
}
\section{Producer price index variables}{
\tabular{lll}{
   Name \tab Unit \tab Meta \cr
   Prices|Index2020|Agriculture|Producer|Primary food products \tab Index 2020=100 \tab Producer price index for primary food products \cr
   Prices|Index2020|Agriculture|Producer|Crops \tab Index 2020=100 \tab Producer price index for crops \cr
   Prices|Index2020|Agriculture|Producer|Livestock products \tab Index 2020=100 \tab Producer price index for livestock products \cr
   Prices|Index2020|Agriculture|Producer|Bioenergy \tab Index 2020=100 \tab Producer price index for bioenergy crops \cr
   Prices|Index2020|Agriculture|Producer|All agricultural products \tab Index 2020=100 \tab Producer price index for all agricultural products \cr
}
}

\examples{

  \dontrun{
    x <- reportProducerPriceIndex(gdx)
  }


}
\author{
Isabelle Weindl, David M CHen
}
