% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportProduction.R
\name{reportProduction}
\alias{reportProduction}
\title{reportProduction}
\usage{
reportProduction(gdx, detail = FALSE, agmip = FALSE, level = "regglo")
}
\arguments{
\item{gdx}{GDX file}

\item{detail}{if detail=FALSE, the subcategories of groups are not reported
(e.g. "soybean" within "oilcrops")}

\item{agmip}{if agmip = TRUE, additional sector aggregates
required for AgMIP are reported (e.g. "AGR")}

\item{level}{aggregation level of returned data ("regglo" by default)}
}
\value{
production as MAgPIE object. Unit: see names
}
\description{
reports production
}
\section{Production variables}{
\tabular{lll}{
   Name \tab Unit \tab Meta \cr
   Production \tab Mt DM/yr \tab Total agricultural production \cr
   Production|+|Crops \tab Mt DM/yr \tab Production of crops \cr
   Production|+|Livestock products \tab Mt DM/yr \tab Production of livestock products (excluding fish) \cr
   Production|+|Secondary products \tab Mt DM/yr \tab Production of secondary products (processed agricultural goods) \cr
   Production|+|Pasture \tab Mt DM/yr \tab Production of pasture biomass \cr
   Production|+|Bioenergy crops \tab Mt DM/yr \tab Production of second-generation bioenergy crops (short rotation grasses, short rotation trees) \cr
}
}

\examples{

  \dontrun{
    x <- reportProduction(gdx)
  }


}
\author{
Benjamin Leon Bodirsky, Isabelle Weindl
}
