% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportProductionBioenergy.R
\name{reportProductionBioenergy}
\alias{reportProductionBioenergy}
\title{reportProductionBioenergy}
\usage{
reportProductionBioenergy(gdx, detail = FALSE)
}
\arguments{
\item{gdx}{GDX file}

\item{detail}{if detail=FALSE, the subcategories of groups are not reported (e.g. "soybean" within "oilcrops")}
}
\value{
production as MAgPIE object. Unit: see names
}
\description{
reports 2nd gen bioenergy production
}
\examples{

  \dontrun{
    x <- reportProductionBioenergy(gdx)
  }

}
\author{
Florian Humpenoeder
}
