% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportProductionGrowth.R
\name{reportProductionGrowth}
\alias{reportProductionGrowth}
\title{reportProductionGrowth}
\usage{
reportProductionGrowth(gdx, detail = FALSE)
}
\arguments{
\item{gdx}{GDX file}

\item{detail}{if true, provides results for all commodities, otherwhise aggregates some groups}
}
\value{
Production growth rates (index)
}
\description{
reports production growth rate
}
\section{Production growth variables}{
\tabular{lll}{
   Name \tab Unit \tab Meta \cr
   Production|Production Growth Rate \tab Index \tab Production volume index relative to base year \cr
   Production|Production Growth Rate|+|Crop products \tab Index \tab Crop production growth index \cr
   Production|Production Growth Rate|+|Livestock products \tab Index \tab Livestock production growth index \cr
}
}

\examples{

  \dontrun{
    x <- reportProductionGrowth(gdx="fulldata.gdx",detail=TRUE)
  }

}
\author{
Xiaoxi Wang
}
