% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportProductionNr.R
\name{reportProductionNr}
\alias{reportProductionNr}
\title{reportProductionNr}
\usage{
reportProductionNr(gdx, detail = FALSE)
}
\arguments{
\item{gdx}{GDX file}

\item{detail}{if detail = FALSE, the subcategories of groups are not reported (e.g. "soybean" within "oilcrops")}
}
\value{
production as MAgPIE object. Unit: see names
}
\description{
reports production in Nr analogous to reportProduction
}
\examples{

  \dontrun{
    x <- reportProductionNr(gdx)
  }

}
\author{
Benjamin Leon Bodirsky
}
