% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportRelativeHourlyLaborCosts.R
\name{reportRelativeHourlyLaborCosts}
\alias{reportRelativeHourlyLaborCosts}
\title{reportRelativeHourlyLaborCosts}
\usage{
reportRelativeHourlyLaborCosts(gdx, level = "regglo")
}
\arguments{
\item{gdx}{GDX file}
}
\value{
labor costs per ag. worker in relation to GDP pc as MAgPIE object
}
\description{
reports labor costs per ag. worker in relation to GDP pc from MAgPIE results
}
\section{Relative hourly labor cost variables}{
\tabular{lll}{
   Name \tab Unit \tab Meta \cr
   Labor|Wages|Labor costs per worker relative to GDP pc \tab \% \tab Agricultural labor costs relative to GDP per capita \cr
}
}

\examples{

  \dontrun{
    x <- reportRelativeHourlyLaborCosts(gdx)
  }


}
\author{
Debbora Leip
}
