% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportRotationLength.R
\name{reportRotationLength}
\alias{reportRotationLength}
\title{reportRotationLength}
\usage{
reportRotationLength(gdx, level = "regglo")
}
\arguments{
\item{gdx}{GDX file}

\item{level}{aggregation level of returned data ("regglo" by default)}
}
\value{
Forest rotation length
}
\description{
reports Forest rotation length.
}
\section{Rotation length variables}{
\tabular{lll}{
   Name \tab Unit \tab Meta \cr
   Rotation lengths|Forestry \tab years \tab Optimal rotation length for timber plantations \cr
}
}

\examples{

  \dontrun{
    x <- reportRotationLength(gdx)
  }

}
\author{
Abhijeet Mishra
}
