% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportRuralDemandShares.R
\name{reportRuralDemandShares}
\alias{reportRuralDemandShares}
\title{reportRuralDemandShares}
\usage{
reportRuralDemandShares(gdx, type = "tradOnly", level = "regglo")
}
\arguments{
\item{gdx}{GDX file}

\item{type}{Type of ratio that should be calculated
\itemize{
       \item \code{all}: How much rural & trad demand as a share of all demand is satisfied locally
       \item \code{tradOnly}: How much rural & trad demand as a share of rural & trad demand is satisfied locally
       \item \code{potential}: How much total gridded demand is potentially
                               satisfied by gridded production
       }}

\item{level}{spatial aggregation: "reg", "glo", "regglo"}
}
\value{
share of food demand at disaggregated level coming from local production as MAgPIE object
}
\description{
reports rural demand and production shares based on local consumption
}
\examples{
\dontrun{
x <- reportruralDemandShares(gdx)
}

}
\author{
David M Chen
}
