% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportSDG1.R
\name{reportSDG1}
\alias{reportSDG1}
\title{reportSDG1}
\usage{
reportSDG1(gdx, level = "regglo")
}
\arguments{
\item{gdx}{GDX file}

\item{level}{An aggregation level for the spatial dimension. Can be any level
available via superAggregateX.}
}
\value{
MAgPIE object
}
\description{
reports all SDG indicators relevant for SDG1 - Poverty
}
\section{SDG1 Poverty variables}{
\tabular{lll}{
   Name \tab Unit \tab Meta \cr
   SDG|SDG01|Per-capita income \tab US$2017 PPP/cap/yr \tab GDP per capita (after shock) \cr
}
}

\examples{

  \dontrun{
    x <- reportSDG3(gdx)
  }


}
\author{
Benjamin Bodirsky
}
