% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportSDG2.R
\name{reportSDG2}
\alias{reportSDG2}
\title{reportSDG2}
\usage{
reportSDG2(gdx)
}
\arguments{
\item{gdx}{GDX file}
}
\value{
MAgPIE object
}
\description{
reports all SDG indicators relevant for SD2 - Hunger
}
\section{SDG2 Hunger variables}{
\tabular{lll}{
   Name \tab Unit \tab Meta \cr
   SDG|SDG02|Prevalence of underweight \tab million \tab Population with underweight BMI \cr
   SDG|SDG02|Prevalence of underweight|Children \tab million \tab Children under 5 with underweight BMI \cr
   SDG|SDG02|Food availability \tab kcal/cap/day \tab Daily per-capita caloric availability \cr
   SDG|SDG02|Food expenditure share \tab income \tab Share of income spent on food (value added) \cr
   SDG|SDG02|Agricultural primary product expenditure share \tab income \tab Share of income spent on agricultural primary products \cr
   SDG|SDG02|Agricultural commodity price index wrt 2020 \tab 1 \tab Food price index relative to 2020 baseline \cr
   SDG|SDG02|Prevalence of obesity|Children \tab million \tab Children under 5 with obesity \cr
   SDG|SDG02|Investment in AgR&D \tab million USD05/yr \tab Investment in agricultural research and development \cr
}
}

\examples{

  \dontrun{
    x <- reportSDG2(gdx)
  }


}
\author{
Benjamin Bodirsky
}
