% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportSDG6.R
\name{reportSDG6}
\alias{reportSDG6}
\title{reportSDG6}
\usage{
reportSDG6(gdx, level = "regglo")
}
\arguments{
\item{gdx}{GDX file}

\item{level}{An aggregation level for the spatial dimension. Can be any level
available via superAggregateX.}
}
\value{
MAgPIE object
}
\description{
reports all SDG indicators relevant for SDG6 - Access to Water
}
\section{SDG6 Water variables}{
\tabular{lll}{
   Name \tab Unit \tab Meta \cr
   SDG|SDG06|Fertilizer use \tab Mt N/yr \tab Nitrogen fertilizer application on cropland \cr
   SDG|SDG06|Nitrogen surplus on cropland \tab Mt N/yr \tab Nitrogen surplus from cropland budget \cr
   SDG|SDG06|Agricultural water use \tab km3/yr \tab Agricultural water withdrawal during growing period \cr
}
}

\examples{

  \dontrun{
    x <- reportSDG6(gdx)
  }


}
\author{
Felicitas Beier, Isabelle Weindl
}
