% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportSOM.R
\name{reportSOM}
\alias{reportSOM}
\title{reportSOM}
\usage{
reportSOM(gdx, baseyear = 1995)
}
\arguments{
\item{gdx}{GDX file}

\item{baseyear}{baseyear for calculating carbon stock change}
}
\description{
Report soil organic carbon stock size for future MAgPIE projections
}
\examples{
  \dontrun{
    x <- reportSOM(gdx)
  }

}
\author{
Kristine Karstens
}
