% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportSOM2.R
\name{reportSOM2}
\alias{reportSOM2}
\title{reportSOM2}
\usage{
reportSOM2(gdx, baseyear = 1995)
}
\arguments{
\item{gdx}{GDX file}

\item{baseyear}{baseyear for calculating carbon stock change}
}
\description{
Report soil organic carbon stock size for future MAgPIE projections (new som realization)
}
\examples{
  \dontrun{
    x <- reportSOM2(gdx)
  }

}
\author{
Kristine Karstens
}
