% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportTau.R
\name{reportTau}
\alias{reportTau}
\title{reportTau}
\usage{
reportTau(gdx)
}
\arguments{
\item{gdx}{GDX file}
}
\value{
tau values as MAgPIE object (Index)
}
\description{
reports Tau
}
\examples{
\dontrun{
x <- reportTau(gdx)
}

}
\author{
Florian Humpenoeder, Patrick v. Jeetze
}
