% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportTau.R
\name{reportTau}
\alias{reportTau}
\title{reportTau}
\usage{
reportTau(gdx)
}
\arguments{
\item{gdx}{GDX file}
}
\value{
tau values as MAgPIE object (Index)
}
\description{
reports Tau
}
\section{Tau variables}{
\tabular{lll}{
   Name \tab Unit \tab Meta \cr
   Productivity|Landuse Intensity Indicator Tau \tab Index \tab Agricultural land-use intensity indicator for crops \cr
   Productivity|Landuse Intensity Indicator Tau managed pastures \tab Index \tab Land-use intensity indicator for managed pastures \cr
}
}

\examples{
\dontrun{
x <- reportTau(gdx)
}


}
\author{
Florian Humpenoeder, Patrick v. Jeetze
}
