% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportTc.R
\name{reportTc}
\alias{reportTc}
\title{reportTc}
\usage{
reportTc(gdx, level = "regglo")
}
\arguments{
\item{gdx}{GDX file}

\item{level}{Aggregation level of the returned Tc report}
}
\value{
tc values as MAgPIE object (%/yr)
}
\description{
reports Tc
}
\examples{

  \dontrun{
    x <- reportTc(gdx)
  }

}
\author{
Florian Humpenoeder
}
