% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportTotalHoursWorked.R
\name{reportTotalHoursWorked}
\alias{reportTotalHoursWorked}
\title{reportTotalHoursWorked}
\usage{
reportTotalHoursWorked(gdx, level = "regglo")
}
\arguments{
\item{gdx}{GDX file}

\item{level}{spatial aggregation: "reg", "glo", "regglo"}
}
\value{
total hours worked as MAgPIE object
}
\description{
reports total hours worked in crop+livestock production from MAgPIE results
}
\examples{

  \dontrun{
    x <- reportTotalHoursWorked(gdx)
  }

}
\author{
Debbora Leip
}
