% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportTrade.R
\name{reportTrade}
\alias{reportTrade}
\title{reportTrade}
\usage{
reportTrade(gdx, detail = FALSE)
}
\arguments{
\item{gdx}{GDX file}

\item{detail}{if true, provides estimates for all commodities, otherwise aggregates some groups}
}
\value{
Net-Exports and self sufficiency (exports/domestic supply) as MAgPIE object. Unit: see names
}
\description{
reports trade
}
\examples{

  \dontrun{
    x <- reportTrade(gdx="fulldata.gdx",detail=TRUE)
  }
}
\author{
Benjamin Leon Bodirsky, Mishko Stevanovic
}
