% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportTradeGrowth.R
\name{reportTradeGrowth}
\alias{reportTradeGrowth}
\title{reportTradeGrowth}
\usage{
reportTradeGrowth(gdx, detail = FALSE)
}
\arguments{
\item{gdx}{GDX file}

\item{detail}{if true, provides results for all commodities, otherwhise aggregates some groups}
}
\value{
Trade growth rates (index)
}
\description{
reports trade growth rate
}
\examples{

  \dontrun{
    x <- reportTradeGrowth(gdx="fulldata.gdx",detail=TRUE)
  }

}
\author{
Xiaoxi Wang
}
