% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportValueMaterialDemand.R
\name{reportValueMaterialDemand}
\alias{reportValueMaterialDemand}
\title{reportValueMaterialDemand}
\usage{
reportValueMaterialDemand(gdx)
}
\arguments{
\item{gdx}{GDX file}
}
\value{
magpie object
}
\description{
reports value of material demand
}
\examples{
\dontrun{
x <- reportValueMaterialDemand(gdx)
}


}
\author{
David Chen
}
