% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportVegfruitShare.R
\name{reportVegfruitShare}
\alias{reportVegfruitShare}
\title{reportVegfruitShare}
\usage{
reportVegfruitShare(gdx, level = "regglo")
}
\arguments{
\item{gdx}{GDX file}

\item{level}{An aggregation level for the spatial dimension. Can be any level
available via superAggregateX.}
}
\value{
per-capita calories as MAgPIE object (kcal/cap/day)
}
\description{
reports the share of livestock products (including fish) in total calorie food supply
}
\section{Vegfruit share variables}{
\tabular{lll}{
   Name \tab Unit \tab Meta \cr
   Nutrition|Dietary Composition|Vegetables Fruits Nuts Share \tab kcal/kcal \tab Share of vegetables, fruits, and nuts in total calorie supply \cr
}
}

\examples{
  \dontrun{
    x <- reportLivestockShare(gdx)
  }

}
\author{
Benjamin Leon Bodirsky
}
