% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportWageDevelopment.R
\name{reportWageDevelopment}
\alias{reportWageDevelopment}
\title{reportWageDevelopment}
\usage{
reportWageDevelopment(gdx, baseYear = 2000, level = "regglo")
}
\arguments{
\item{gdx}{GDX file}

\item{baseYear}{year relative to which the wage development should be calculated}

\item{level}{spatial aggregation: "reg", "glo", "regglo"}
}
\value{
indicator on wage development as MAgPIE object
}
\description{
reports indicator on wage development: hourly labor costs in each time step relative to hourly labor
costs in 2000
}
\examples{
\dontrun{
x <- reportWageDevelopment(gdx)
}

}
\author{
Debbora Leip
}
