% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportWaterAvailability.R
\name{reportWaterAvailability}
\alias{reportWaterAvailability}
\title{reportWaterAvailability}
\usage{
reportWaterAvailability(gdx)
}
\arguments{
\item{gdx}{GDX file}
}
\value{
water availability as MAgPIE object Unit: see names
}
\description{
reports water availability
}
\section{Water availability variables}{
\tabular{lll}{
   Name \tab Unit \tab Meta \cr
   Resources|Water|Availability|Agriculture \tab km3/yr \tab Water available for agricultural use \cr
}
}

\examples{

  \dontrun{
    x <- reportWaterAvailability(gdx)
  }


}
\author{
Felicitas Beier
}
