% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportWaterIndicators.R
\name{reportWaterIndicators}
\alias{reportWaterIndicators}
\title{reportWaterIndicators}
\usage{
reportWaterIndicators(gdx, level = "regglo", dir = ".")
}
\arguments{
\item{gdx}{GDX file}

\item{level}{level of aggregation (cluster: "cell",
regional: "regglo", "reg", "glo")}

\item{dir}{for gridded outputs: magpie output directory which contains
a mapping file (rds or spam)}
}
\value{
MAgPIE object
}
\description{
reports a set of water indicators
}
\examples{
\dontrun{
x <- reportWaterIndicators(gdx)
}
}
\author{
Felicitas Beier
}
