% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportWaterIndicators.R
\name{reportWaterIndicators}
\alias{reportWaterIndicators}
\title{reportWaterIndicators}
\usage{
reportWaterIndicators(gdx)
}
\arguments{
\item{gdx}{GDX file}
}
\value{
MAgPIE object
}
\description{
reports a set of water indicators
}
\examples{
\dontrun{
x <- reportWaterIndicators(gdx)
}
}
\author{
Felicitas Beier
}
