% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportWaterUsage.R
\name{reportWaterUsage}
\alias{reportWaterUsage}
\title{reportWaterUsage}
\usage{
reportWaterUsage(gdx, detail = TRUE)
}
\arguments{
\item{gdx}{GDX file}

\item{detail}{logical. Setting to FALSE reports for agricultural sector, TRUE reports for combined, crops and livestock separately}
}
\value{
water usage as MAgPIE object Unit: see names
}
\description{
reports water usage for agricultural sector, crops and livestock
}
\examples{

  \dontrun{
    x <- reportWaterUsage(gdx)
  }

}
\author{
Florian Humpenoeder, Vartika Singh, Miodrag Stevanovic
}
