% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportYieldsCropCalib.R
\name{reportYieldsCropCalib}
\alias{reportYieldsCropCalib}
\title{reportYieldsCropCalib}
\usage{
reportYieldsCropCalib(gdx, detail = FALSE)
}
\arguments{
\item{gdx}{GDX file}

\item{detail}{if detail=FALSE, the subcategories of groups are not reported (e.g. "soybean" within "oilcrops")}
}
\value{
yield as MAgPIE object (Mt DM/ha)
}
\description{
reports potential yields after calibration
}
\examples{
\dontrun{
x <- reportYieldsCropCalib(gdx)
}

}
\author{
Edna J. Molina Bacca
}
