% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportYieldsCropCalib.R
\name{reportYieldsCropCalib}
\alias{reportYieldsCropCalib}
\title{reportYieldsCropCalib}
\usage{
reportYieldsCropCalib(gdx, detail = FALSE, level = "regglo")
}
\arguments{
\item{gdx}{GDX file}

\item{detail}{if detail=FALSE, the subcategories of groups are not reported (e.g. "soybean" within "oilcrops")}
}
\value{
yield as MAgPIE object (Mt DM/ha)
}
\description{
reports potential yields after calibration
}
\section{Calibrated yield variables}{
\tabular{lll}{
   Name \tab Unit \tab Meta \cr
   Productivity|Yield (after calibration) \tab t DM/ha \tab Potential crop yields after calibration \cr
   Productivity|Yield (after calibration)|+|Cereals \tab t DM/ha \tab Calibrated cereal yields \cr
   Productivity|Yield (including tau) \tab t DM/ha \tab Potential yields including tau factor \cr
}
}

\examples{
\dontrun{
x <- reportYieldsCropCalib(gdx)
}

}
\author{
Edna J. Molina Bacca
}
