% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportYieldsCropRaw.R
\name{reportYieldsCropRaw}
\alias{reportYieldsCropRaw}
\title{reportYieldsCropRaw}
\usage{
reportYieldsCropRaw(gdx, detail = FALSE, level = "regglo")
}
\arguments{
\item{gdx}{GDX file}

\item{detail}{if detail=FALSE, the subcategories of groups are not reported (e.g. "soybean" within "oilcrops")}

\item{level}{aggregation level of returned data ("regglo" by default)}
}
\value{
yield as MAgPIE object (Mt DM/ha)
}
\description{
reports potential yields before calibration
}
\section{Raw yield variables}{
\tabular{lll}{
   Name \tab Unit \tab Meta \cr
   Productivity|Yield (before calibration) \tab t DM/ha \tab Potential crop yields before calibration \cr
   Productivity|Yield (before calibration)|+|Cereals \tab t DM/ha \tab Uncalibrated cereal yields \cr
}
}

\examples{
\dontrun{
x <- reportYieldsCropRaw(gdx)
}

}
\author{
Edna J. Molina Bacca
}
