% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportYieldShifter.R
\name{reportYieldShifter}
\alias{reportYieldShifter}
\title{reportYieldShifter}
\usage{
reportYieldShifter(
  gdx,
  file = NULL,
  level = "reg",
  baseyear = "y2000",
  relative = TRUE,
  dir = ".",
  spamfiledirectory = ""
)
}
\arguments{
\item{gdx}{GDX file}

\item{file}{a file name the output should be written to using write.magpie}

\item{level}{Level of regional aggregation; "reg" (regional), "glo" (global), "regglo" (regional and global) or any other aggregation level defined in superAggregate}

\item{baseyear}{baseyear for the yield shifter. Also fixes land patterns for aggregation to baseyear.}

\item{relative}{relative or absolute changes to baseyear}

\item{dir}{for gridded outputs: magpie output directory which contains a mapping file (rds) for disaggregation}

\item{spamfiledirectory}{deprecated. please use \code{dir} instead}
}
\value{
crop yield as MAgPIE object (unit depends on attributes)
}
\description{
Reports the Crop model input yield shifter
}
\examples{

  \dontrun{
    x <- reportYieldShifter(gdx)
  }

}
\seealso{
\code{\link{reportYieldShifter}}
}
\author{
Benjamin Leon Bodirsky
}
