% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOM.R
\name{SOM}
\alias{SOM}
\title{SOM}
\usage{
SOM(
  gdx,
  file = NULL,
  type = "stock",
  reference = "actual",
  level = "reg",
  noncrop_aggr = TRUE,
  dir = ".",
  spamfiledirectory = ""
)
}
\arguments{
\item{gdx}{GDX file}

\item{file}{a file name the output should be written to using write.magpie}

\item{type}{"stock" (default) for absoulte values,
"density" for per hectar values}

\item{reference}{default is "actual" (cshare in actual carbon stocks).
Other option is "target" (cshare in target carbon stocks).}

\item{level}{Level of regional aggregation;
"reg" (regional),
"glo" (global),
"regglo" (regional and global) or
any other aggregation level defined in superAggregate}

\item{noncrop_aggr}{aggregate non cropland types to 'noncropland'
(if FALSE all land types of pools59 will be reported)}

\item{dir}{for gridded outputs: magpie output directory
which contains a mapping file (rds or spam) disaggregation}

\item{spamfiledirectory}{deprecated. please use \code{dir} instead}
}
\value{
A MAgPIE object containing som values
}
\description{
Calculates soil organic carbon stock size
             based on a MAgPIE gdx file
}
\examples{
\dontrun{
x <- SOM(gdx)
}

}
\author{
Kristine Karstens
}
