% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tradedPrimaries.R
\name{tradedPrimaries}
\alias{tradedPrimaries}
\title{tradedPrimaries}
\usage{
tradedPrimaries(gdx, file = NULL)
}
\arguments{
\item{gdx}{GDX file}

\item{file}{a file name the output should be written to using write.magpie}
}
\value{
MAgPIE object with primary product trade equivalents in dry matter (tDM)
}
\description{
Converts trade flows into primary product equivalents by tracing
             secondary and livestock products back to their primary inputs.
             Trade flows are decomposed into three pathways:
             (1) direct primary trade, (2) primaries embodied in secondary products,
             and (3) primaries needed as feed for livestock.
}
\details{
Processing shares and livestock feed baskets of the importing region are used to 
         calculate primary product requirements. This means that no footprint can be assigned 
         to secondaries or livestock products originating from the producer region's 
         processing pathways or feed baskets. With a bilateral trade implementation, these 
         differences between producer and importer could be tracked.
}
\examples{
\dontrun{
  x <- tradedPrimaries(gdx)
}

}
\seealso{
\code{\link{land}}, \code{\link{croparea}}, \code{\link{trade}}
}
\author{
Kristine Karstens, David M Chen
}
