% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tryList.R
\name{tryList}
\alias{tryList}
\title{tryList}
\usage{
tryList(..., gdx)
}
\arguments{
\item{...}{report function to be run}

\item{gdx}{gdx file to report from}
}
\description{
Internal support function to run a list of reportings in
a \code{\link{tryReport}} environment.
}
\seealso{
\code{\link{tryReport}}
}
\author{
Jan Philipp Dietrich
}
