% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tryReport.R
\name{tryReport}
\alias{tryReport}
\title{tryReport}
\usage{
tryReport(report, gdx, level = "regglo", env = parent.frame())
}
\arguments{
\item{report}{report function to be run}

\item{gdx}{gdx file to report from}

\item{level}{spatial level (either "regglo" for region+global, "iso" for country-level,
or the file of a mapping file).}

\item{env}{environment to evaluate the report in}
}
\value{
A named list with information on the outcome of the report (success, error, validationError, warning)
}
\description{
Internal support function to run a reporting in a try environment
and properly report problems if something goes wrong without stopping the
further processing in case of an error.
}
\seealso{
\code{\link{reportResult}}
}
\author{
Jan Philipp Dietrich
}
