% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tryReport.R
\name{tryReport}
\alias{tryReport}
\title{tryReport}
\usage{
tryReport(report, width, gdx, level = "regglo", n = 1)
}
\arguments{
\item{report}{report function to be run}

\item{width}{max number of characters per line}

\item{gdx}{gdx file to report from}

\item{level}{spatial level (either "regglo" for region+global or "iso" for ISO countries)}

\item{n}{number of parent generations to go back when catching the environment
the report should get evaluated in}
}
\description{
Internal support function to run a reporting in a try environment
and properly report problems if something goes wrong without stopping the
further processing in case of an error
}
\author{
Jan Philipp Dietrich
}
