% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/water_usage.R
\name{water_usage}
\alias{water_usage}
\title{water_usage}
\usage{
water_usage(
  gdx,
  file = NULL,
  level = "reg",
  users = NULL,
  sum = FALSE,
  digits = 0
)
}
\arguments{
\item{gdx}{GDX file}

\item{file}{a file name the output should be written to using write.magpie}

\item{level}{spatial level of aggregation: "cell" (cellular), "reg" (regional), "glo" (global), "regglo" (regional and global) or any other aggregation level defined in superAggregate}

\item{users}{If sectors, will only report for high-level sectors - agriculture, industry, electricity, domestic, ecosystem. Sum not applicable in this case}

\item{sum}{determines whether output should be sector specific (FALSE) or aggregated over all sectors (TRUE)}

\item{digits}{integer. For rounding of the return values}
}
\value{
A MAgPIE object containing the water usage (km^3/yr)
}
\description{
reads area usage from a MAgPIE gdx file
}
\examples{

  \dontrun{
    x <- water_usage(gdx)
  }

}
\author{
Markus Bonsch, Vartika Singh
}
